Module aws_sdk_securityhub::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptAdministratorInvitation

Fluent builder constructing a request to AcceptAdministratorInvitation.

AcceptInvitation

Fluent builder constructing a request to AcceptInvitation.

BatchDisableStandards

Fluent builder constructing a request to BatchDisableStandards.

BatchEnableStandards

Fluent builder constructing a request to BatchEnableStandards.

BatchImportFindings

Fluent builder constructing a request to BatchImportFindings.

BatchUpdateFindings

Fluent builder constructing a request to BatchUpdateFindings.

CreateActionTarget

Fluent builder constructing a request to CreateActionTarget.

CreateFindingAggregator

Fluent builder constructing a request to CreateFindingAggregator.

CreateInsight

Fluent builder constructing a request to CreateInsight.

CreateMembers

Fluent builder constructing a request to CreateMembers.

DeclineInvitations

Fluent builder constructing a request to DeclineInvitations.

DeleteActionTarget

Fluent builder constructing a request to DeleteActionTarget.

DeleteFindingAggregator

Fluent builder constructing a request to DeleteFindingAggregator.

DeleteInsight

Fluent builder constructing a request to DeleteInsight.

DeleteInvitations

Fluent builder constructing a request to DeleteInvitations.

DeleteMembers

Fluent builder constructing a request to DeleteMembers.

DescribeActionTargets

Fluent builder constructing a request to DescribeActionTargets.

DescribeHub

Fluent builder constructing a request to DescribeHub.

DescribeOrganizationConfiguration

Fluent builder constructing a request to DescribeOrganizationConfiguration.

DescribeProducts

Fluent builder constructing a request to DescribeProducts.

DescribeStandards

Fluent builder constructing a request to DescribeStandards.

DescribeStandardsControls

Fluent builder constructing a request to DescribeStandardsControls.

DisableImportFindingsForProduct

Fluent builder constructing a request to DisableImportFindingsForProduct.

DisableOrganizationAdminAccount

Fluent builder constructing a request to DisableOrganizationAdminAccount.

DisableSecurityHub

Fluent builder constructing a request to DisableSecurityHub.

DisassociateFromAdministratorAccount

Fluent builder constructing a request to DisassociateFromAdministratorAccount.

DisassociateFromMasterAccount

Fluent builder constructing a request to DisassociateFromMasterAccount.

DisassociateMembers

Fluent builder constructing a request to DisassociateMembers.

EnableImportFindingsForProduct

Fluent builder constructing a request to EnableImportFindingsForProduct.

EnableOrganizationAdminAccount

Fluent builder constructing a request to EnableOrganizationAdminAccount.

EnableSecurityHub

Fluent builder constructing a request to EnableSecurityHub.

GetAdministratorAccount

Fluent builder constructing a request to GetAdministratorAccount.

GetEnabledStandards

Fluent builder constructing a request to GetEnabledStandards.

GetFindingAggregator

Fluent builder constructing a request to GetFindingAggregator.

GetFindings

Fluent builder constructing a request to GetFindings.

GetInsightResults

Fluent builder constructing a request to GetInsightResults.

GetInsights

Fluent builder constructing a request to GetInsights.

GetInvitationsCount

Fluent builder constructing a request to GetInvitationsCount.

GetMasterAccount

Fluent builder constructing a request to GetMasterAccount.

GetMembers

Fluent builder constructing a request to GetMembers.

InviteMembers

Fluent builder constructing a request to InviteMembers.

ListEnabledProductsForImport

Fluent builder constructing a request to ListEnabledProductsForImport.

ListFindingAggregators

Fluent builder constructing a request to ListFindingAggregators.

ListInvitations

Fluent builder constructing a request to ListInvitations.

ListMembers

Fluent builder constructing a request to ListMembers.

ListOrganizationAdminAccounts

Fluent builder constructing a request to ListOrganizationAdminAccounts.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateActionTarget

Fluent builder constructing a request to UpdateActionTarget.

UpdateFindingAggregator

Fluent builder constructing a request to UpdateFindingAggregator.

UpdateFindings

Fluent builder constructing a request to UpdateFindings.

UpdateInsight

Fluent builder constructing a request to UpdateInsight.

UpdateOrganizationConfiguration

Fluent builder constructing a request to UpdateOrganizationConfiguration.

UpdateSecurityHubConfiguration

Fluent builder constructing a request to UpdateSecurityHubConfiguration.

UpdateStandardsControl

Fluent builder constructing a request to UpdateStandardsControl.