Struct aws_sdk_securityhub::client::fluent_builders::BatchUpdateFindings

pub struct BatchUpdateFindings<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> BatchUpdateFindings<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<BatchUpdateFindingsOutput, SdkError<BatchUpdateFindingsError>> where
R::Policy: SmithyRetryPolicy<BatchUpdateFindingsInputOperationOutputAlias, BatchUpdateFindingsOutput, BatchUpdateFindingsError, BatchUpdateFindingsInputOperationRetryAlias>,

pub fn finding_identifiers(
self,
inp: impl Into<AwsSecurityFindingIdentifier>
) -> Self

Appends an item to FindingIdentifiers.

To override the contents of this collection use set_finding_identifiers.

The list of findings to update. BatchUpdateFindings can be used to update up to 100 findings at a time.

For each finding, the list provides the finding identifier and the ARN of the finding provider.

pub fn set_finding_identifiers(
self,
input: Option<Vec<AwsSecurityFindingIdentifier>>
) -> Self

pub fn note(self, inp: NoteUpdate) -> Self

The updated note.

pub fn set_note(self, input: Option<NoteUpdate>) -> Self

pub fn severity(self, inp: SeverityUpdate) -> Self

Used to update the finding severity.

pub fn set_severity(self, input: Option<SeverityUpdate>) -> Self

pub fn verification_state(self, inp: VerificationState) -> Self

Indicates the veracity of a finding.

The available values for VerificationState are as follows.

UNKNOWN – The default disposition of a security finding

TRUE_POSITIVE – The security finding is confirmed

FALSE_POSITIVE – The security finding was determined to be a false alarm

BENIGN_POSITIVE – A special case of TRUE_POSITIVE where the finding doesn't pose any threat, is expected, or both

pub fn set_verification_state(self, input: Option<VerificationState>) -> Self

pub fn confidence(self, inp: i32) -> Self

The updated value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn set_confidence(self, input: Option<i32>) -> Self

pub fn criticality(self, inp: i32) -> Self

The updated value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn set_criticality(self, input: Option<i32>) -> Self

pub fn types(self, inp: impl Into<String>) -> Self

Appends an item to Types.

To override the contents of this collection use set_types.

One or more finding types in the format of namespace/category/classifier that classify a finding.

Valid namespace values are as follows.

Software and Configuration Checks

TTPs

Effects

Unusual Behaviors

Sensitive Data Identifications

pub fn set_types(self, input: Option<Vec<String>>) -> Self

pub fn user_defined_fields(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to UserDefinedFields.

To override the contents of this collection use set_user_defined_fields.

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

pub fn set_user_defined_fields(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn workflow(self, inp: WorkflowUpdate) -> Self

Used to update the workflow status of a finding.

The workflow status indicates the progress of the investigation into the finding.

pub fn set_workflow(self, input: Option<WorkflowUpdate>) -> Self

pub fn related_findings(self, inp: impl Into<RelatedFinding>) -> Self

Appends an item to RelatedFindings.

To override the contents of this collection use set_related_findings.

A list of findings that are related to the updated findings.

pub fn set_related_findings(self, input: Option<Vec<RelatedFinding>>) -> Self