Struct aws_sdk_securityhub::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for SecurityHubAPIService.

This client allows ergonomic access to a SecurityHubAPIService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_administrator_invitation(
&self
) -> AcceptAdministratorInvitation<C, M, R>

pub fn accept_invitation(&self) -> AcceptInvitation<C, M, R>

pub fn batch_disable_standards(&self) -> BatchDisableStandards<C, M, R>

pub fn batch_enable_standards(&self) -> BatchEnableStandards<C, M, R>

pub fn batch_import_findings(&self) -> BatchImportFindings<C, M, R>

pub fn batch_update_findings(&self) -> BatchUpdateFindings<C, M, R>

pub fn create_action_target(&self) -> CreateActionTarget<C, M, R>

pub fn create_insight(&self) -> CreateInsight<C, M, R>

pub fn create_members(&self) -> CreateMembers<C, M, R>

pub fn decline_invitations(&self) -> DeclineInvitations<C, M, R>

pub fn delete_action_target(&self) -> DeleteActionTarget<C, M, R>

pub fn delete_insight(&self) -> DeleteInsight<C, M, R>

pub fn delete_invitations(&self) -> DeleteInvitations<C, M, R>

pub fn delete_members(&self) -> DeleteMembers<C, M, R>

pub fn describe_action_targets(&self) -> DescribeActionTargets<C, M, R>

pub fn describe_hub(&self) -> DescribeHub<C, M, R>

pub fn describe_organization_configuration(
&self
) -> DescribeOrganizationConfiguration<C, M, R>

pub fn describe_products(&self) -> DescribeProducts<C, M, R>

pub fn describe_standards(&self) -> DescribeStandards<C, M, R>

pub fn describe_standards_controls(&self) -> DescribeStandardsControls<C, M, R>

pub fn disable_import_findings_for_product(
&self
) -> DisableImportFindingsForProduct<C, M, R>

pub fn disable_organization_admin_account(
&self
) -> DisableOrganizationAdminAccount<C, M, R>

pub fn disable_security_hub(&self) -> DisableSecurityHub<C, M, R>

pub fn disassociate_from_administrator_account(
&self
) -> DisassociateFromAdministratorAccount<C, M, R>

pub fn disassociate_from_master_account(
&self
) -> DisassociateFromMasterAccount<C, M, R>

pub fn disassociate_members(&self) -> DisassociateMembers<C, M, R>

pub fn enable_import_findings_for_product(
&self
) -> EnableImportFindingsForProduct<C, M, R>

pub fn enable_organization_admin_account(
&self
) -> EnableOrganizationAdminAccount<C, M, R>

pub fn enable_security_hub(&self) -> EnableSecurityHub<C, M, R>

pub fn get_administrator_account(&self) -> GetAdministratorAccount<C, M, R>

pub fn get_enabled_standards(&self) -> GetEnabledStandards<C, M, R>

pub fn get_findings(&self) -> GetFindings<C, M, R>

pub fn get_insight_results(&self) -> GetInsightResults<C, M, R>

pub fn get_insights(&self) -> GetInsights<C, M, R>

pub fn get_invitations_count(&self) -> GetInvitationsCount<C, M, R>

pub fn get_master_account(&self) -> GetMasterAccount<C, M, R>

pub fn get_members(&self) -> GetMembers<C, M, R>

pub fn invite_members(&self) -> InviteMembers<C, M, R>

pub fn list_enabled_products_for_import(
&self
) -> ListEnabledProductsForImport<C, M, R>

pub fn list_invitations(&self) -> ListInvitations<C, M, R>

pub fn list_members(&self) -> ListMembers<C, M, R>

pub fn list_organization_admin_accounts(
&self
) -> ListOrganizationAdminAccounts<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_action_target(&self) -> UpdateActionTarget<C, M, R>

pub fn update_findings(&self) -> UpdateFindings<C, M, R>

pub fn update_insight(&self) -> UpdateInsight<C, M, R>

pub fn update_organization_configuration(
&self
) -> UpdateOrganizationConfiguration<C, M, R>

pub fn update_security_hub_configuration(
&self
) -> UpdateSecurityHubConfiguration<C, M, R>

pub fn update_standards_control(&self) -> UpdateStandardsControl<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self