Module aws_sdk_securityhub::error

Modules

access_denied_exception

See AccessDeniedException

internal_exception

See InternalException

invalid_access_exception

See InvalidAccessException

invalid_input_exception

See InvalidInputException

limit_exceeded_exception

See LimitExceededException

resource_conflict_exception

See ResourceConflictException

resource_not_found_exception

See ResourceNotFoundException

Structs

AcceptAdministratorInvitationError

AcceptInvitationError

AccessDeniedException

You don't have permission to perform the action specified in the request.

BatchDisableStandardsError

BatchEnableStandardsError

BatchImportFindingsError

BatchUpdateFindingsError

CreateActionTargetError

CreateInsightError

CreateMembersError

DeclineInvitationsError

DeleteActionTargetError

DeleteInsightError

DeleteInvitationsError

DeleteMembersError

DescribeActionTargetsError

DescribeHubError

DescribeOrganizationConfigurationError

DescribeProductsError

DescribeStandardsControlsError

DescribeStandardsError

DisableImportFindingsForProductError

DisableOrganizationAdminAccountError

DisableSecurityHubError

DisassociateFromAdministratorAccountError

DisassociateFromMasterAccountError

DisassociateMembersError

EnableImportFindingsForProductError

EnableOrganizationAdminAccountError

EnableSecurityHubError

GetAdministratorAccountError

GetEnabledStandardsError

GetFindingsError

GetInsightResultsError

GetInsightsError

GetInvitationsCountError

GetMasterAccountError

GetMembersError

InternalException

Internal server error.

InvalidAccessException

There is an issue with the account used to make the request. Either Security Hub is not enabled

InvalidInputException

The request was rejected because you supplied an invalid or out-of-range value for an

InviteMembersError

LimitExceededException

The request was rejected because it attempted to create resources beyond the current Amazon Web Services

ListEnabledProductsForImportError

ListInvitationsError

ListMembersError

ListOrganizationAdminAccountsError

ListTagsForResourceError

ResourceConflictException

The resource specified in the request conflicts with an existing resource.

ResourceNotFoundException

The request was rejected because we can't find the specified resource.

TagResourceError

UntagResourceError

UpdateActionTargetError

UpdateFindingsError

UpdateInsightError

UpdateOrganizationConfigurationError

UpdateSecurityHubConfigurationError

UpdateStandardsControlError

Enums

AcceptAdministratorInvitationErrorKind

AcceptInvitationErrorKind

BatchDisableStandardsErrorKind

BatchEnableStandardsErrorKind

BatchImportFindingsErrorKind

BatchUpdateFindingsErrorKind

CreateActionTargetErrorKind

CreateInsightErrorKind

CreateMembersErrorKind

DeclineInvitationsErrorKind

DeleteActionTargetErrorKind

DeleteInsightErrorKind

DeleteInvitationsErrorKind

DeleteMembersErrorKind

DescribeActionTargetsErrorKind

DescribeHubErrorKind

DescribeOrganizationConfigurationErrorKind

DescribeProductsErrorKind

DescribeStandardsControlsErrorKind

DescribeStandardsErrorKind

DisableImportFindingsForProductErrorKind

DisableOrganizationAdminAccountErrorKind

DisableSecurityHubErrorKind

DisassociateFromAdministratorAccountErrorKind

DisassociateFromMasterAccountErrorKind

DisassociateMembersErrorKind

EnableImportFindingsForProductErrorKind

EnableOrganizationAdminAccountErrorKind

EnableSecurityHubErrorKind

GetAdministratorAccountErrorKind

GetEnabledStandardsErrorKind

GetFindingsErrorKind

GetInsightResultsErrorKind

GetInsightsErrorKind

GetInvitationsCountErrorKind

GetMasterAccountErrorKind

GetMembersErrorKind

InviteMembersErrorKind

ListEnabledProductsForImportErrorKind

ListInvitationsErrorKind

ListMembersErrorKind

ListOrganizationAdminAccountsErrorKind

ListTagsForResourceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateActionTargetErrorKind

UpdateFindingsErrorKind

UpdateInsightErrorKind

UpdateOrganizationConfigurationErrorKind

UpdateSecurityHubConfigurationErrorKind

UpdateStandardsControlErrorKind