Module aws_sdk_securityhub::model

Modules

account_details

See AccountDetails

action

See Action

action_local_ip_details

See ActionLocalIpDetails

action_local_port_details

See ActionLocalPortDetails

action_remote_ip_details

See ActionRemoteIpDetails

action_remote_port_details

See ActionRemotePortDetails

action_target

See ActionTarget

adjustment

See Adjustment

admin_account

See AdminAccount

availability_zone

See AvailabilityZone

aws_api_call_action

See AwsApiCallAction

aws_api_call_action_domain_details

See AwsApiCallActionDomainDetails

aws_api_gateway_access_log_settings

See AwsApiGatewayAccessLogSettings

aws_api_gateway_canary_settings

See AwsApiGatewayCanarySettings

aws_api_gateway_endpoint_configuration

See AwsApiGatewayEndpointConfiguration

aws_api_gateway_method_settings

See AwsApiGatewayMethodSettings

aws_api_gateway_rest_api_details

See AwsApiGatewayRestApiDetails

aws_api_gateway_stage_details

See AwsApiGatewayStageDetails

aws_api_gateway_v2_api_details

See AwsApiGatewayV2ApiDetails

aws_api_gateway_v2_route_settings

See AwsApiGatewayV2RouteSettings

aws_api_gateway_v2_stage_details

See AwsApiGatewayV2StageDetails

aws_auto_scaling_auto_scaling_group_details

See AwsAutoScalingAutoScalingGroupDetails

aws_auto_scaling_launch_configuration_block_device_mappings_details

See AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails

aws_auto_scaling_launch_configuration_block_device_mappings_ebs_details

See AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

aws_auto_scaling_launch_configuration_details

See AwsAutoScalingLaunchConfigurationDetails

aws_auto_scaling_launch_configuration_instance_monitoring_details

See AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails

aws_certificate_manager_certificate_details

See AwsCertificateManagerCertificateDetails

aws_certificate_manager_certificate_domain_validation_option

See AwsCertificateManagerCertificateDomainValidationOption

aws_certificate_manager_certificate_extended_key_usage

See AwsCertificateManagerCertificateExtendedKeyUsage

aws_certificate_manager_certificate_key_usage

See AwsCertificateManagerCertificateKeyUsage

aws_certificate_manager_certificate_options

See AwsCertificateManagerCertificateOptions

aws_certificate_manager_certificate_renewal_summary

See AwsCertificateManagerCertificateRenewalSummary

aws_certificate_manager_certificate_resource_record

See AwsCertificateManagerCertificateResourceRecord

aws_cloud_front_distribution_cache_behavior

See AwsCloudFrontDistributionCacheBehavior

aws_cloud_front_distribution_cache_behaviors

See AwsCloudFrontDistributionCacheBehaviors

aws_cloud_front_distribution_default_cache_behavior

See AwsCloudFrontDistributionDefaultCacheBehavior

aws_cloud_front_distribution_details

See AwsCloudFrontDistributionDetails

aws_cloud_front_distribution_logging

See AwsCloudFrontDistributionLogging

aws_cloud_front_distribution_origin_group

See AwsCloudFrontDistributionOriginGroup

aws_cloud_front_distribution_origin_group_failover

See AwsCloudFrontDistributionOriginGroupFailover

aws_cloud_front_distribution_origin_group_failover_status_codes

See AwsCloudFrontDistributionOriginGroupFailoverStatusCodes

aws_cloud_front_distribution_origin_groups

See AwsCloudFrontDistributionOriginGroups

aws_cloud_front_distribution_origin_item

See AwsCloudFrontDistributionOriginItem

aws_cloud_front_distribution_origin_s3_origin_config

See AwsCloudFrontDistributionOriginS3OriginConfig

aws_cloud_front_distribution_origins

See AwsCloudFrontDistributionOrigins

aws_cloud_trail_trail_details

See AwsCloudTrailTrailDetails

aws_code_build_project_details

See AwsCodeBuildProjectDetails

aws_code_build_project_environment

See AwsCodeBuildProjectEnvironment

aws_code_build_project_environment_registry_credential

See AwsCodeBuildProjectEnvironmentRegistryCredential

aws_code_build_project_source

See AwsCodeBuildProjectSource

aws_code_build_project_vpc_config

See AwsCodeBuildProjectVpcConfig

aws_cors_configuration

See AwsCorsConfiguration

aws_dynamo_db_table_attribute_definition

See AwsDynamoDbTableAttributeDefinition

aws_dynamo_db_table_billing_mode_summary

See AwsDynamoDbTableBillingModeSummary

aws_dynamo_db_table_details

See AwsDynamoDbTableDetails

aws_dynamo_db_table_global_secondary_index

See AwsDynamoDbTableGlobalSecondaryIndex

aws_dynamo_db_table_key_schema

See AwsDynamoDbTableKeySchema

aws_dynamo_db_table_local_secondary_index

See AwsDynamoDbTableLocalSecondaryIndex

aws_dynamo_db_table_projection

See AwsDynamoDbTableProjection

aws_dynamo_db_table_provisioned_throughput

See AwsDynamoDbTableProvisionedThroughput

aws_dynamo_db_table_provisioned_throughput_override

See AwsDynamoDbTableProvisionedThroughputOverride

aws_dynamo_db_table_replica

See AwsDynamoDbTableReplica

aws_dynamo_db_table_replica_global_secondary_index

See AwsDynamoDbTableReplicaGlobalSecondaryIndex

aws_dynamo_db_table_restore_summary

See AwsDynamoDbTableRestoreSummary

aws_dynamo_db_table_sse_description

See AwsDynamoDbTableSseDescription

aws_dynamo_db_table_stream_specification

See AwsDynamoDbTableStreamSpecification

aws_ec2_eip_details

See AwsEc2EipDetails

aws_ec2_instance_details

See AwsEc2InstanceDetails

aws_ec2_instance_network_interfaces_details

See AwsEc2InstanceNetworkInterfacesDetails

aws_ec2_network_acl_association

See AwsEc2NetworkAclAssociation

aws_ec2_network_acl_details

See AwsEc2NetworkAclDetails

aws_ec2_network_acl_entry

See AwsEc2NetworkAclEntry

aws_ec2_network_interface_attachment

See AwsEc2NetworkInterfaceAttachment

aws_ec2_network_interface_details

See AwsEc2NetworkInterfaceDetails

aws_ec2_network_interface_ip_v6_address_detail

See AwsEc2NetworkInterfaceIpV6AddressDetail

aws_ec2_network_interface_private_ip_address_detail

See AwsEc2NetworkInterfacePrivateIpAddressDetail

aws_ec2_network_interface_security_group

See AwsEc2NetworkInterfaceSecurityGroup

aws_ec2_security_group_details

See AwsEc2SecurityGroupDetails

aws_ec2_security_group_ip_permission

See AwsEc2SecurityGroupIpPermission

aws_ec2_security_group_ip_range

See AwsEc2SecurityGroupIpRange

aws_ec2_security_group_ipv6_range

See AwsEc2SecurityGroupIpv6Range

aws_ec2_security_group_prefix_list_id

See AwsEc2SecurityGroupPrefixListId

aws_ec2_security_group_user_id_group_pair

See AwsEc2SecurityGroupUserIdGroupPair

aws_ec2_subnet_details

See AwsEc2SubnetDetails

aws_ec2_volume_attachment

See AwsEc2VolumeAttachment

aws_ec2_volume_details

See AwsEc2VolumeDetails

aws_ec2_vpc_details

See AwsEc2VpcDetails

aws_ec2_vpn_connection_details

See AwsEc2VpnConnectionDetails

aws_ec2_vpn_connection_options_details

See AwsEc2VpnConnectionOptionsDetails

aws_ec2_vpn_connection_options_tunnel_options_details

See AwsEc2VpnConnectionOptionsTunnelOptionsDetails

aws_ec2_vpn_connection_routes_details

See AwsEc2VpnConnectionRoutesDetails

aws_ec2_vpn_connection_vgw_telemetry_details

See AwsEc2VpnConnectionVgwTelemetryDetails

aws_ecr_container_image_details

See AwsEcrContainerImageDetails

aws_ecs_cluster_cluster_settings_details

See AwsEcsClusterClusterSettingsDetails

aws_ecs_cluster_configuration_details

See AwsEcsClusterConfigurationDetails

aws_ecs_cluster_configuration_execute_command_configuration_details

See AwsEcsClusterConfigurationExecuteCommandConfigurationDetails

aws_ecs_cluster_configuration_execute_command_configuration_log_configuration_details

See AwsEcsClusterConfigurationExecuteCommandConfigurationLogConfigurationDetails

aws_ecs_cluster_default_capacity_provider_strategy_details

See AwsEcsClusterDefaultCapacityProviderStrategyDetails

aws_ecs_cluster_details

See AwsEcsClusterDetails

aws_ecs_service_capacity_provider_strategy_details

See AwsEcsServiceCapacityProviderStrategyDetails

aws_ecs_service_deployment_configuration_deployment_circuit_breaker_details

See AwsEcsServiceDeploymentConfigurationDeploymentCircuitBreakerDetails

aws_ecs_service_deployment_configuration_details

See AwsEcsServiceDeploymentConfigurationDetails

aws_ecs_service_deployment_controller_details

See AwsEcsServiceDeploymentControllerDetails

aws_ecs_service_details

See AwsEcsServiceDetails

aws_ecs_service_load_balancers_details

See AwsEcsServiceLoadBalancersDetails

aws_ecs_service_network_configuration_aws_vpc_configuration_details

See AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails

aws_ecs_service_network_configuration_details

See AwsEcsServiceNetworkConfigurationDetails

aws_ecs_service_placement_constraints_details

See AwsEcsServicePlacementConstraintsDetails

aws_ecs_service_placement_strategies_details

See AwsEcsServicePlacementStrategiesDetails

aws_ecs_service_service_registries_details

See AwsEcsServiceServiceRegistriesDetails

aws_ecs_task_definition_container_definitions_depends_on_details

See AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails

aws_ecs_task_definition_container_definitions_details

See AwsEcsTaskDefinitionContainerDefinitionsDetails

aws_ecs_task_definition_container_definitions_environment_details

See AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails

aws_ecs_task_definition_container_definitions_environment_files_details

See AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails

aws_ecs_task_definition_container_definitions_extra_hosts_details

See AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails

aws_ecs_task_definition_container_definitions_firelens_configuration_details

See AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails

aws_ecs_task_definition_container_definitions_health_check_details

See AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

aws_ecs_task_definition_container_definitions_linux_parameters_capabilities_details

See AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersCapabilitiesDetails

aws_ecs_task_definition_container_definitions_linux_parameters_details

See AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

aws_ecs_task_definition_container_definitions_linux_parameters_devices_details

See AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDevicesDetails

aws_ecs_task_definition_container_definitions_linux_parameters_tmpfs_details

See AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersTmpfsDetails

aws_ecs_task_definition_container_definitions_log_configuration_details

See AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

aws_ecs_task_definition_container_definitions_log_configuration_secret_options_details

See AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationSecretOptionsDetails

aws_ecs_task_definition_container_definitions_mount_points_details

See AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails

aws_ecs_task_definition_container_definitions_port_mappings_details

See AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails

aws_ecs_task_definition_container_definitions_repository_credentials_details

See AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails

aws_ecs_task_definition_container_definitions_resource_requirements_details

See AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails

aws_ecs_task_definition_container_definitions_secrets_details

See AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails

aws_ecs_task_definition_container_definitions_system_controls_details

See AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails

aws_ecs_task_definition_container_definitions_ulimits_details

See AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails

aws_ecs_task_definition_container_definitions_volumes_from_details

See AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails

aws_ecs_task_definition_details

See AwsEcsTaskDefinitionDetails

aws_ecs_task_definition_inference_accelerators_details

See AwsEcsTaskDefinitionInferenceAcceleratorsDetails

aws_ecs_task_definition_placement_constraints_details

See AwsEcsTaskDefinitionPlacementConstraintsDetails

aws_ecs_task_definition_proxy_configuration_details

See AwsEcsTaskDefinitionProxyConfigurationDetails

aws_ecs_task_definition_proxy_configuration_proxy_configuration_properties_details

See AwsEcsTaskDefinitionProxyConfigurationProxyConfigurationPropertiesDetails

aws_ecs_task_definition_volumes_details

See AwsEcsTaskDefinitionVolumesDetails

aws_ecs_task_definition_volumes_docker_volume_configuration_details

See AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

aws_ecs_task_definition_volumes_efs_volume_configuration_authorization_config_details

See AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationAuthorizationConfigDetails

aws_ecs_task_definition_volumes_efs_volume_configuration_details

See AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationDetails

aws_ecs_task_definition_volumes_host_details

See AwsEcsTaskDefinitionVolumesHostDetails

aws_elastic_beanstalk_environment_details

See AwsElasticBeanstalkEnvironmentDetails

aws_elastic_beanstalk_environment_environment_link

See AwsElasticBeanstalkEnvironmentEnvironmentLink

aws_elastic_beanstalk_environment_option_setting

See AwsElasticBeanstalkEnvironmentOptionSetting

aws_elastic_beanstalk_environment_tier

See AwsElasticBeanstalkEnvironmentTier

aws_elasticsearch_domain_details

See AwsElasticsearchDomainDetails

aws_elasticsearch_domain_domain_endpoint_options

See AwsElasticsearchDomainDomainEndpointOptions

aws_elasticsearch_domain_elasticsearch_cluster_config_details

See AwsElasticsearchDomainElasticsearchClusterConfigDetails

aws_elasticsearch_domain_elasticsearch_cluster_config_zone_awareness_config_details

See AwsElasticsearchDomainElasticsearchClusterConfigZoneAwarenessConfigDetails

aws_elasticsearch_domain_encryption_at_rest_options

See AwsElasticsearchDomainEncryptionAtRestOptions

aws_elasticsearch_domain_log_publishing_options

See AwsElasticsearchDomainLogPublishingOptions

aws_elasticsearch_domain_log_publishing_options_log_config

See AwsElasticsearchDomainLogPublishingOptionsLogConfig

aws_elasticsearch_domain_node_to_node_encryption_options

See AwsElasticsearchDomainNodeToNodeEncryptionOptions

aws_elasticsearch_domain_service_software_options

See AwsElasticsearchDomainServiceSoftwareOptions

aws_elasticsearch_domain_vpc_options

See AwsElasticsearchDomainVpcOptions

aws_elb_app_cookie_stickiness_policy

See AwsElbAppCookieStickinessPolicy

aws_elb_lb_cookie_stickiness_policy

See AwsElbLbCookieStickinessPolicy

aws_elb_load_balancer_access_log

See AwsElbLoadBalancerAccessLog

aws_elb_load_balancer_attributes

See AwsElbLoadBalancerAttributes

aws_elb_load_balancer_backend_server_description

See AwsElbLoadBalancerBackendServerDescription

aws_elb_load_balancer_connection_draining

See AwsElbLoadBalancerConnectionDraining

aws_elb_load_balancer_connection_settings

See AwsElbLoadBalancerConnectionSettings

aws_elb_load_balancer_cross_zone_load_balancing

See AwsElbLoadBalancerCrossZoneLoadBalancing

aws_elb_load_balancer_details

See AwsElbLoadBalancerDetails

aws_elb_load_balancer_health_check

See AwsElbLoadBalancerHealthCheck

aws_elb_load_balancer_instance

See AwsElbLoadBalancerInstance

aws_elb_load_balancer_listener

See AwsElbLoadBalancerListener

aws_elb_load_balancer_listener_description

See AwsElbLoadBalancerListenerDescription

aws_elb_load_balancer_policies

See AwsElbLoadBalancerPolicies

aws_elb_load_balancer_source_security_group

See AwsElbLoadBalancerSourceSecurityGroup

aws_elbv2_load_balancer_details

See AwsElbv2LoadBalancerDetails

aws_iam_access_key_details

See AwsIamAccessKeyDetails

aws_iam_access_key_session_context

See AwsIamAccessKeySessionContext

aws_iam_access_key_session_context_attributes

See AwsIamAccessKeySessionContextAttributes

aws_iam_access_key_session_context_session_issuer

See AwsIamAccessKeySessionContextSessionIssuer

aws_iam_attached_managed_policy

See AwsIamAttachedManagedPolicy

aws_iam_group_details

See AwsIamGroupDetails

aws_iam_group_policy

See AwsIamGroupPolicy

aws_iam_instance_profile

See AwsIamInstanceProfile

aws_iam_instance_profile_role

See AwsIamInstanceProfileRole

aws_iam_permissions_boundary

See AwsIamPermissionsBoundary

aws_iam_policy_details

See AwsIamPolicyDetails

aws_iam_policy_version

See AwsIamPolicyVersion

aws_iam_role_details

See AwsIamRoleDetails

aws_iam_role_policy

See AwsIamRolePolicy

aws_iam_user_details

See AwsIamUserDetails

aws_iam_user_policy

See AwsIamUserPolicy

aws_kms_key_details

See AwsKmsKeyDetails

aws_lambda_function_code

See AwsLambdaFunctionCode

aws_lambda_function_dead_letter_config

See AwsLambdaFunctionDeadLetterConfig

aws_lambda_function_details

See AwsLambdaFunctionDetails

aws_lambda_function_environment

See AwsLambdaFunctionEnvironment

aws_lambda_function_environment_error

See AwsLambdaFunctionEnvironmentError

aws_lambda_function_layer

See AwsLambdaFunctionLayer

aws_lambda_function_tracing_config

See AwsLambdaFunctionTracingConfig

aws_lambda_function_vpc_config

See AwsLambdaFunctionVpcConfig

aws_lambda_layer_version_details

See AwsLambdaLayerVersionDetails

aws_rds_db_cluster_associated_role

See AwsRdsDbClusterAssociatedRole

aws_rds_db_cluster_details

See AwsRdsDbClusterDetails

aws_rds_db_cluster_member

See AwsRdsDbClusterMember

aws_rds_db_cluster_option_group_membership

See AwsRdsDbClusterOptionGroupMembership

aws_rds_db_cluster_snapshot_details

See AwsRdsDbClusterSnapshotDetails

aws_rds_db_domain_membership

See AwsRdsDbDomainMembership

aws_rds_db_instance_associated_role

See AwsRdsDbInstanceAssociatedRole

aws_rds_db_instance_details

See AwsRdsDbInstanceDetails

aws_rds_db_instance_endpoint

See AwsRdsDbInstanceEndpoint

aws_rds_db_instance_vpc_security_group

See AwsRdsDbInstanceVpcSecurityGroup

aws_rds_db_option_group_membership

See AwsRdsDbOptionGroupMembership

aws_rds_db_parameter_group

See AwsRdsDbParameterGroup

aws_rds_db_pending_modified_values

See AwsRdsDbPendingModifiedValues

aws_rds_db_processor_feature

See AwsRdsDbProcessorFeature

aws_rds_db_snapshot_details

See AwsRdsDbSnapshotDetails

aws_rds_db_status_info

See AwsRdsDbStatusInfo

aws_rds_db_subnet_group

See AwsRdsDbSubnetGroup

aws_rds_db_subnet_group_subnet

See AwsRdsDbSubnetGroupSubnet

aws_rds_db_subnet_group_subnet_availability_zone

See AwsRdsDbSubnetGroupSubnetAvailabilityZone

aws_rds_event_subscription_details

See AwsRdsEventSubscriptionDetails

aws_rds_pending_cloud_watch_logs_exports

See AwsRdsPendingCloudWatchLogsExports

aws_redshift_cluster_cluster_node

See AwsRedshiftClusterClusterNode

aws_redshift_cluster_cluster_parameter_group

See AwsRedshiftClusterClusterParameterGroup

aws_redshift_cluster_cluster_parameter_status

See AwsRedshiftClusterClusterParameterStatus

aws_redshift_cluster_cluster_security_group

See AwsRedshiftClusterClusterSecurityGroup

aws_redshift_cluster_cluster_snapshot_copy_status

See AwsRedshiftClusterClusterSnapshotCopyStatus

aws_redshift_cluster_deferred_maintenance_window

See AwsRedshiftClusterDeferredMaintenanceWindow

aws_redshift_cluster_details

See AwsRedshiftClusterDetails

aws_redshift_cluster_elastic_ip_status

See AwsRedshiftClusterElasticIpStatus

aws_redshift_cluster_endpoint

See AwsRedshiftClusterEndpoint

aws_redshift_cluster_hsm_status

See AwsRedshiftClusterHsmStatus

aws_redshift_cluster_iam_role

See AwsRedshiftClusterIamRole

aws_redshift_cluster_pending_modified_values

See AwsRedshiftClusterPendingModifiedValues

aws_redshift_cluster_resize_info

See AwsRedshiftClusterResizeInfo

aws_redshift_cluster_restore_status

See AwsRedshiftClusterRestoreStatus

aws_redshift_cluster_vpc_security_group

See AwsRedshiftClusterVpcSecurityGroup

aws_s3_account_public_access_block_details

See AwsS3AccountPublicAccessBlockDetails

aws_s3_bucket_bucket_lifecycle_configuration_details

See AwsS3BucketBucketLifecycleConfigurationDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_abort_incomplete_multipart_upload_details

See AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_details

See AwsS3BucketBucketLifecycleConfigurationRulesDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_filter_details

See AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_filter_predicate_details

See AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_filter_predicate_operands_details

See AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateOperandsDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_filter_predicate_operands_tag_details

See AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateOperandsTagDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_filter_predicate_tag_details

See AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateTagDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_noncurrent_version_transitions_details

See AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails

aws_s3_bucket_bucket_lifecycle_configuration_rules_transitions_details

See AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails

aws_s3_bucket_details

See AwsS3BucketDetails

aws_s3_bucket_logging_configuration

See AwsS3BucketLoggingConfiguration

aws_s3_bucket_notification_configuration

See AwsS3BucketNotificationConfiguration

aws_s3_bucket_notification_configuration_detail

See AwsS3BucketNotificationConfigurationDetail

aws_s3_bucket_notification_configuration_filter

See AwsS3BucketNotificationConfigurationFilter

aws_s3_bucket_notification_configuration_s3_key_filter

See AwsS3BucketNotificationConfigurationS3KeyFilter

aws_s3_bucket_notification_configuration_s3_key_filter_rule

See AwsS3BucketNotificationConfigurationS3KeyFilterRule

aws_s3_bucket_server_side_encryption_by_default

See AwsS3BucketServerSideEncryptionByDefault

aws_s3_bucket_server_side_encryption_configuration

See AwsS3BucketServerSideEncryptionConfiguration

aws_s3_bucket_server_side_encryption_rule

See AwsS3BucketServerSideEncryptionRule

aws_s3_bucket_website_configuration

See AwsS3BucketWebsiteConfiguration

aws_s3_bucket_website_configuration_redirect_to

See AwsS3BucketWebsiteConfigurationRedirectTo

aws_s3_bucket_website_configuration_routing_rule

See AwsS3BucketWebsiteConfigurationRoutingRule

aws_s3_bucket_website_configuration_routing_rule_condition

See AwsS3BucketWebsiteConfigurationRoutingRuleCondition

aws_s3_bucket_website_configuration_routing_rule_redirect

See AwsS3BucketWebsiteConfigurationRoutingRuleRedirect

aws_s3_object_details

See AwsS3ObjectDetails

aws_secrets_manager_secret_details

See AwsSecretsManagerSecretDetails

aws_secrets_manager_secret_rotation_rules

See AwsSecretsManagerSecretRotationRules

aws_security_finding

See AwsSecurityFinding

aws_security_finding_filters

See AwsSecurityFindingFilters

aws_security_finding_identifier

See AwsSecurityFindingIdentifier

aws_sns_topic_details

See AwsSnsTopicDetails

aws_sns_topic_subscription

See AwsSnsTopicSubscription

aws_sqs_queue_details

See AwsSqsQueueDetails

aws_ssm_compliance_summary

See AwsSsmComplianceSummary

aws_ssm_patch

See AwsSsmPatch

aws_ssm_patch_compliance_details

See AwsSsmPatchComplianceDetails

aws_waf_web_acl_details

See AwsWafWebAclDetails

aws_waf_web_acl_rule

See AwsWafWebAclRule

batch_update_findings_unprocessed_finding

See BatchUpdateFindingsUnprocessedFinding

cell

See Cell

cidr_block_association

See CidrBlockAssociation

city

See City

classification_result

See ClassificationResult

classification_status

See ClassificationStatus

compliance

See Compliance

container_details

See ContainerDetails

country

See Country

custom_data_identifiers_detections

See CustomDataIdentifiersDetections

custom_data_identifiers_result

See CustomDataIdentifiersResult

cvss

See Cvss

data_classification_details

See DataClassificationDetails

date_filter

See DateFilter

date_range

See DateRange

dns_request_action

See DnsRequestAction

finding_provider_fields

See FindingProviderFields

finding_provider_severity

See FindingProviderSeverity

geo_location

See GeoLocation

icmp_type_code

See IcmpTypeCode

import_findings_error

See ImportFindingsError

insight

See Insight

insight_result_value

See InsightResultValue

insight_results

See InsightResults

invitation

See Invitation

ip_filter

See IpFilter

ip_organization_details

See IpOrganizationDetails

ipv6_cidr_block_association

See Ipv6CidrBlockAssociation

keyword_filter

See KeywordFilter

load_balancer_state

See LoadBalancerState

malware

See Malware

map_filter

See MapFilter

member

See Member

network

See Network

network_connection_action

See NetworkConnectionAction

network_header

See NetworkHeader

network_path_component

See NetworkPathComponent

network_path_component_details

See NetworkPathComponentDetails

note

See Note

note_update

See NoteUpdate

number_filter

See NumberFilter

occurrences

See Occurrences

page

See Page

patch_summary

See PatchSummary

port_probe_action

See PortProbeAction

port_probe_detail

See PortProbeDetail

port_range

See PortRange

port_range_from_to

See PortRangeFromTo

process_details

See ProcessDetails

product

See Product

range

See Range

recommendation

See Recommendation

record

See Record

related_finding

See RelatedFinding

remediation

See Remediation

resource

See Resource

resource_details

See ResourceDetails

result

See Result

sensitive_data_detections

See SensitiveDataDetections

sensitive_data_result

See SensitiveDataResult

severity

See Severity

severity_update

See SeverityUpdate

software_package

See SoftwarePackage

sort_criterion

See SortCriterion

standard

See Standard

standards_control

See StandardsControl

standards_subscription

See StandardsSubscription

standards_subscription_request

See StandardsSubscriptionRequest

status_reason

See StatusReason

string_filter

See StringFilter

threat_intel_indicator

See ThreatIntelIndicator

vulnerability

See Vulnerability

vulnerability_vendor

See VulnerabilityVendor

waf_action

See WafAction

waf_excluded_rule

See WafExcludedRule

waf_override_action

See WafOverrideAction

workflow

See Workflow

workflow_update

See WorkflowUpdate

Structs

AccountDetails

The details of an Amazon Web Services account.

Action

Provides details about one of the following actions that affects or that was taken on a resource:

ActionLocalIpDetails

Provides information about the IP address where the scanned port is located.

ActionLocalPortDetails

For NetworkConnectionAction and PortProbeDetails,

ActionRemoteIpDetails

For AwsApiAction, NetworkConnectionAction, and

ActionRemotePortDetails

Provides information about the remote port that was involved in an attempted network

ActionTarget

An ActionTarget object.

Adjustment

An adjustment to the CVSS metric.

AdminAccount

Represents a Security Hub administrator account designated by an organization management

AvailabilityZone

Information about an Availability Zone.

AwsApiCallAction

Provided if ActionType is AWS_API_CALL. It provides details

AwsApiCallActionDomainDetails

Provided if CallerType is domain. It provides information

AwsApiGatewayAccessLogSettings

Contains information about settings for logging access for the stage.

AwsApiGatewayCanarySettings

Contains information about settings for canary deployment in the stage.

AwsApiGatewayEndpointConfiguration

Contains information about the endpoints for the API.

AwsApiGatewayMethodSettings

Defines settings for a method for the stage.

AwsApiGatewayRestApiDetails

Contains information about a REST API in version 1 of Amazon API Gateway.

AwsApiGatewayStageDetails

Provides information about a version 1 Amazon API Gateway stage.

AwsApiGatewayV2ApiDetails

Contains information about a version 2 API in Amazon API Gateway.

AwsApiGatewayV2RouteSettings

Contains route settings for a stage.

AwsApiGatewayV2StageDetails

Contains information about a version 2 stage for Amazon API Gateway.

AwsAutoScalingAutoScalingGroupDetails

Provides details about an auto scaling group.

AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails

A block device for the instance.

AwsAutoScalingLaunchConfigurationBlockDeviceMappingsEbsDetails

Parameters that are used to automatically set up EBS volumes when an instance is launched.

AwsAutoScalingLaunchConfigurationDetails

Details about a launch configuration.

AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails

Information about the type of monitoring for instances in the group.

AwsCertificateManagerCertificateDetails

Provides details about an Certificate Manager certificate.

AwsCertificateManagerCertificateDomainValidationOption

Contains information about one of the following:

AwsCertificateManagerCertificateExtendedKeyUsage

Contains information about an extended key usage X.509 v3 extension object.

AwsCertificateManagerCertificateKeyUsage

Contains information about a key usage X.509 v3 extension object.

AwsCertificateManagerCertificateOptions

Contains other options for the certificate.

AwsCertificateManagerCertificateRenewalSummary

Contains information about the Certificate Manager managed renewal for an

AwsCertificateManagerCertificateResourceRecord

Provides details about the CNAME record that is added to the DNS database for domain

AwsCloudFrontDistributionCacheBehavior

Information about a cache behavior for the distribution.

AwsCloudFrontDistributionCacheBehaviors

Provides information about caching for the distribution.

AwsCloudFrontDistributionDefaultCacheBehavior

Contains information about the default cache configuration for the distribution.

AwsCloudFrontDistributionDetails

A distribution configuration.

AwsCloudFrontDistributionLogging

A complex type that controls whether access logs are written for the distribution.

AwsCloudFrontDistributionOriginGroup

Information about an origin group for the distribution.

AwsCloudFrontDistributionOriginGroupFailover

Provides information about when an origin group fails over.

AwsCloudFrontDistributionOriginGroupFailoverStatusCodes

The status codes that cause an origin group to fail over.

AwsCloudFrontDistributionOriginGroups

Provides information about origin groups that are associated with the

AwsCloudFrontDistributionOriginItem

A complex type that describes the S3 bucket, HTTP server (for example, a web

AwsCloudFrontDistributionOriginS3OriginConfig

Information about an origin that is an S3 bucket that is not configured with static

AwsCloudFrontDistributionOrigins

A complex type that contains information about origins and origin groups for this distribution.

AwsCloudTrailTrailDetails

Provides details about a CloudTrail trail.

AwsCodeBuildProjectDetails

Information about an CodeBuild project.

AwsCodeBuildProjectEnvironment

Information about the build environment for this build project.

AwsCodeBuildProjectEnvironmentRegistryCredential

The credentials for access to a private registry.

AwsCodeBuildProjectSource

Information about the build input source code for this build project.

AwsCodeBuildProjectVpcConfig

Information about the VPC configuration that CodeBuild accesses.

AwsCorsConfiguration

Contains the cross-origin resource sharing (CORS) configuration for the API. CORS is

AwsDynamoDbTableAttributeDefinition

Contains a definition of an attribute for the table.

AwsDynamoDbTableBillingModeSummary

Provides information about the billing for read/write capacity on the table.

AwsDynamoDbTableDetails

Provides details about a DynamoDB table.

AwsDynamoDbTableGlobalSecondaryIndex

Information abut a global secondary index for the table.

AwsDynamoDbTableKeySchema

A component of the key schema for the DynamoDB table, a global secondary index, or a

AwsDynamoDbTableLocalSecondaryIndex

Information about a local secondary index for a DynamoDB table.

AwsDynamoDbTableProjection

For global and local secondary indexes, identifies the attributes that are copied from

AwsDynamoDbTableProvisionedThroughput

Information about the provisioned throughput for the table or for a global secondary

AwsDynamoDbTableProvisionedThroughputOverride

Replica-specific configuration for the provisioned throughput.

AwsDynamoDbTableReplica

Information about a replica of a DynamoDB table.

AwsDynamoDbTableReplicaGlobalSecondaryIndex

Information about a global secondary index for a DynamoDB table replica.

AwsDynamoDbTableRestoreSummary

Information about the restore for the table.

AwsDynamoDbTableSseDescription

Information about the server-side encryption for the table.

AwsDynamoDbTableStreamSpecification

The current DynamoDB Streams configuration for the table.

AwsEc2EipDetails

Information about an Elastic IP address.

AwsEc2InstanceDetails

The details of an EC2 instance.

AwsEc2InstanceNetworkInterfacesDetails

Identifies a network interface for the EC2 instance.

AwsEc2NetworkAclAssociation

An association between the network ACL and a subnet.

AwsEc2NetworkAclDetails

Contains details about an EC2 network access control list (ACL).

AwsEc2NetworkAclEntry

A rule for the network ACL. Each rule allows or denies access based on the IP address, traffic direction, port, and protocol.

AwsEc2NetworkInterfaceAttachment

Information about the network interface attachment.

AwsEc2NetworkInterfaceDetails

Details about the network interface

AwsEc2NetworkInterfaceIpV6AddressDetail

Provides information about an IPV6 address that is associated with the network

AwsEc2NetworkInterfacePrivateIpAddressDetail

Provides information about a private IPv4 address that is with the network

AwsEc2NetworkInterfaceSecurityGroup

A security group associated with the network interface.

AwsEc2SecurityGroupDetails

Details about an EC2 security group.

AwsEc2SecurityGroupIpPermission

An IP permission for an EC2 security group.

AwsEc2SecurityGroupIpRange

A range of IPv4 addresses.

AwsEc2SecurityGroupIpv6Range

A range of IPv6 addresses.

AwsEc2SecurityGroupPrefixListId

A prefix list ID.

AwsEc2SecurityGroupUserIdGroupPair

A relationship between a security group and a user.

AwsEc2SubnetDetails

Contains information about a subnet in Amazon EC2.

AwsEc2VolumeAttachment

An attachment to an Amazon EC2 volume.

AwsEc2VolumeDetails

Details about an EC2 volume.

AwsEc2VpcDetails

Details about an EC2 VPC.

AwsEc2VpnConnectionDetails

Details about an Amazon EC2 VPN

AwsEc2VpnConnectionOptionsDetails

VPN connection options.

AwsEc2VpnConnectionOptionsTunnelOptionsDetails

The VPN tunnel options.

AwsEc2VpnConnectionRoutesDetails

A static routes associated with

AwsEc2VpnConnectionVgwTelemetryDetails

Information about the VPN tunnel.

AwsEcrContainerImageDetails

Information about an Amazon ECR image.

AwsEcsClusterClusterSettingsDetails

Indicates whether to enable CloudWatch Container Insights for the ECS cluster.

AwsEcsClusterConfigurationDetails

The run command configuration for the cluster.

AwsEcsClusterConfigurationExecuteCommandConfigurationDetails

Contains the run command configuration for the cluster.

AwsEcsClusterConfigurationExecuteCommandConfigurationLogConfigurationDetails

The log configuration for the results of the run command actions.

AwsEcsClusterDefaultCapacityProviderStrategyDetails

The default capacity provider strategy for the cluster. The default capacity provider strategy is used when services or tasks are run without a specified launch type or capacity provider strategy.

AwsEcsClusterDetails

provides details about an ECS cluster.

AwsEcsServiceCapacityProviderStrategyDetails

Strategy item for the capacity provider strategy that the service uses.

AwsEcsServiceDeploymentConfigurationDeploymentCircuitBreakerDetails

Determines whether a service deployment fails if a service cannot reach a steady state.

AwsEcsServiceDeploymentConfigurationDetails

Optional deployment parameters for the service.

AwsEcsServiceDeploymentControllerDetails

Information about the deployment controller type that the service uses.

AwsEcsServiceDetails

Provides details about a service within an ECS cluster.

AwsEcsServiceLoadBalancersDetails

Information about a load balancer that the service uses.

AwsEcsServiceNetworkConfigurationAwsVpcConfigurationDetails

For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

AwsEcsServiceNetworkConfigurationDetails

For tasks that use the awsvpc networking mode, the VPC subnet and security group configuration.

AwsEcsServicePlacementConstraintsDetails

A placement constraint for the tasks in the service.

AwsEcsServicePlacementStrategiesDetails

A placement strategy that determines how to place the tasks for the service.

AwsEcsServiceServiceRegistriesDetails

Information about a service discovery registry to assign to the service.

AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails

A dependency that is defined for container startup and shutdown.

AwsEcsTaskDefinitionContainerDefinitionsDetails

A container definition that describes a container in the task.

AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails

An environment variable to pass to the container.

AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails

A file that contain environment variables to pass to a container.

AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails

A hostname and IP address mapping to append to the /etc/hosts file on the container.

AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails

The FireLens configuration for the container. The configuration specifies and configures a log router for container logs.

AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails

The container health check command and associated configuration parameters for the container.

AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersCapabilitiesDetails

The Linux capabilities for the container that are added to or dropped from the default configuration provided by Docker.

AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails

>Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDevicesDetails

A host device to expose to the container.

AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersTmpfsDetails

The container path, mount options, and size (in MiB) of a tmpfs mount.

AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails

The log configuration specification for the container.

AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationSecretOptionsDetails

A secret to pass to the log configuration.

AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails

A mount point for the data volumes in the container.

AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails

A port mapping for the container.

AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails

The private repository authentication credentials to use.

AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails

A resource to assign to a container.

AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails

A secret to pass to the container.

AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails

A namespaced kernel parameter to set in the container.

AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails

A ulimit to set in the container.

AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails

A data volume to mount from another container.

AwsEcsTaskDefinitionDetails

details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

AwsEcsTaskDefinitionInferenceAcceleratorsDetails

An Elastic Inference accelerator to use

AwsEcsTaskDefinitionPlacementConstraintsDetails

A placement constraint object to use for tasks.

AwsEcsTaskDefinitionProxyConfigurationDetails

The configuration details for the App Mesh

AwsEcsTaskDefinitionProxyConfigurationProxyConfigurationPropertiesDetails

A network configuration parameter to provide to the Container Network Interface (CNI) plugin.

AwsEcsTaskDefinitionVolumesDetails

A data volume to mount from another container.

AwsEcsTaskDefinitionVolumesDockerVolumeConfigurationDetails

Information about a Docker volume.

AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationAuthorizationConfigDetails

AwsEcsTaskDefinitionVolumesEfsVolumeConfigurationDetails

Information about the Amazon Elastic File System file system that is used for task storage.

AwsEcsTaskDefinitionVolumesHostDetails

Information about a bind mount host volume.

AwsElasticBeanstalkEnvironmentDetails

Contains details about an Elastic Beanstalk environment.

AwsElasticBeanstalkEnvironmentEnvironmentLink

Contains information about a link to another environment that is in the same group.

AwsElasticBeanstalkEnvironmentOptionSetting

A configuration option setting for the environment.

AwsElasticBeanstalkEnvironmentTier

Contains information about the tier of the environment.

AwsElasticsearchDomainDetails

Information about an Amazon Elasticsearch Service domain.

AwsElasticsearchDomainDomainEndpointOptions

Additional options for the domain endpoint, such as whether to require HTTPS for all

AwsElasticsearchDomainElasticsearchClusterConfigDetails

details about the configuration of an Elasticsearch cluster.

AwsElasticsearchDomainElasticsearchClusterConfigZoneAwarenessConfigDetails

Configuration options for zone awareness.

AwsElasticsearchDomainEncryptionAtRestOptions

Details about the configuration for encryption at rest.

AwsElasticsearchDomainLogPublishingOptions

configures the CloudWatch Logs to publish for the

AwsElasticsearchDomainLogPublishingOptionsLogConfig

The log configuration.

AwsElasticsearchDomainNodeToNodeEncryptionOptions

Details about the configuration for node-to-node encryption.

AwsElasticsearchDomainServiceSoftwareOptions

Information about the state of the domain relative to the latest service software.

AwsElasticsearchDomainVpcOptions

Information that Elasticsearch derives based on VPCOptions for the

AwsElbAppCookieStickinessPolicy

Contains information about a stickiness policy that was created using

AwsElbLbCookieStickinessPolicy

Contains information about a stickiness policy that was created using

AwsElbLoadBalancerAccessLog

Contains information about the access log configuration for the load balancer.

AwsElbLoadBalancerAttributes

Contains attributes for the load balancer.

AwsElbLoadBalancerBackendServerDescription

Provides information about the configuration of an EC2 instance for the load

AwsElbLoadBalancerConnectionDraining

Contains information about the connection draining configuration for the load

AwsElbLoadBalancerConnectionSettings

Contains connection settings for the load balancer.

AwsElbLoadBalancerCrossZoneLoadBalancing

Contains cross-zone load balancing settings for the load balancer.

AwsElbLoadBalancerDetails

Contains details about a Classic Load Balancer.

AwsElbLoadBalancerHealthCheck

Contains information about the health checks that are conducted on the load

AwsElbLoadBalancerInstance

Provides information about an EC2 instance for a load balancer.

AwsElbLoadBalancerListener

Information about a load balancer listener.

AwsElbLoadBalancerListenerDescription

Lists the policies that are enabled for a load balancer listener.

AwsElbLoadBalancerPolicies

Contains information about the policies for a load balancer.

AwsElbLoadBalancerSourceSecurityGroup

Contains information about the security group for the load balancer.

AwsElbv2LoadBalancerDetails

Information about a load balancer.

AwsIamAccessKeyDetails

IAM access key details related to a finding.

AwsIamAccessKeySessionContext

Provides information about the session that the key was used for.

AwsIamAccessKeySessionContextAttributes

Attributes of the session that the key was used for.

AwsIamAccessKeySessionContextSessionIssuer

Information about the entity that created the session.

AwsIamAttachedManagedPolicy

A managed policy that is attached to an IAM principal.

AwsIamGroupDetails

Contains details about an IAM group.

AwsIamGroupPolicy

A managed policy that is attached to the IAM group.

AwsIamInstanceProfile

Information about an instance profile.

AwsIamInstanceProfileRole

Information about a role associated with an instance profile.

AwsIamPermissionsBoundary

Information about the policy used to set the permissions boundary for an IAM

AwsIamPolicyDetails

Represents an IAM permissions policy.

AwsIamPolicyVersion

A version of an IAM policy.

AwsIamRoleDetails

Contains information about an IAM role, including all of the role's policies.

AwsIamRolePolicy

An inline policy that is embedded in the role.

AwsIamUserDetails

Information about an IAM user.

AwsIamUserPolicy

Information about an inline policy that is embedded in the user.

AwsKmsKeyDetails

Contains metadata about an KMS key.

AwsLambdaFunctionCode

The code for the Lambda function. You can specify either an object in Amazon S3, or upload a deployment package directly.

AwsLambdaFunctionDeadLetterConfig

The dead-letter queue for failed asynchronous invocations.

AwsLambdaFunctionDetails

Details about a function's configuration.

AwsLambdaFunctionEnvironment

A function's environment variable settings.

AwsLambdaFunctionEnvironmentError

Error messages for environment variables that could not be applied.

AwsLambdaFunctionLayer

An Lambda layer.

AwsLambdaFunctionTracingConfig

The function's X-Ray tracing configuration.

AwsLambdaFunctionVpcConfig

The VPC security groups and subnets that are attached to a Lambda function.

AwsLambdaLayerVersionDetails

Details about a Lambda layer version.

AwsRdsDbClusterAssociatedRole

An IAM role that is associated with the Amazon RDS DB cluster.

AwsRdsDbClusterDetails

Information about an Amazon RDS DB cluster.

AwsRdsDbClusterMember

Information about an instance in the DB cluster.

AwsRdsDbClusterOptionGroupMembership

Information about an option group membership for a DB cluster.

AwsRdsDbClusterSnapshotDetails

Information about an Amazon RDS DB cluster snapshot.

AwsRdsDbDomainMembership

Information about an Active Directory domain membership record associated with the DB

AwsRdsDbInstanceAssociatedRole

An IAM role associated with the DB instance.

AwsRdsDbInstanceDetails

Contains the details of an Amazon RDS DB instance.

AwsRdsDbInstanceEndpoint

Specifies the connection endpoint.

AwsRdsDbInstanceVpcSecurityGroup

A VPC security groups that the DB instance belongs to.

AwsRdsDbOptionGroupMembership

An option group membership.

AwsRdsDbParameterGroup

Provides information about a parameter group for a DB instance.

AwsRdsDbPendingModifiedValues

Changes to a DB instance that are currently pending.

AwsRdsDbProcessorFeature

A processor feature.

AwsRdsDbSnapshotDetails

Provides details about an Amazon RDS DB cluster snapshot.

AwsRdsDbStatusInfo

Information about the status of a read replica.

AwsRdsDbSubnetGroup

Information about the subnet group for the database instance.

AwsRdsDbSubnetGroupSubnet

Information about a subnet in a subnet group.

AwsRdsDbSubnetGroupSubnetAvailabilityZone

An Availability Zone for a subnet in a subnet group.

AwsRdsEventSubscriptionDetails

Details about an Amazon RDS event notification subscription. The subscription allows Amazon RDS to post events to an SNS topic.

AwsRdsPendingCloudWatchLogsExports

Identifies the log types to enable and disable.

AwsRedshiftClusterClusterNode

A node in an Amazon Redshift cluster.

AwsRedshiftClusterClusterParameterGroup

A cluster parameter group that is associated with an Amazon Redshift cluster.

AwsRedshiftClusterClusterParameterStatus

The status of a parameter in a cluster parameter group for an Amazon Redshift

AwsRedshiftClusterClusterSecurityGroup

A security group that is associated with the cluster.

AwsRedshiftClusterClusterSnapshotCopyStatus

Information about a cross-Region snapshot copy.

AwsRedshiftClusterDeferredMaintenanceWindow

A time windows during which maintenance was deferred for an Amazon Redshift

AwsRedshiftClusterDetails

Details about an Amazon Redshift cluster.

AwsRedshiftClusterElasticIpStatus

The status of the elastic IP (EIP) address for an Amazon Redshift cluster.

AwsRedshiftClusterEndpoint

The connection endpoint for an Amazon Redshift cluster.

AwsRedshiftClusterHsmStatus

Information about whether an Amazon Redshift cluster finished applying any hardware

AwsRedshiftClusterIamRole

An IAM role that the cluster can use to access other Amazon Web Services services.

AwsRedshiftClusterPendingModifiedValues

Changes to the Amazon Redshift cluster that are currently pending.

AwsRedshiftClusterResizeInfo

Information about the resize operation for the cluster.

AwsRedshiftClusterRestoreStatus

Information about the status of a cluster restore action. It only applies if the cluster

AwsRedshiftClusterVpcSecurityGroup

A VPC security group that the cluster belongs to, if the cluster is in a VPC.

AwsS3AccountPublicAccessBlockDetails

provides information about the Amazon S3 Public Access Block configuration for accounts.

AwsS3BucketBucketLifecycleConfigurationDetails

The lifecycle configuration for the objects in the S3 bucket.

AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails

Information about what Amazon S3

AwsS3BucketBucketLifecycleConfigurationRulesDetails

Configuration for a lifecycle rule.

AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails

Identifies the objects that a rule applies to.

AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateDetails

The configuration for the filter.

AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateOperandsDetails

A value to use for the filter.

AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateOperandsTagDetails

A tag that is assigned to matching objects.

AwsS3BucketBucketLifecycleConfigurationRulesFilterPredicateTagDetails

A tag filter.

AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails

A transition rule that describes when noncurrent objects transition to a specified storage class.

AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails

A rule for when objects transition to specific storage classes.

AwsS3BucketDetails

The details of an Amazon S3 bucket.

AwsS3BucketLoggingConfiguration

Information about logging for

AwsS3BucketNotificationConfiguration

The notification

AwsS3BucketNotificationConfigurationDetail

Details for an S3 bucket notification configuration.

AwsS3BucketNotificationConfigurationFilter

Filtering information for the notifications. The

AwsS3BucketNotificationConfigurationS3KeyFilter

Details for an Amazon S3 filter.

AwsS3BucketNotificationConfigurationS3KeyFilterRule

Details for a filter rule.

AwsS3BucketServerSideEncryptionByDefault

Specifies the default server-side encryption to apply to new objects in the

AwsS3BucketServerSideEncryptionConfiguration

The encryption configuration for the S3 bucket.

AwsS3BucketServerSideEncryptionRule

An encryption rule to apply to the S3 bucket.

AwsS3BucketWebsiteConfiguration

Website parameters for the S3

AwsS3BucketWebsiteConfigurationRedirectTo

The redirect behavior for requests

AwsS3BucketWebsiteConfigurationRoutingRule

A rule for redirecting requests

AwsS3BucketWebsiteConfigurationRoutingRuleCondition

The condition that must be met in order to apply the routing rule.

AwsS3BucketWebsiteConfigurationRoutingRuleRedirect

The rules to redirect the request if the condition in Condition is

AwsS3ObjectDetails

Details about an Amazon S3 object.

AwsSecretsManagerSecretDetails

Details about an Secrets Manager secret.

AwsSecretsManagerSecretRotationRules

Defines the rotation schedule for the secret.

AwsSecurityFinding

Provides consistent format for the contents of the Security Hub-aggregated findings.

AwsSecurityFindingFilters

A collection of attributes that are applied to all active Security Hub-aggregated findings and

AwsSecurityFindingIdentifier

Identifies a finding to update using BatchUpdateFindings.

AwsSnsTopicDetails

A wrapper type for the topic's ARN.

AwsSnsTopicSubscription

A wrapper type for the attributes of an Amazon SNS subscription.

AwsSqsQueueDetails

Data about a queue.

AwsSsmComplianceSummary

Provides the details about the compliance status for a patch.

AwsSsmPatch

Provides details about the compliance for a patch.

AwsSsmPatchComplianceDetails

Provides information about the state of a patch on an instance based on the patch

AwsWafWebAclDetails

Details about an WAF WebACL.

AwsWafWebAclRule

Details for a rule in an WAF WebACL.

BatchUpdateFindingsUnprocessedFinding

A finding from a BatchUpdateFindings request that Security Hub was unable to

Cell

An occurrence of sensitive data detected in a Microsoft Excel workbook, comma-separated value (CSV) file, or tab-separated value (TSV) file.

CidrBlockAssociation

An IPv4 CIDR block association.

City

Information about a city.

ClassificationResult

Details about the sensitive data that was detected on the resource.

ClassificationStatus

Provides details about the current status of the sensitive data detection.

Compliance

Contains finding details that are specific to control-based findings. Only returned for

ContainerDetails

Container details related to a finding.

Country

Information about a country.

CustomDataIdentifiersDetections

The list of detected instances of sensitive data.

CustomDataIdentifiersResult

Contains an instance of sensitive data that was detected by a customer-defined identifier.

Cvss

CVSS scores from the advisory related to the vulnerability.

DataClassificationDetails

Provides details about sensitive data that was detected on a resource.

DateFilter

A date filter for querying findings.

DateRange

A date range for the date filter.

DnsRequestAction

Provided if ActionType is DNS_REQUEST. It provides details

FindingProviderFields

In a BatchImportFindings request, finding providers use FindingProviderFields to provide and update values for confidence, criticality, related findings, severity, and types.

FindingProviderSeverity

The severity assigned to the finding by the finding provider.

GeoLocation

Provides the latitude and longitude coordinates of a location.

IcmpTypeCode

An Internet Control Message Protocol (ICMP) type and code.

ImportFindingsError

The list of the findings that cannot be imported. For each finding, the list provides

Insight

Contains information about a Security Hub insight.

InsightResultValue

The insight result values returned by the GetInsightResults

InsightResults

The insight results returned by the GetInsightResults operation.

Invitation

Details about an invitation.

IpFilter

The IP filter for querying findings.

IpOrganizationDetails

Provides information about an internet provider.

Ipv6CidrBlockAssociation

An IPV6 CIDR block association.

KeywordFilter

A keyword filter for querying findings.

LoadBalancerState

Information about the state of the load balancer.

Malware

A list of malware related to a finding.

MapFilter

A map filter for querying findings. Each map filter provides the field to check, the

Member

The details about a member account.

Network

The details of network-related information about a finding.

NetworkConnectionAction

Provided if ActionType is NETWORK_CONNECTION. It provides

NetworkHeader

Details about a network path component that occurs before or after the current

NetworkPathComponent

Information about a network path component.

NetworkPathComponentDetails

Information about the destination of the next component in the network path.

Note

A user-defined note added to a finding.

NoteUpdate

The updated note.

NumberFilter

A number filter for querying findings.

Occurrences

The detected occurrences of sensitive data.

Page

An occurrence of sensitive data in an Adobe Portable Document Format (PDF) file.

PatchSummary

Provides an overview of the patch compliance status for an instance against a selected

PortProbeAction

Provided if ActionType is PORT_PROBE. It provides details

PortProbeDetail

A port scan that was part of the port probe. For each scan, PortProbeDetails provides

PortRange

A range of ports.

PortRangeFromTo

A range of ports.

ProcessDetails

The details of process-related information about a finding.

Product

Contains details about a product.

Range

Identifies where the sensitive data begins and ends.

Recommendation

A recommendation on how to remediate the issue identified in a finding.

Record

An occurrence of sensitive data in an Apache Avro object container or an Apache Parquet file.

RelatedFinding

Details about a related finding.

Remediation

Details about the remediation steps for a finding.

Resource

A resource related to a finding.

ResourceDetails

Additional details about a resource related to a finding.

Result

Details about the account that was not processed.

SensitiveDataDetections

The list of detected instances of sensitive data.

SensitiveDataResult

Contains a detected instance of sensitive data that are based on built-in identifiers.

Severity

The severity of the finding.

SeverityUpdate

Updates to the severity information for a finding.

SoftwarePackage

Information about a software package.

SortCriterion

A collection of finding attributes used to sort findings.

Standard

Provides information about a specific standard.

StandardsControl

Details for an individual security standard control.

StandardsSubscription

A resource that represents your subscription to a supported standard.

StandardsSubscriptionRequest

The standard that you want to enable.

StatusReason

Provides additional context for the value of Compliance.Status.

StringFilter

A string filter for querying findings.

ThreatIntelIndicator

Details about the threat intelligence related to a finding.

Vulnerability

A vulnerability associated with a finding.

VulnerabilityVendor

A vendor that generates a vulnerability report.

WafAction

Details about the action that CloudFront or WAF takes when a web request matches the

WafExcludedRule

Details about a rule to exclude from a rule group.

WafOverrideAction

Details about an override action for a rule.

Workflow

Provides information about the status of the investigation into a finding.

WorkflowUpdate

Used to update information about the investigation into the finding.

Enums

AdminStatus

AwsIamAccessKeyStatus

AwsS3BucketNotificationConfigurationS3KeyFilterRuleName

ComplianceStatus

ControlStatus

DateRangeUnit

IntegrationType

MalwareState

MalwareType

MapFilterComparison

NetworkDirection

Partition

RecordState

SeverityLabel

SeverityRating

SortOrder

StandardsStatus

StringFilterComparison

ThreatIntelIndicatorCategory

ThreatIntelIndicatorType

VerificationState

NOTE: VerificationState::Unknown has been renamed to ::UnknownValue.

WorkflowState

WorkflowStatus