Struct aws_sdk_securityhub::model::aws_auto_scaling_launch_configuration_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsAutoScalingLaunchConfigurationDetails

Implementations

impl Builder

pub fn associate_public_ip_address(self, input: bool) -> Self

For Auto Scaling groups that run in a VPC, specifies whether to assign a public IP address to the group's instances.

pub fn set_associate_public_ip_address(self, input: Option<bool>) -> Self

For Auto Scaling groups that run in a VPC, specifies whether to assign a public IP address to the group's instances.

pub fn block_device_mappings(
self,
input: impl Into<AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails>
) -> Self

Appends an item to block_device_mappings.

To override the contents of this collection use set_block_device_mappings.

Specifies the block devices for the instance.

pub fn set_block_device_mappings(
self,
input: Option<Vec<AwsAutoScalingLaunchConfigurationBlockDeviceMappingsDetails>>
) -> Self

Specifies the block devices for the instance.

pub fn classic_link_vpc_id(self, input: impl Into<String>) -> Self

The identifier of a ClassicLink-enabled VPC that EC2-Classic instances are linked to.

pub fn set_classic_link_vpc_id(self, input: Option<String>) -> Self

The identifier of a ClassicLink-enabled VPC that EC2-Classic instances are linked to.

pub fn classic_link_vpc_security_groups(self, input: impl Into<String>) -> Self

Appends an item to classic_link_vpc_security_groups.

To override the contents of this collection use set_classic_link_vpc_security_groups.

The identifiers of one or more security groups for the VPC that is specified in ClassicLinkVPCId.

pub fn set_classic_link_vpc_security_groups(
self,
input: Option<Vec<String>>
) -> Self

The identifiers of one or more security groups for the VPC that is specified in ClassicLinkVPCId.

pub fn created_time(self, input: impl Into<String>) -> Self

The creation date and time for the launch configuration.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_created_time(self, input: Option<String>) -> Self

The creation date and time for the launch configuration.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn ebs_optimized(self, input: bool) -> Self

Whether the launch configuration is optimized for Amazon EBS I/O.

pub fn set_ebs_optimized(self, input: Option<bool>) -> Self

Whether the launch configuration is optimized for Amazon EBS I/O.

pub fn iam_instance_profile(self, input: impl Into<String>) -> Self

The name or the ARN of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

pub fn set_iam_instance_profile(self, input: Option<String>) -> Self

The name or the ARN of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

pub fn image_id(self, input: impl Into<String>) -> Self

The identifier of the Amazon Machine Image (AMI) that is used to launch EC2 instances.

pub fn set_image_id(self, input: Option<String>) -> Self

The identifier of the Amazon Machine Image (AMI) that is used to launch EC2 instances.

pub fn instance_monitoring(
self,
input: AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails
) -> Self

Indicates the type of monitoring for instances in the group.

pub fn set_instance_monitoring(
self,
input: Option<AwsAutoScalingLaunchConfigurationInstanceMonitoringDetails>
) -> Self

Indicates the type of monitoring for instances in the group.

pub fn instance_type(self, input: impl Into<String>) -> Self

The instance type for the instances.

pub fn set_instance_type(self, input: Option<String>) -> Self

The instance type for the instances.

pub fn kernel_id(self, input: impl Into<String>) -> Self

The identifier of the kernel associated with the AMI.

pub fn set_kernel_id(self, input: Option<String>) -> Self

The identifier of the kernel associated with the AMI.

pub fn key_name(self, input: impl Into<String>) -> Self

The name of the key pair.

pub fn set_key_name(self, input: Option<String>) -> Self

The name of the key pair.

pub fn launch_configuration_name(self, input: impl Into<String>) -> Self

The name of the launch configuration.

pub fn set_launch_configuration_name(self, input: Option<String>) -> Self

The name of the launch configuration.

pub fn placement_tenancy(self, input: impl Into<String>) -> Self

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

pub fn set_placement_tenancy(self, input: Option<String>) -> Self

The tenancy of the instance. An instance with dedicated tenancy runs on isolated, single-tenant hardware and can only be launched into a VPC.

pub fn ramdisk_id(self, input: impl Into<String>) -> Self

The identifier of the RAM disk associated with the AMI.

pub fn set_ramdisk_id(self, input: Option<String>) -> Self

The identifier of the RAM disk associated with the AMI.

pub fn security_groups(self, input: impl Into<String>) -> Self

Appends an item to security_groups.

To override the contents of this collection use set_security_groups.

The security groups to assign to the instances in the Auto Scaling group.

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

The security groups to assign to the instances in the Auto Scaling group.

pub fn spot_price(self, input: impl Into<String>) -> Self

The maximum hourly price to be paid for any Spot Instance that is launched to fulfill the request.

pub fn set_spot_price(self, input: Option<String>) -> Self

The maximum hourly price to be paid for any Spot Instance that is launched to fulfill the request.

pub fn user_data(self, input: impl Into<String>) -> Self

The user data to make available to the launched EC2 instances. Must be base64-encoded text.

pub fn set_user_data(self, input: Option<String>) -> Self

The user data to make available to the launched EC2 instances. Must be base64-encoded text.

pub fn build(self) -> AwsAutoScalingLaunchConfigurationDetails

Consumes the builder and constructs a AwsAutoScalingLaunchConfigurationDetails