Struct aws_sdk_securityhub::model::aws_certificate_manager_certificate_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsCertificateManagerCertificateDetails

Implementations

impl Builder

pub fn certificate_authority_arn(self, input: impl Into<String>) -> Self

The ARN of the private certificate authority (CA) that will be used to issue the certificate.

pub fn set_certificate_authority_arn(self, input: Option<String>) -> Self

pub fn created_at(self, input: impl Into<String>) -> Self

Indicates when the certificate was requested.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_created_at(self, input: Option<String>) -> Self

pub fn domain_name(self, input: impl Into<String>) -> Self

The fully qualified domain name (FQDN), such as www.example.com, that is secured by the certificate.

pub fn set_domain_name(self, input: Option<String>) -> Self

pub fn domain_validation_options(
self,
input: impl Into<AwsCertificateManagerCertificateDomainValidationOption>
) -> Self

pub fn set_domain_validation_options(
self,
input: Option<Vec<AwsCertificateManagerCertificateDomainValidationOption>>
) -> Self

pub fn extended_key_usages(
self,
input: impl Into<AwsCertificateManagerCertificateExtendedKeyUsage>
) -> Self

pub fn set_extended_key_usages(
self,
input: Option<Vec<AwsCertificateManagerCertificateExtendedKeyUsage>>
) -> Self

pub fn failure_reason(self, input: impl Into<String>) -> Self

For a failed certificate request, the reason for the failure.

Valid values: NO_AVAILABLE_CONTACTS | ADDITIONAL_VERIFICATION_REQUIRED | DOMAIN_NOT_ALLOWED | INVALID_PUBLIC_DOMAIN | DOMAIN_VALIDATION_DENIED | CAA_ERROR | PCA_LIMIT_EXCEEDED | PCA_INVALID_ARN | PCA_INVALID_STATE | PCA_REQUEST_FAILED | PCA_NAME_CONSTRAINTS_VALIDATION | PCA_RESOURCE_NOT_FOUND | PCA_INVALID_ARGS | PCA_INVALID_DURATION | PCA_ACCESS_DENIED | SLR_NOT_FOUND | OTHER

pub fn set_failure_reason(self, input: Option<String>) -> Self

pub fn imported_at(self, input: impl Into<String>) -> Self

Indicates when the certificate was imported. Provided if the certificate type is IMPORTED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_imported_at(self, input: Option<String>) -> Self

pub fn in_use_by(self, input: impl Into<String>) -> Self

pub fn set_in_use_by(self, input: Option<Vec<String>>) -> Self

pub fn issued_at(self, input: impl Into<String>) -> Self

Indicates when the certificate was issued. Provided if the certificate type is AMAZON_ISSUED.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_issued_at(self, input: Option<String>) -> Self

pub fn issuer(self, input: impl Into<String>) -> Self

The name of the certificate authority that issued and signed the certificate.

pub fn set_issuer(self, input: Option<String>) -> Self

pub fn key_algorithm(self, input: impl Into<String>) -> Self

The algorithm that was used to generate the public-private key pair.

Valid values: RSA_2048 | RSA_1024 | RSA_4096 | EC_prime256v1 | EC_secp384r1 | EC_secp521r1

pub fn set_key_algorithm(self, input: Option<String>) -> Self

pub fn key_usages(
self,
input: impl Into<AwsCertificateManagerCertificateKeyUsage>
) -> Self

pub fn set_key_usages(
self,
input: Option<Vec<AwsCertificateManagerCertificateKeyUsage>>
) -> Self

pub fn not_after(self, input: impl Into<String>) -> Self

The time after which the certificate becomes invalid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_not_after(self, input: Option<String>) -> Self

pub fn not_before(self, input: impl Into<String>) -> Self

The time before which the certificate is not valid.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_not_before(self, input: Option<String>) -> Self

pub fn options(self, input: AwsCertificateManagerCertificateOptions) -> Self

Provides a value that specifies whether to add the certificate to a transparency log.

pub fn set_options(
self,
input: Option<AwsCertificateManagerCertificateOptions>
) -> Self

pub fn renewal_eligibility(self, input: impl Into<String>) -> Self

Whether the certificate is eligible for renewal.

Valid values: ELIGIBLE | INELIGIBLE

pub fn set_renewal_eligibility(self, input: Option<String>) -> Self

pub fn renewal_summary(
self,
input: AwsCertificateManagerCertificateRenewalSummary
) -> Self

Information about the status of the Certificate Manager managed renewal for the certificate. Provided only when the certificate type is AMAZON_ISSUED.

pub fn set_renewal_summary(
self,
input: Option<AwsCertificateManagerCertificateRenewalSummary>
) -> Self

pub fn serial(self, input: impl Into<String>) -> Self

The serial number of the certificate.

pub fn set_serial(self, input: Option<String>) -> Self

pub fn signature_algorithm(self, input: impl Into<String>) -> Self

The algorithm that was used to sign the certificate.

pub fn set_signature_algorithm(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The status of the certificate.

Valid values: PENDING_VALIDATION | ISSUED | INACTIVE | EXPIRED | VALIDATION_TIMED_OUT | REVOKED | FAILED

pub fn set_status(self, input: Option<String>) -> Self

pub fn subject(self, input: impl Into<String>) -> Self

The name of the entity that is associated with the public key contained in the certificate.

pub fn set_subject(self, input: Option<String>) -> Self

pub fn subject_alternative_names(self, input: impl Into<String>) -> Self

pub fn set_subject_alternative_names(self, input: Option<Vec<String>>) -> Self

pub fn type(self, input: impl Into<String>) -> Self

The source of the certificate. For certificates that Certificate Manager provides, Type is AMAZON_ISSUED. For certificates that are imported with ImportCertificate, Type is IMPORTED.

Valid values: IMPORTED | AMAZON_ISSUED | PRIVATE

pub fn set_type(self, input: Option<String>) -> Self

pub fn build(self) -> AwsCertificateManagerCertificateDetails

Consumes the builder and constructs a AwsCertificateManagerCertificateDetails