Struct aws_sdk_securityhub::model::aws_elb_load_balancer_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsElbLoadBalancerDetails

Implementations

impl Builder

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn backend_server_descriptions(
self,
input: impl Into<AwsElbLoadBalancerBackendServerDescription>
) -> Self

pub fn set_backend_server_descriptions(
self,
input: Option<Vec<AwsElbLoadBalancerBackendServerDescription>>
) -> Self

pub fn canonical_hosted_zone_name(self, input: impl Into<String>) -> Self

The name of the Amazon Route 53 hosted zone for the load balancer.

pub fn set_canonical_hosted_zone_name(self, input: Option<String>) -> Self

pub fn canonical_hosted_zone_name_id(self, input: impl Into<String>) -> Self

The ID of the Amazon Route 53 hosted zone for the load balancer.

pub fn set_canonical_hosted_zone_name_id(self, input: Option<String>) -> Self

pub fn created_time(self, input: impl Into<String>) -> Self

Indicates when the load balancer was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_created_time(self, input: Option<String>) -> Self

pub fn dns_name(self, input: impl Into<String>) -> Self

The DNS name of the load balancer.

pub fn set_dns_name(self, input: Option<String>) -> Self

pub fn health_check(self, input: AwsElbLoadBalancerHealthCheck) -> Self

Information about the health checks that are conducted on the load balancer.

pub fn set_health_check(
self,
input: Option<AwsElbLoadBalancerHealthCheck>
) -> Self

pub fn instances(self, input: impl Into<AwsElbLoadBalancerInstance>) -> Self

pub fn set_instances(
self,
input: Option<Vec<AwsElbLoadBalancerInstance>>
) -> Self

pub fn listener_descriptions(
self,
input: impl Into<AwsElbLoadBalancerListenerDescription>
) -> Self

pub fn set_listener_descriptions(
self,
input: Option<Vec<AwsElbLoadBalancerListenerDescription>>
) -> Self

pub fn load_balancer_attributes(
self,
input: AwsElbLoadBalancerAttributes
) -> Self

The attributes for a load balancer.

pub fn set_load_balancer_attributes(
self,
input: Option<AwsElbLoadBalancerAttributes>
) -> Self

pub fn load_balancer_name(self, input: impl Into<String>) -> Self

The name of the load balancer.

pub fn set_load_balancer_name(self, input: Option<String>) -> Self

pub fn policies(self, input: AwsElbLoadBalancerPolicies) -> Self

The policies for a load balancer.

pub fn set_policies(self, input: Option<AwsElbLoadBalancerPolicies>) -> Self

pub fn scheme(self, input: impl Into<String>) -> Self

The type of load balancer. Only provided if the load balancer is in a VPC.

If Scheme is internet-facing, the load balancer has a public DNS name that resolves to a public IP address.

If Scheme is internal, the load balancer has a public DNS name that resolves to a private IP address.

pub fn set_scheme(self, input: Option<String>) -> Self

pub fn security_groups(self, input: impl Into<String>) -> Self

pub fn set_security_groups(self, input: Option<Vec<String>>) -> Self

pub fn source_security_group(
self,
input: AwsElbLoadBalancerSourceSecurityGroup
) -> Self

Information about the security group for the load balancer. This is the security group that is used for inbound rules.

pub fn set_source_security_group(
self,
input: Option<AwsElbLoadBalancerSourceSecurityGroup>
) -> Self

pub fn subnets(self, input: impl Into<String>) -> Self

pub fn set_subnets(self, input: Option<Vec<String>>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The identifier of the VPC for the load balancer.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn build(self) -> AwsElbLoadBalancerDetails

Consumes the builder and constructs a AwsElbLoadBalancerDetails