Struct aws_sdk_securityhub::model::aws_rds_db_cluster_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsRdsDbClusterDetails

Implementations

impl Builder

pub fn allocated_storage(self, input: i32) -> Self

For all database engines except Aurora, specifies the allocated storage size in gibibytes (GiB).

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

pub fn availability_zones(self, input: impl Into<String>) -> Self

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which automated backups are retained.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

pub fn database_name(self, input: impl Into<String>) -> Self

The name of the database.

pub fn set_database_name(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The current status of this DB cluster.

pub fn set_status(self, input: Option<String>) -> Self

pub fn endpoint(self, input: impl Into<String>) -> Self

The connection endpoint for the primary instance of the DB cluster.

pub fn set_endpoint(self, input: Option<String>) -> Self

pub fn reader_endpoint(self, input: impl Into<String>) -> Self

The reader endpoint for the DB cluster.

pub fn set_reader_endpoint(self, input: Option<String>) -> Self

pub fn custom_endpoints(self, input: impl Into<String>) -> Self

pub fn set_custom_endpoints(self, input: Option<Vec<String>>) -> Self

pub fn multi_az(self, input: bool) -> Self

Whether the DB cluster has instances in multiple Availability Zones.

pub fn set_multi_az(self, input: Option<bool>) -> Self

pub fn engine(self, input: impl Into<String>) -> Self

The name of the database engine to use for this DB cluster.

pub fn set_engine(self, input: Option<String>) -> Self

pub fn engine_version(self, input: impl Into<String>) -> Self

The version number of the database engine to use.

pub fn set_engine_version(self, input: Option<String>) -> Self

pub fn port(self, input: i32) -> Self

The port number on which the DB instances in the DB cluster accept connections.

pub fn set_port(self, input: Option<i32>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The name of the master user for the DB cluster.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format :HH:MM-:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn read_replica_identifiers(self, input: impl Into<String>) -> Self

pub fn set_read_replica_identifiers(self, input: Option<Vec<String>>) -> Self

pub fn vpc_security_groups(
self,
input: impl Into<AwsRdsDbInstanceVpcSecurityGroup>
) -> Self

pub fn set_vpc_security_groups(
self,
input: Option<Vec<AwsRdsDbInstanceVpcSecurityGroup>>
) -> Self

pub fn hosted_zone_id(self, input: impl Into<String>) -> Self

Specifies the identifier that Amazon Route 53 assigns when you create a hosted zone.

pub fn set_hosted_zone_id(self, input: Option<String>) -> Self

pub fn storage_encrypted(self, input: bool) -> Self

Whether the DB cluster is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The ARN of the KMS master key that is used to encrypt the database instances in the DB cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn db_cluster_resource_id(self, input: impl Into<String>) -> Self

The identifier of the DB cluster. The identifier must be unique within each Amazon Web Services Region and is immutable.

pub fn set_db_cluster_resource_id(self, input: Option<String>) -> Self

pub fn associated_roles(
self,
input: impl Into<AwsRdsDbClusterAssociatedRole>
) -> Self

pub fn set_associated_roles(
self,
input: Option<Vec<AwsRdsDbClusterAssociatedRole>>
) -> Self

pub fn cluster_create_time(self, input: impl Into<String>) -> Self

Indicates when the DB cluster was created, in Universal Coordinated Time (UTC).

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_cluster_create_time(self, input: Option<String>) -> Self

pub fn enabled_cloud_watch_logs_exports(self, input: impl Into<String>) -> Self

pub fn set_enabled_cloud_watch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

pub fn engine_mode(self, input: impl Into<String>) -> Self

The database engine mode of the DB cluster.

pub fn set_engine_mode(self, input: Option<String>) -> Self

pub fn deletion_protection(self, input: bool) -> Self

Whether the DB cluster has deletion protection enabled.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

pub fn http_endpoint_enabled(self, input: bool) -> Self

Whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

pub fn set_http_endpoint_enabled(self, input: Option<bool>) -> Self

pub fn activity_stream_status(self, input: impl Into<String>) -> Self

The status of the database activity stream.

pub fn set_activity_stream_status(self, input: Option<String>) -> Self

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Whether tags are copied from the DB cluster to snapshots of the DB cluster.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

pub fn cross_account_clone(self, input: bool) -> Self

Whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

pub fn set_cross_account_clone(self, input: Option<bool>) -> Self

pub fn domain_memberships(
self,
input: impl Into<AwsRdsDbDomainMembership>
) -> Self

pub fn set_domain_memberships(
self,
input: Option<Vec<AwsRdsDbDomainMembership>>
) -> Self

pub fn db_cluster_parameter_group(self, input: impl Into<String>) -> Self

The name of the DB cluster parameter group for the DB cluster.

pub fn set_db_cluster_parameter_group(self, input: Option<String>) -> Self

pub fn db_subnet_group(self, input: impl Into<String>) -> Self

The subnet group that is associated with the DB cluster, including the name, description, and subnets in the subnet group.

pub fn set_db_subnet_group(self, input: Option<String>) -> Self

pub fn db_cluster_option_group_memberships(
self,
input: impl Into<AwsRdsDbClusterOptionGroupMembership>
) -> Self

pub fn set_db_cluster_option_group_memberships(
self,
input: Option<Vec<AwsRdsDbClusterOptionGroupMembership>>
) -> Self

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

The DB cluster identifier that the user assigned to the cluster. This identifier is the unique key that identifies a DB cluster.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

pub fn db_cluster_members(self, input: impl Into<AwsRdsDbClusterMember>) -> Self

pub fn set_db_cluster_members(
self,
input: Option<Vec<AwsRdsDbClusterMember>>
) -> Self

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

Whether the mapping of IAM accounts to database accounts is enabled.

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

pub fn build(self) -> AwsRdsDbClusterDetails

Consumes the builder and constructs a AwsRdsDbClusterDetails