Struct aws_sdk_securityhub::model::aws_redshift_cluster_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsRedshiftClusterDetails

Implementations

impl Builder

pub fn allow_version_upgrade(self, input: bool) -> Self

Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

pub fn set_allow_version_upgrade(self, input: Option<bool>) -> Self

pub fn automated_snapshot_retention_period(self, input: i32) -> Self

The number of days that automatic cluster snapshots are retained.

pub fn set_automated_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn availability_zone(self, input: impl Into<String>) -> Self

The name of the Availability Zone in which the cluster is located.

pub fn set_availability_zone(self, input: Option<String>) -> Self

pub fn cluster_availability_status(self, input: impl Into<String>) -> Self

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

pub fn set_cluster_availability_status(self, input: Option<String>) -> Self

pub fn cluster_create_time(self, input: impl Into<String>) -> Self

Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_cluster_create_time(self, input: Option<String>) -> Self

pub fn cluster_identifier(self, input: impl Into<String>) -> Self

The unique identifier of the cluster.

pub fn set_cluster_identifier(self, input: Option<String>) -> Self

pub fn cluster_nodes(
self,
input: impl Into<AwsRedshiftClusterClusterNode>
) -> Self

pub fn set_cluster_nodes(
self,
input: Option<Vec<AwsRedshiftClusterClusterNode>>
) -> Self

pub fn cluster_parameter_groups(
self,
input: impl Into<AwsRedshiftClusterClusterParameterGroup>
) -> Self

pub fn set_cluster_parameter_groups(
self,
input: Option<Vec<AwsRedshiftClusterClusterParameterGroup>>
) -> Self

pub fn cluster_public_key(self, input: impl Into<String>) -> Self

The public key for the cluster.

pub fn set_cluster_public_key(self, input: Option<String>) -> Self

pub fn cluster_revision_number(self, input: impl Into<String>) -> Self

The specific revision number of the database in the cluster.

pub fn set_cluster_revision_number(self, input: Option<String>) -> Self

pub fn cluster_security_groups(
self,
input: impl Into<AwsRedshiftClusterClusterSecurityGroup>
) -> Self

pub fn set_cluster_security_groups(
self,
input: Option<Vec<AwsRedshiftClusterClusterSecurityGroup>>
) -> Self

pub fn cluster_snapshot_copy_status(
self,
input: AwsRedshiftClusterClusterSnapshotCopyStatus
) -> Self

Information about the destination Region and retention period for the cross-Region snapshot copy.

pub fn set_cluster_snapshot_copy_status(
self,
input: Option<AwsRedshiftClusterClusterSnapshotCopyStatus>
) -> Self

pub fn cluster_status(self, input: impl Into<String>) -> Self

The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

pub fn set_cluster_status(self, input: Option<String>) -> Self

pub fn cluster_subnet_group_name(self, input: impl Into<String>) -> Self

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

pub fn set_cluster_subnet_group_name(self, input: Option<String>) -> Self

pub fn cluster_version(self, input: impl Into<String>) -> Self

The version ID of the Amazon Redshift engine that runs on the cluster.

pub fn set_cluster_version(self, input: Option<String>) -> Self

pub fn db_name(self, input: impl Into<String>) -> Self

The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.

pub fn set_db_name(self, input: Option<String>) -> Self

pub fn deferred_maintenance_windows(
self,
input: impl Into<AwsRedshiftClusterDeferredMaintenanceWindow>
) -> Self

pub fn set_deferred_maintenance_windows(
self,
input: Option<Vec<AwsRedshiftClusterDeferredMaintenanceWindow>>
) -> Self

pub fn elastic_ip_status(self, input: AwsRedshiftClusterElasticIpStatus) -> Self

Information about the status of the Elastic IP (EIP) address.

pub fn set_elastic_ip_status(
self,
input: Option<AwsRedshiftClusterElasticIpStatus>
) -> Self

pub fn elastic_resize_number_of_node_options(
self,
input: impl Into<String>
) -> Self

The number of nodes that you can use the elastic resize method to resize the cluster to.

pub fn set_elastic_resize_number_of_node_options(
self,
input: Option<String>
) -> Self

pub fn encrypted(self, input: bool) -> Self

Indicates whether the data in the cluster is encrypted at rest.

pub fn set_encrypted(self, input: Option<bool>) -> Self

pub fn endpoint(self, input: AwsRedshiftClusterEndpoint) -> Self

The connection endpoint.

pub fn set_endpoint(self, input: Option<AwsRedshiftClusterEndpoint>) -> Self

pub fn enhanced_vpc_routing(self, input: bool) -> Self

Indicates whether to create the cluster with enhanced VPC routing enabled.

pub fn set_enhanced_vpc_routing(self, input: Option<bool>) -> Self

pub fn expected_next_snapshot_schedule_time(
self,
input: impl Into<String>
) -> Self

Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_expected_next_snapshot_schedule_time(
self,
input: Option<String>
) -> Self

pub fn expected_next_snapshot_schedule_time_status(
self,
input: impl Into<String>
) -> Self

The status of the next expected snapshot.

Valid values: OnTrack | Pending

pub fn set_expected_next_snapshot_schedule_time_status(
self,
input: Option<String>
) -> Self

pub fn hsm_status(self, input: AwsRedshiftClusterHsmStatus) -> Self

Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

pub fn set_hsm_status(self, input: Option<AwsRedshiftClusterHsmStatus>) -> Self

pub fn iam_roles(self, input: impl Into<AwsRedshiftClusterIamRole>) -> Self

pub fn set_iam_roles(
self,
input: Option<Vec<AwsRedshiftClusterIamRole>>
) -> Self

pub fn kms_key_id(self, input: impl Into<String>) -> Self

The identifier of the KMS encryption key that is used to encrypt data in the cluster.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

pub fn maintenance_track_name(self, input: impl Into<String>) -> Self

The name of the maintenance track for the cluster.

pub fn set_maintenance_track_name(self, input: Option<String>) -> Self

pub fn manual_snapshot_retention_period(self, input: i32) -> Self

The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

pub fn set_manual_snapshot_retention_period(self, input: Option<i32>) -> Self

pub fn master_username(self, input: impl Into<String>) -> Self

The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

pub fn set_master_username(self, input: Option<String>) -> Self

pub fn next_maintenance_window_start_time(
self,
input: impl Into<String>
) -> Self

Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_next_maintenance_window_start_time(
self,
input: Option<String>
) -> Self

pub fn node_type(self, input: impl Into<String>) -> Self

The node type for the nodes in the cluster.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn number_of_nodes(self, input: i32) -> Self

The number of compute nodes in the cluster.

pub fn set_number_of_nodes(self, input: Option<i32>) -> Self

pub fn pending_actions(self, input: impl Into<String>) -> Self

pub fn set_pending_actions(self, input: Option<Vec<String>>) -> Self

pub fn pending_modified_values(
self,
input: AwsRedshiftClusterPendingModifiedValues
) -> Self

A list of changes to the cluster that are currently pending.

pub fn set_pending_modified_values(
self,
input: Option<AwsRedshiftClusterPendingModifiedValues>
) -> Self

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: :HH:MM-:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn publicly_accessible(self, input: bool) -> Self

Whether the cluster can be accessed from a public network.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

pub fn resize_info(self, input: AwsRedshiftClusterResizeInfo) -> Self

Information about the resize operation for the cluster.

pub fn set_resize_info(
self,
input: Option<AwsRedshiftClusterResizeInfo>
) -> Self

pub fn restore_status(self, input: AwsRedshiftClusterRestoreStatus) -> Self

Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

pub fn set_restore_status(
self,
input: Option<AwsRedshiftClusterRestoreStatus>
) -> Self

pub fn snapshot_schedule_identifier(self, input: impl Into<String>) -> Self

A unique identifier for the cluster snapshot schedule.

pub fn set_snapshot_schedule_identifier(self, input: Option<String>) -> Self

pub fn snapshot_schedule_state(self, input: impl Into<String>) -> Self

The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

pub fn set_snapshot_schedule_state(self, input: Option<String>) -> Self

pub fn vpc_id(self, input: impl Into<String>) -> Self

The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

pub fn set_vpc_id(self, input: Option<String>) -> Self

pub fn vpc_security_groups(
self,
input: impl Into<AwsRedshiftClusterVpcSecurityGroup>
) -> Self

pub fn set_vpc_security_groups(
self,
input: Option<Vec<AwsRedshiftClusterVpcSecurityGroup>>
) -> Self

pub fn build(self) -> AwsRedshiftClusterDetails

Consumes the builder and constructs a AwsRedshiftClusterDetails