Struct aws_sdk_securityhub::model::aws_s3_bucket_bucket_lifecycle_configuration_rules_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsS3BucketBucketLifecycleConfigurationRulesDetails

Implementations

impl Builder

pub fn abort_incomplete_multipart_upload(
self,
input: AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails
) -> Self

How Amazon S3 responds when a multipart upload is incomplete. Specifically, provides a number of days before Amazon S3 cancels the entire upload.

pub fn set_abort_incomplete_multipart_upload(
self,
input: Option<AwsS3BucketBucketLifecycleConfigurationRulesAbortIncompleteMultipartUploadDetails>
) -> Self

pub fn expiration_date(self, input: impl Into<String>) -> Self

The date when objects are moved or deleted.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_expiration_date(self, input: Option<String>) -> Self

pub fn expiration_in_days(self, input: i32) -> Self

The length in days of the lifetime for objects that are subject to the rule.

pub fn set_expiration_in_days(self, input: Option<i32>) -> Self

pub fn expired_object_delete_marker(self, input: bool) -> Self

Whether Amazon S3 removes a delete marker that has no noncurrent versions. If set to true, the delete marker is expired. If set to false, the policy takes no action.

If you provide ExpiredObjectDeleteMarker, you cannot provide ExpirationInDays or ExpirationDate.

pub fn set_expired_object_delete_marker(self, input: Option<bool>) -> Self

pub fn filter(
self,
input: AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails
) -> Self

Identifies the objects that a rule applies to.

pub fn set_filter(
self,
input: Option<AwsS3BucketBucketLifecycleConfigurationRulesFilterDetails>
) -> Self

pub fn id(self, input: impl Into<String>) -> Self

The unique identifier of the rule.

pub fn set_id(self, input: Option<String>) -> Self

pub fn noncurrent_version_expiration_in_days(self, input: i32) -> Self

The number of days that an object is noncurrent before Amazon S3 can perform the associated action.

pub fn set_noncurrent_version_expiration_in_days(
self,
input: Option<i32>
) -> Self

pub fn noncurrent_version_transitions(
self,
input: impl Into<AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails>
) -> Self

pub fn set_noncurrent_version_transitions(
self,
input: Option<Vec<AwsS3BucketBucketLifecycleConfigurationRulesNoncurrentVersionTransitionsDetails>>
) -> Self

pub fn prefix(self, input: impl Into<String>) -> Self

A prefix that identifies one or more objects that the rule applies to.

pub fn set_prefix(self, input: Option<String>) -> Self

pub fn status(self, input: impl Into<String>) -> Self

The current status of the rule. Indicates whether the rule is currently being applied.

pub fn set_status(self, input: Option<String>) -> Self

pub fn transitions(
self,
input: impl Into<AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails>
) -> Self

pub fn set_transitions(
self,
input: Option<Vec<AwsS3BucketBucketLifecycleConfigurationRulesTransitionsDetails>>
) -> Self

pub fn build(self) -> AwsS3BucketBucketLifecycleConfigurationRulesDetails

Consumes the builder and constructs a AwsS3BucketBucketLifecycleConfigurationRulesDetails