Struct aws_sdk_securityhub::model::aws_security_finding::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsSecurityFinding

Implementations

impl Builder

pub fn schema_version(self, input: impl Into<String>) -> Self

The schema version that a finding is formatted for.

pub fn set_schema_version(self, input: Option<String>) -> Self

pub fn id(self, input: impl Into<String>) -> Self

The security findings provider-specific identifier for a finding.

pub fn set_id(self, input: Option<String>) -> Self

pub fn product_arn(self, input: impl Into<String>) -> Self

The ARN generated by Security Hub that uniquely identifies a product that generates findings. This can be the ARN for a third-party product that is integrated with Security Hub, or the ARN for a custom integration.

pub fn set_product_arn(self, input: Option<String>) -> Self

pub fn product_name(self, input: impl Into<String>) -> Self

The name of the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by product name, you use this attribute.

When you use the Security Hub API to filter findings by product name, you use the aws/securityhub/ProductName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

pub fn set_product_name(self, input: Option<String>) -> Self

pub fn company_name(self, input: impl Into<String>) -> Self

The name of the company for the product that generated the finding.

Security Hub populates this attribute automatically for each finding. You cannot be updated using BatchImportFindings or BatchUpdateFindings. The exception to this is when you use a custom integration.

When you use the Security Hub console to filter findings by company name, you use this attribute.

When you use the Security Hub API to filter findings by company name, you use the aws/securityhub/CompanyName attribute under ProductFields.

Security Hub does not synchronize those two attributes.

pub fn set_company_name(self, input: Option<String>) -> Self

pub fn region(self, input: impl Into<String>) -> Self

The Region from which the finding was generated.

Security Hub populates this attribute automatically for each finding. You cannot update it using BatchImportFindings or BatchUpdateFindings.

pub fn set_region(self, input: Option<String>) -> Self

pub fn generator_id(self, input: impl Into<String>) -> Self

The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

pub fn set_generator_id(self, input: Option<String>) -> Self

pub fn aws_account_id(self, input: impl Into<String>) -> Self

The Amazon Web Services account ID that a finding is generated in.

pub fn set_aws_account_id(self, input: Option<String>) -> Self

pub fn types(self, input: impl Into<String>) -> Self

pub fn set_types(self, input: Option<Vec<String>>) -> Self

pub fn first_observed_at(self, input: impl Into<String>) -> Self

Indicates when the security-findings provider first observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_first_observed_at(self, input: Option<String>) -> Self

pub fn last_observed_at(self, input: impl Into<String>) -> Self

Indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_last_observed_at(self, input: Option<String>) -> Self

pub fn created_at(self, input: impl Into<String>) -> Self

Indicates when the security-findings provider created the potential security issue that a finding captured.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_created_at(self, input: Option<String>) -> Self

pub fn updated_at(self, input: impl Into<String>) -> Self

Indicates when the security-findings provider last updated the finding record.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_updated_at(self, input: Option<String>) -> Self

pub fn severity(self, input: Severity) -> Self

A finding's severity.

pub fn set_severity(self, input: Option<Severity>) -> Self

pub fn confidence(self, input: i32) -> Self

A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn set_confidence(self, input: Option<i32>) -> Self

pub fn criticality(self, input: i32) -> Self

The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn set_criticality(self, input: Option<i32>) -> Self

pub fn title(self, input: impl Into<String>) -> Self

A finding's title.

In this release, Title is a required property.

pub fn set_title(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A finding's description.

In this release, Description is a required property.

pub fn set_description(self, input: Option<String>) -> Self

pub fn remediation(self, input: Remediation) -> Self

A data type that describes the remediation options for a finding.

pub fn set_remediation(self, input: Option<Remediation>) -> Self

pub fn source_url(self, input: impl Into<String>) -> Self

A URL that links to a page about the current finding in the security-findings provider's solution.

pub fn set_source_url(self, input: Option<String>) -> Self

pub fn product_fields(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_product_fields(self, input: Option<HashMap<String, String>>) -> Self

pub fn user_defined_fields(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_user_defined_fields(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn malware(self, input: impl Into<Malware>) -> Self

pub fn set_malware(self, input: Option<Vec<Malware>>) -> Self

pub fn network(self, input: Network) -> Self

The details of network-related information about a finding.

pub fn set_network(self, input: Option<Network>) -> Self

pub fn network_path(self, input: impl Into<NetworkPathComponent>) -> Self

pub fn set_network_path(self, input: Option<Vec<NetworkPathComponent>>) -> Self

pub fn process(self, input: ProcessDetails) -> Self

The details of process-related information about a finding.

pub fn set_process(self, input: Option<ProcessDetails>) -> Self

pub fn threat_intel_indicators(
self,
input: impl Into<ThreatIntelIndicator>
) -> Self

pub fn set_threat_intel_indicators(
self,
input: Option<Vec<ThreatIntelIndicator>>
) -> Self

pub fn resources(self, input: impl Into<Resource>) -> Self

pub fn set_resources(self, input: Option<Vec<Resource>>) -> Self

pub fn compliance(self, input: Compliance) -> Self

This data type is exclusive to findings that are generated as the result of a check run against a specific rule in a supported security standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

pub fn set_compliance(self, input: Option<Compliance>) -> Self

pub fn verification_state(self, input: VerificationState) -> Self

Indicates the veracity of a finding.

pub fn set_verification_state(self, input: Option<VerificationState>) -> Self

pub fn workflow_state(self, input: WorkflowState) -> Self

The workflow state of a finding.

pub fn set_workflow_state(self, input: Option<WorkflowState>) -> Self

pub fn workflow(self, input: Workflow) -> Self

Provides information about the status of the investigation into a finding.

pub fn set_workflow(self, input: Option<Workflow>) -> Self

pub fn record_state(self, input: RecordState) -> Self

The record state of a finding.

pub fn set_record_state(self, input: Option<RecordState>) -> Self

pub fn related_findings(self, input: impl Into<RelatedFinding>) -> Self

pub fn set_related_findings(self, input: Option<Vec<RelatedFinding>>) -> Self

pub fn note(self, input: Note) -> Self

A user-defined note added to a finding.

pub fn set_note(self, input: Option<Note>) -> Self

pub fn vulnerabilities(self, input: impl Into<Vulnerability>) -> Self

pub fn set_vulnerabilities(self, input: Option<Vec<Vulnerability>>) -> Self

pub fn patch_summary(self, input: PatchSummary) -> Self

Provides an overview of the patch compliance status for an instance against a selected compliance standard.

pub fn set_patch_summary(self, input: Option<PatchSummary>) -> Self

pub fn action(self, input: Action) -> Self

Provides details about an action that affects or that was taken on a resource.

pub fn set_action(self, input: Option<Action>) -> Self

pub fn finding_provider_fields(self, input: FindingProviderFields) -> Self

In a BatchImportFindings request, finding providers use FindingProviderFields to provide and update their own values for confidence, criticality, related findings, severity, and types.

pub fn set_finding_provider_fields(
self,
input: Option<FindingProviderFields>
) -> Self

pub fn build(self) -> AwsSecurityFinding

Consumes the builder and constructs a AwsSecurityFinding