Struct aws_sdk_securityhub::model::aws_security_finding_filters::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsSecurityFindingFilters

Implementations

impl Builder

pub fn product_arn(self, input: impl Into<StringFilter>) -> Self

Appends an item to product_arn.

To override the contents of this collection use set_product_arn.

The ARN generated by Security Hub that uniquely identifies a third-party company (security findings provider) after this provider's product (solution that generates findings) is registered with Security Hub.

pub fn set_product_arn(self, input: Option<Vec<StringFilter>>) -> Self

The ARN generated by Security Hub that uniquely identifies a third-party company (security findings provider) after this provider's product (solution that generates findings) is registered with Security Hub.

pub fn aws_account_id(self, input: impl Into<StringFilter>) -> Self

Appends an item to aws_account_id.

To override the contents of this collection use set_aws_account_id.

The Amazon Web Services account ID that a finding is generated in.

pub fn set_aws_account_id(self, input: Option<Vec<StringFilter>>) -> Self

The Amazon Web Services account ID that a finding is generated in.

pub fn id(self, input: impl Into<StringFilter>) -> Self

Appends an item to id.

To override the contents of this collection use set_id.

The security findings provider-specific identifier for a finding.

pub fn set_id(self, input: Option<Vec<StringFilter>>) -> Self

The security findings provider-specific identifier for a finding.

pub fn generator_id(self, input: impl Into<StringFilter>) -> Self

Appends an item to generator_id.

To override the contents of this collection use set_generator_id.

The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

pub fn set_generator_id(self, input: Option<Vec<StringFilter>>) -> Self

The identifier for the solution-specific component (a discrete unit of logic) that generated a finding. In various security-findings providers' solutions, this generator can be called a rule, a check, a detector, a plugin, etc.

pub fn region(self, input: impl Into<StringFilter>) -> Self

Appends an item to region.

To override the contents of this collection use set_region.

The Region from which the finding was generated.

pub fn set_region(self, input: Option<Vec<StringFilter>>) -> Self

The Region from which the finding was generated.

pub fn type(self, input: impl Into<StringFilter>) -> Self

Appends an item to r#type.

To override the contents of this collection use set_type.

A finding type in the format of namespace/category/classifier that classifies a finding.

pub fn set_type(self, input: Option<Vec<StringFilter>>) -> Self

A finding type in the format of namespace/category/classifier that classifies a finding.

pub fn first_observed_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to first_observed_at.

To override the contents of this collection use set_first_observed_at.

An ISO8601-formatted timestamp that indicates when the security-findings provider first observed the potential security issue that a finding captured.

pub fn set_first_observed_at(self, input: Option<Vec<DateFilter>>) -> Self

An ISO8601-formatted timestamp that indicates when the security-findings provider first observed the potential security issue that a finding captured.

pub fn last_observed_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to last_observed_at.

To override the contents of this collection use set_last_observed_at.

An ISO8601-formatted timestamp that indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

pub fn set_last_observed_at(self, input: Option<Vec<DateFilter>>) -> Self

An ISO8601-formatted timestamp that indicates when the security-findings provider most recently observed the potential security issue that a finding captured.

pub fn created_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to created_at.

To override the contents of this collection use set_created_at.

An ISO8601-formatted timestamp that indicates when the security-findings provider captured the potential security issue that a finding captured.

pub fn set_created_at(self, input: Option<Vec<DateFilter>>) -> Self

An ISO8601-formatted timestamp that indicates when the security-findings provider captured the potential security issue that a finding captured.

pub fn updated_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to updated_at.

To override the contents of this collection use set_updated_at.

An ISO8601-formatted timestamp that indicates when the security-findings provider last updated the finding record.

pub fn set_updated_at(self, input: Option<Vec<DateFilter>>) -> Self

An ISO8601-formatted timestamp that indicates when the security-findings provider last updated the finding record.

pub fn severity_product(self, input: impl Into<NumberFilter>) -> Self

Appends an item to severity_product.

To override the contents of this collection use set_severity_product.

The native severity as defined by the security-findings provider's solution that generated the finding.

pub fn set_severity_product(self, input: Option<Vec<NumberFilter>>) -> Self

The native severity as defined by the security-findings provider's solution that generated the finding.

pub fn severity_normalized(self, input: impl Into<NumberFilter>) -> Self

Appends an item to severity_normalized.

To override the contents of this collection use set_severity_normalized.

The normalized severity of a finding.

pub fn set_severity_normalized(self, input: Option<Vec<NumberFilter>>) -> Self

The normalized severity of a finding.

pub fn severity_label(self, input: impl Into<StringFilter>) -> Self

Appends an item to severity_label.

To override the contents of this collection use set_severity_label.

The label of a finding's severity.

pub fn set_severity_label(self, input: Option<Vec<StringFilter>>) -> Self

The label of a finding's severity.

pub fn confidence(self, input: impl Into<NumberFilter>) -> Self

Appends an item to confidence.

To override the contents of this collection use set_confidence.

A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn set_confidence(self, input: Option<Vec<NumberFilter>>) -> Self

A finding's confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn criticality(self, input: impl Into<NumberFilter>) -> Self

Appends an item to criticality.

To override the contents of this collection use set_criticality.

The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn set_criticality(self, input: Option<Vec<NumberFilter>>) -> Self

The level of importance assigned to the resources associated with the finding.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn title(self, input: impl Into<StringFilter>) -> Self

Appends an item to title.

To override the contents of this collection use set_title.

A finding's title.

pub fn set_title(self, input: Option<Vec<StringFilter>>) -> Self

A finding's title.

pub fn description(self, input: impl Into<StringFilter>) -> Self

Appends an item to description.

To override the contents of this collection use set_description.

A finding's description.

pub fn set_description(self, input: Option<Vec<StringFilter>>) -> Self

A finding's description.

pub fn recommendation_text(self, input: impl Into<StringFilter>) -> Self

Appends an item to recommendation_text.

To override the contents of this collection use set_recommendation_text.

The recommendation of what to do about the issue described in a finding.

pub fn set_recommendation_text(self, input: Option<Vec<StringFilter>>) -> Self

The recommendation of what to do about the issue described in a finding.

pub fn source_url(self, input: impl Into<StringFilter>) -> Self

Appends an item to source_url.

To override the contents of this collection use set_source_url.

A URL that links to a page about the current finding in the security-findings provider's solution.

pub fn set_source_url(self, input: Option<Vec<StringFilter>>) -> Self

A URL that links to a page about the current finding in the security-findings provider's solution.

pub fn product_fields(self, input: impl Into<MapFilter>) -> Self

Appends an item to product_fields.

To override the contents of this collection use set_product_fields.

A data type where security-findings providers can include additional solution-specific details that aren't part of the defined AwsSecurityFinding format.

pub fn set_product_fields(self, input: Option<Vec<MapFilter>>) -> Self

A data type where security-findings providers can include additional solution-specific details that aren't part of the defined AwsSecurityFinding format.

pub fn product_name(self, input: impl Into<StringFilter>) -> Self

Appends an item to product_name.

To override the contents of this collection use set_product_name.

The name of the solution (product) that generates findings.

Note that this is a filter against the aws/securityhub/ProductName field in ProductFields. It is not a filter for the top-level ProductName field.

pub fn set_product_name(self, input: Option<Vec<StringFilter>>) -> Self

The name of the solution (product) that generates findings.

Note that this is a filter against the aws/securityhub/ProductName field in ProductFields. It is not a filter for the top-level ProductName field.

pub fn company_name(self, input: impl Into<StringFilter>) -> Self

Appends an item to company_name.

To override the contents of this collection use set_company_name.

The name of the findings provider (company) that owns the solution (product) that generates findings.

Note that this is a filter against the aws/securityhub/CompanyName field in ProductFields. It is not a filter for the top-level CompanyName field.

pub fn set_company_name(self, input: Option<Vec<StringFilter>>) -> Self

The name of the findings provider (company) that owns the solution (product) that generates findings.

Note that this is a filter against the aws/securityhub/CompanyName field in ProductFields. It is not a filter for the top-level CompanyName field.

pub fn user_defined_fields(self, input: impl Into<MapFilter>) -> Self

Appends an item to user_defined_fields.

To override the contents of this collection use set_user_defined_fields.

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

pub fn set_user_defined_fields(self, input: Option<Vec<MapFilter>>) -> Self

A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

pub fn malware_name(self, input: impl Into<StringFilter>) -> Self

Appends an item to malware_name.

To override the contents of this collection use set_malware_name.

The name of the malware that was observed.

pub fn set_malware_name(self, input: Option<Vec<StringFilter>>) -> Self

The name of the malware that was observed.

pub fn malware_type(self, input: impl Into<StringFilter>) -> Self

Appends an item to malware_type.

To override the contents of this collection use set_malware_type.

The type of the malware that was observed.

pub fn set_malware_type(self, input: Option<Vec<StringFilter>>) -> Self

The type of the malware that was observed.

pub fn malware_path(self, input: impl Into<StringFilter>) -> Self

Appends an item to malware_path.

To override the contents of this collection use set_malware_path.

The filesystem path of the malware that was observed.

pub fn set_malware_path(self, input: Option<Vec<StringFilter>>) -> Self

The filesystem path of the malware that was observed.

pub fn malware_state(self, input: impl Into<StringFilter>) -> Self

Appends an item to malware_state.

To override the contents of this collection use set_malware_state.

The state of the malware that was observed.

pub fn set_malware_state(self, input: Option<Vec<StringFilter>>) -> Self

The state of the malware that was observed.

pub fn network_direction(self, input: impl Into<StringFilter>) -> Self

Appends an item to network_direction.

To override the contents of this collection use set_network_direction.

Indicates the direction of network traffic associated with a finding.

pub fn set_network_direction(self, input: Option<Vec<StringFilter>>) -> Self

Indicates the direction of network traffic associated with a finding.

pub fn network_protocol(self, input: impl Into<StringFilter>) -> Self

Appends an item to network_protocol.

To override the contents of this collection use set_network_protocol.

The protocol of network-related information about a finding.

pub fn set_network_protocol(self, input: Option<Vec<StringFilter>>) -> Self

The protocol of network-related information about a finding.

pub fn network_source_ip_v4(self, input: impl Into<IpFilter>) -> Self

Appends an item to network_source_ip_v4.

To override the contents of this collection use set_network_source_ip_v4.

The source IPv4 address of network-related information about a finding.

pub fn set_network_source_ip_v4(self, input: Option<Vec<IpFilter>>) -> Self

The source IPv4 address of network-related information about a finding.

pub fn network_source_ip_v6(self, input: impl Into<IpFilter>) -> Self

Appends an item to network_source_ip_v6.

To override the contents of this collection use set_network_source_ip_v6.

The source IPv6 address of network-related information about a finding.

pub fn set_network_source_ip_v6(self, input: Option<Vec<IpFilter>>) -> Self

The source IPv6 address of network-related information about a finding.

pub fn network_source_port(self, input: impl Into<NumberFilter>) -> Self

Appends an item to network_source_port.

To override the contents of this collection use set_network_source_port.

The source port of network-related information about a finding.

pub fn set_network_source_port(self, input: Option<Vec<NumberFilter>>) -> Self

The source port of network-related information about a finding.

pub fn network_source_domain(self, input: impl Into<StringFilter>) -> Self

Appends an item to network_source_domain.

To override the contents of this collection use set_network_source_domain.

The source domain of network-related information about a finding.

pub fn set_network_source_domain(self, input: Option<Vec<StringFilter>>) -> Self

The source domain of network-related information about a finding.

pub fn network_source_mac(self, input: impl Into<StringFilter>) -> Self

Appends an item to network_source_mac.

To override the contents of this collection use set_network_source_mac.

The source media access control (MAC) address of network-related information about a finding.

pub fn set_network_source_mac(self, input: Option<Vec<StringFilter>>) -> Self

The source media access control (MAC) address of network-related information about a finding.

pub fn network_destination_ip_v4(self, input: impl Into<IpFilter>) -> Self

Appends an item to network_destination_ip_v4.

To override the contents of this collection use set_network_destination_ip_v4.

The destination IPv4 address of network-related information about a finding.

pub fn set_network_destination_ip_v4(self, input: Option<Vec<IpFilter>>) -> Self

The destination IPv4 address of network-related information about a finding.

pub fn network_destination_ip_v6(self, input: impl Into<IpFilter>) -> Self

Appends an item to network_destination_ip_v6.

To override the contents of this collection use set_network_destination_ip_v6.

The destination IPv6 address of network-related information about a finding.

pub fn set_network_destination_ip_v6(self, input: Option<Vec<IpFilter>>) -> Self

The destination IPv6 address of network-related information about a finding.

pub fn network_destination_port(self, input: impl Into<NumberFilter>) -> Self

Appends an item to network_destination_port.

To override the contents of this collection use set_network_destination_port.

The destination port of network-related information about a finding.

pub fn set_network_destination_port(
self,
input: Option<Vec<NumberFilter>>
) -> Self

The destination port of network-related information about a finding.

pub fn network_destination_domain(self, input: impl Into<StringFilter>) -> Self

Appends an item to network_destination_domain.

To override the contents of this collection use set_network_destination_domain.

The destination domain of network-related information about a finding.

pub fn set_network_destination_domain(
self,
input: Option<Vec<StringFilter>>
) -> Self

The destination domain of network-related information about a finding.

pub fn process_name(self, input: impl Into<StringFilter>) -> Self

Appends an item to process_name.

To override the contents of this collection use set_process_name.

The name of the process.

pub fn set_process_name(self, input: Option<Vec<StringFilter>>) -> Self

The name of the process.

pub fn process_path(self, input: impl Into<StringFilter>) -> Self

Appends an item to process_path.

To override the contents of this collection use set_process_path.

The path to the process executable.

pub fn set_process_path(self, input: Option<Vec<StringFilter>>) -> Self

The path to the process executable.

pub fn process_pid(self, input: impl Into<NumberFilter>) -> Self

Appends an item to process_pid.

To override the contents of this collection use set_process_pid.

The process ID.

pub fn set_process_pid(self, input: Option<Vec<NumberFilter>>) -> Self

The process ID.

pub fn process_parent_pid(self, input: impl Into<NumberFilter>) -> Self

Appends an item to process_parent_pid.

To override the contents of this collection use set_process_parent_pid.

The parent process ID.

pub fn set_process_parent_pid(self, input: Option<Vec<NumberFilter>>) -> Self

The parent process ID.

pub fn process_launched_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to process_launched_at.

To override the contents of this collection use set_process_launched_at.

The date/time that the process was launched.

pub fn set_process_launched_at(self, input: Option<Vec<DateFilter>>) -> Self

The date/time that the process was launched.

pub fn process_terminated_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to process_terminated_at.

To override the contents of this collection use set_process_terminated_at.

The date/time that the process was terminated.

pub fn set_process_terminated_at(self, input: Option<Vec<DateFilter>>) -> Self

The date/time that the process was terminated.

pub fn threat_intel_indicator_type(self, input: impl Into<StringFilter>) -> Self

Appends an item to threat_intel_indicator_type.

To override the contents of this collection use set_threat_intel_indicator_type.

The type of a threat intelligence indicator.

pub fn set_threat_intel_indicator_type(
self,
input: Option<Vec<StringFilter>>
) -> Self

The type of a threat intelligence indicator.

pub fn threat_intel_indicator_value(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to threat_intel_indicator_value.

To override the contents of this collection use set_threat_intel_indicator_value.

The value of a threat intelligence indicator.

pub fn set_threat_intel_indicator_value(
self,
input: Option<Vec<StringFilter>>
) -> Self

The value of a threat intelligence indicator.

pub fn threat_intel_indicator_category(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to threat_intel_indicator_category.

To override the contents of this collection use set_threat_intel_indicator_category.

The category of a threat intelligence indicator.

pub fn set_threat_intel_indicator_category(
self,
input: Option<Vec<StringFilter>>
) -> Self

The category of a threat intelligence indicator.

pub fn threat_intel_indicator_last_observed_at(
self,
input: impl Into<DateFilter>
) -> Self

Appends an item to threat_intel_indicator_last_observed_at.

To override the contents of this collection use set_threat_intel_indicator_last_observed_at.

The date/time of the last observation of a threat intelligence indicator.

pub fn set_threat_intel_indicator_last_observed_at(
self,
input: Option<Vec<DateFilter>>
) -> Self

The date/time of the last observation of a threat intelligence indicator.

pub fn threat_intel_indicator_source(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to threat_intel_indicator_source.

To override the contents of this collection use set_threat_intel_indicator_source.

The source of the threat intelligence.

pub fn set_threat_intel_indicator_source(
self,
input: Option<Vec<StringFilter>>
) -> Self

The source of the threat intelligence.

pub fn threat_intel_indicator_source_url(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to threat_intel_indicator_source_url.

To override the contents of this collection use set_threat_intel_indicator_source_url.

The URL for more details from the source of the threat intelligence.

pub fn set_threat_intel_indicator_source_url(
self,
input: Option<Vec<StringFilter>>
) -> Self

The URL for more details from the source of the threat intelligence.

pub fn resource_type(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_type.

To override the contents of this collection use set_resource_type.

Specifies the type of the resource that details are provided for.

pub fn set_resource_type(self, input: Option<Vec<StringFilter>>) -> Self

Specifies the type of the resource that details are provided for.

pub fn resource_id(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_id.

To override the contents of this collection use set_resource_id.

The canonical identifier for the given resource type.

pub fn set_resource_id(self, input: Option<Vec<StringFilter>>) -> Self

The canonical identifier for the given resource type.

pub fn resource_partition(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_partition.

To override the contents of this collection use set_resource_partition.

The canonical Amazon Web Services partition name that the Region is assigned to.

pub fn set_resource_partition(self, input: Option<Vec<StringFilter>>) -> Self

The canonical Amazon Web Services partition name that the Region is assigned to.

pub fn resource_region(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_region.

To override the contents of this collection use set_resource_region.

The canonical Amazon Web Services external Region name where this resource is located.

pub fn set_resource_region(self, input: Option<Vec<StringFilter>>) -> Self

The canonical Amazon Web Services external Region name where this resource is located.

pub fn resource_tags(self, input: impl Into<MapFilter>) -> Self

Appends an item to resource_tags.

To override the contents of this collection use set_resource_tags.

A list of Amazon Web Services tags associated with a resource at the time the finding was processed.

pub fn set_resource_tags(self, input: Option<Vec<MapFilter>>) -> Self

A list of Amazon Web Services tags associated with a resource at the time the finding was processed.

pub fn resource_aws_ec2_instance_type(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_type.

To override the contents of this collection use set_resource_aws_ec2_instance_type.

The instance type of the instance.

pub fn set_resource_aws_ec2_instance_type(
self,
input: Option<Vec<StringFilter>>
) -> Self

The instance type of the instance.

pub fn resource_aws_ec2_instance_image_id(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_image_id.

To override the contents of this collection use set_resource_aws_ec2_instance_image_id.

The Amazon Machine Image (AMI) ID of the instance.

pub fn set_resource_aws_ec2_instance_image_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The Amazon Machine Image (AMI) ID of the instance.

pub fn resource_aws_ec2_instance_ip_v4_addresses(
self,
input: impl Into<IpFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_ip_v4_addresses.

To override the contents of this collection use set_resource_aws_ec2_instance_ip_v4_addresses.

The IPv4 addresses associated with the instance.

pub fn set_resource_aws_ec2_instance_ip_v4_addresses(
self,
input: Option<Vec<IpFilter>>
) -> Self

The IPv4 addresses associated with the instance.

pub fn resource_aws_ec2_instance_ip_v6_addresses(
self,
input: impl Into<IpFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_ip_v6_addresses.

To override the contents of this collection use set_resource_aws_ec2_instance_ip_v6_addresses.

The IPv6 addresses associated with the instance.

pub fn set_resource_aws_ec2_instance_ip_v6_addresses(
self,
input: Option<Vec<IpFilter>>
) -> Self

The IPv6 addresses associated with the instance.

pub fn resource_aws_ec2_instance_key_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_key_name.

To override the contents of this collection use set_resource_aws_ec2_instance_key_name.

The key name associated with the instance.

pub fn set_resource_aws_ec2_instance_key_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The key name associated with the instance.

pub fn resource_aws_ec2_instance_iam_instance_profile_arn(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_iam_instance_profile_arn.

To override the contents of this collection use set_resource_aws_ec2_instance_iam_instance_profile_arn.

The IAM profile ARN of the instance.

pub fn set_resource_aws_ec2_instance_iam_instance_profile_arn(
self,
input: Option<Vec<StringFilter>>
) -> Self

The IAM profile ARN of the instance.

pub fn resource_aws_ec2_instance_vpc_id(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_vpc_id.

To override the contents of this collection use set_resource_aws_ec2_instance_vpc_id.

The identifier of the VPC that the instance was launched in.

pub fn set_resource_aws_ec2_instance_vpc_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The identifier of the VPC that the instance was launched in.

pub fn resource_aws_ec2_instance_subnet_id(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_subnet_id.

To override the contents of this collection use set_resource_aws_ec2_instance_subnet_id.

The identifier of the subnet that the instance was launched in.

pub fn set_resource_aws_ec2_instance_subnet_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The identifier of the subnet that the instance was launched in.

pub fn resource_aws_ec2_instance_launched_at(
self,
input: impl Into<DateFilter>
) -> Self

Appends an item to resource_aws_ec2_instance_launched_at.

To override the contents of this collection use set_resource_aws_ec2_instance_launched_at.

The date and time the instance was launched.

pub fn set_resource_aws_ec2_instance_launched_at(
self,
input: Option<Vec<DateFilter>>
) -> Self

The date and time the instance was launched.

pub fn resource_aws_s3_bucket_owner_id(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_s3_bucket_owner_id.

To override the contents of this collection use set_resource_aws_s3_bucket_owner_id.

The canonical user ID of the owner of the S3 bucket.

pub fn set_resource_aws_s3_bucket_owner_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The canonical user ID of the owner of the S3 bucket.

pub fn resource_aws_s3_bucket_owner_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_s3_bucket_owner_name.

To override the contents of this collection use set_resource_aws_s3_bucket_owner_name.

The display name of the owner of the S3 bucket.

pub fn set_resource_aws_s3_bucket_owner_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The display name of the owner of the S3 bucket.

pub fn resource_aws_iam_access_key_user_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_iam_access_key_user_name.

To override the contents of this collection use set_resource_aws_iam_access_key_user_name.

The user associated with the IAM access key related to a finding.

pub fn set_resource_aws_iam_access_key_user_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The user associated with the IAM access key related to a finding.

pub fn resource_aws_iam_access_key_principal_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_iam_access_key_principal_name.

To override the contents of this collection use set_resource_aws_iam_access_key_principal_name.

The name of the principal that is associated with an IAM access key.

pub fn set_resource_aws_iam_access_key_principal_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The name of the principal that is associated with an IAM access key.

pub fn resource_aws_iam_access_key_status(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_iam_access_key_status.

To override the contents of this collection use set_resource_aws_iam_access_key_status.

The status of the IAM access key related to a finding.

pub fn set_resource_aws_iam_access_key_status(
self,
input: Option<Vec<StringFilter>>
) -> Self

The status of the IAM access key related to a finding.

pub fn resource_aws_iam_access_key_created_at(
self,
input: impl Into<DateFilter>
) -> Self

Appends an item to resource_aws_iam_access_key_created_at.

To override the contents of this collection use set_resource_aws_iam_access_key_created_at.

The creation date/time of the IAM access key related to a finding.

pub fn set_resource_aws_iam_access_key_created_at(
self,
input: Option<Vec<DateFilter>>
) -> Self

The creation date/time of the IAM access key related to a finding.

pub fn resource_aws_iam_user_user_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_aws_iam_user_user_name.

To override the contents of this collection use set_resource_aws_iam_user_user_name.

The name of an IAM user.

pub fn set_resource_aws_iam_user_user_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The name of an IAM user.

pub fn resource_container_name(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_container_name.

To override the contents of this collection use set_resource_container_name.

The name of the container related to a finding.

pub fn set_resource_container_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The name of the container related to a finding.

pub fn resource_container_image_id(self, input: impl Into<StringFilter>) -> Self

Appends an item to resource_container_image_id.

To override the contents of this collection use set_resource_container_image_id.

The identifier of the image related to a finding.

pub fn set_resource_container_image_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The identifier of the image related to a finding.

pub fn resource_container_image_name(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to resource_container_image_name.

To override the contents of this collection use set_resource_container_image_name.

The name of the image related to a finding.

pub fn set_resource_container_image_name(
self,
input: Option<Vec<StringFilter>>
) -> Self

The name of the image related to a finding.

pub fn resource_container_launched_at(
self,
input: impl Into<DateFilter>
) -> Self

Appends an item to resource_container_launched_at.

To override the contents of this collection use set_resource_container_launched_at.

The date/time that the container was started.

pub fn set_resource_container_launched_at(
self,
input: Option<Vec<DateFilter>>
) -> Self

The date/time that the container was started.

pub fn resource_details_other(self, input: impl Into<MapFilter>) -> Self

Appends an item to resource_details_other.

To override the contents of this collection use set_resource_details_other.

The details of a resource that doesn't have a specific subfield for the resource type defined.

pub fn set_resource_details_other(self, input: Option<Vec<MapFilter>>) -> Self

The details of a resource that doesn't have a specific subfield for the resource type defined.

pub fn compliance_status(self, input: impl Into<StringFilter>) -> Self

Appends an item to compliance_status.

To override the contents of this collection use set_compliance_status.

Exclusive to findings that are generated as the result of a check run against a specific rule in a supported standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

pub fn set_compliance_status(self, input: Option<Vec<StringFilter>>) -> Self

Exclusive to findings that are generated as the result of a check run against a specific rule in a supported standard, such as CIS Amazon Web Services Foundations. Contains security standard-related finding details.

pub fn verification_state(self, input: impl Into<StringFilter>) -> Self

Appends an item to verification_state.

To override the contents of this collection use set_verification_state.

The veracity of a finding.

pub fn set_verification_state(self, input: Option<Vec<StringFilter>>) -> Self

The veracity of a finding.

pub fn workflow_state(self, input: impl Into<StringFilter>) -> Self

Appends an item to workflow_state.

To override the contents of this collection use set_workflow_state.

The workflow state of a finding.

Note that this field is deprecated. To search for a finding based on its workflow status, use WorkflowStatus.

pub fn set_workflow_state(self, input: Option<Vec<StringFilter>>) -> Self

The workflow state of a finding.

Note that this field is deprecated. To search for a finding based on its workflow status, use WorkflowStatus.

pub fn workflow_status(self, input: impl Into<StringFilter>) -> Self

Appends an item to workflow_status.

To override the contents of this collection use set_workflow_status.

The status of the investigation into a finding. Allowed values are the following.

NEW - The initial state of a finding, before it is reviewed.

Security Hub also resets the workflow status from NOTIFIED or RESOLVED to NEW in the following cases:

The record state changes from ARCHIVED to ACTIVE.

The compliance status changes from PASSED to either WARNING, FAILED, or NOT_AVAILABLE.

NOTIFIED - Indicates that the resource owner has been notified about the security issue. Used when the initial reviewer is not the resource owner, and needs intervention from the resource owner.

SUPPRESSED - The finding will not be reviewed again and will not be acted upon.

RESOLVED - The finding was reviewed and remediated and is now considered resolved.

pub fn set_workflow_status(self, input: Option<Vec<StringFilter>>) -> Self

The status of the investigation into a finding. Allowed values are the following.

NEW - The initial state of a finding, before it is reviewed.

Security Hub also resets the workflow status from NOTIFIED or RESOLVED to NEW in the following cases:

The record state changes from ARCHIVED to ACTIVE.

The compliance status changes from PASSED to either WARNING, FAILED, or NOT_AVAILABLE.

NOTIFIED - Indicates that the resource owner has been notified about the security issue. Used when the initial reviewer is not the resource owner, and needs intervention from the resource owner.

SUPPRESSED - The finding will not be reviewed again and will not be acted upon.

RESOLVED - The finding was reviewed and remediated and is now considered resolved.

pub fn record_state(self, input: impl Into<StringFilter>) -> Self

Appends an item to record_state.

To override the contents of this collection use set_record_state.

The updated record state for the finding.

pub fn set_record_state(self, input: Option<Vec<StringFilter>>) -> Self

The updated record state for the finding.

pub fn related_findings_product_arn(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to related_findings_product_arn.

To override the contents of this collection use set_related_findings_product_arn.

The ARN of the solution that generated a related finding.

pub fn set_related_findings_product_arn(
self,
input: Option<Vec<StringFilter>>
) -> Self

The ARN of the solution that generated a related finding.

pub fn related_findings_id(self, input: impl Into<StringFilter>) -> Self

Appends an item to related_findings_id.

To override the contents of this collection use set_related_findings_id.

The solution-generated identifier for a related finding.

pub fn set_related_findings_id(self, input: Option<Vec<StringFilter>>) -> Self

The solution-generated identifier for a related finding.

pub fn note_text(self, input: impl Into<StringFilter>) -> Self

Appends an item to note_text.

To override the contents of this collection use set_note_text.

The text of a note.

pub fn set_note_text(self, input: Option<Vec<StringFilter>>) -> Self

The text of a note.

pub fn note_updated_at(self, input: impl Into<DateFilter>) -> Self

Appends an item to note_updated_at.

To override the contents of this collection use set_note_updated_at.

The timestamp of when the note was updated.

pub fn set_note_updated_at(self, input: Option<Vec<DateFilter>>) -> Self

The timestamp of when the note was updated.

pub fn note_updated_by(self, input: impl Into<StringFilter>) -> Self

Appends an item to note_updated_by.

To override the contents of this collection use set_note_updated_by.

The principal that created a note.

pub fn set_note_updated_by(self, input: Option<Vec<StringFilter>>) -> Self

The principal that created a note.

pub fn keyword(self, input: impl Into<KeywordFilter>) -> Self

Appends an item to keyword.

To override the contents of this collection use set_keyword.

A keyword for a finding.

pub fn set_keyword(self, input: Option<Vec<KeywordFilter>>) -> Self

A keyword for a finding.

pub fn finding_provider_fields_confidence(
self,
input: impl Into<NumberFilter>
) -> Self

Appends an item to finding_provider_fields_confidence.

To override the contents of this collection use set_finding_provider_fields_confidence.

The finding provider value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn set_finding_provider_fields_confidence(
self,
input: Option<Vec<NumberFilter>>
) -> Self

The finding provider value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify.

Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

pub fn finding_provider_fields_criticality(
self,
input: impl Into<NumberFilter>
) -> Self

Appends an item to finding_provider_fields_criticality.

To override the contents of this collection use set_finding_provider_fields_criticality.

The finding provider value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn set_finding_provider_fields_criticality(
self,
input: Option<Vec<NumberFilter>>
) -> Self

The finding provider value for the level of importance assigned to the resources associated with the findings.

A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

pub fn finding_provider_fields_related_findings_id(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to finding_provider_fields_related_findings_id.

To override the contents of this collection use set_finding_provider_fields_related_findings_id.

The finding identifier of a related finding that is identified by the finding provider.

pub fn set_finding_provider_fields_related_findings_id(
self,
input: Option<Vec<StringFilter>>
) -> Self

The finding identifier of a related finding that is identified by the finding provider.

pub fn finding_provider_fields_related_findings_product_arn(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to finding_provider_fields_related_findings_product_arn.

To override the contents of this collection use set_finding_provider_fields_related_findings_product_arn.

The ARN of the solution that generated a related finding that is identified by the finding provider.

pub fn set_finding_provider_fields_related_findings_product_arn(
self,
input: Option<Vec<StringFilter>>
) -> Self

The ARN of the solution that generated a related finding that is identified by the finding provider.

pub fn finding_provider_fields_severity_label(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to finding_provider_fields_severity_label.

To override the contents of this collection use set_finding_provider_fields_severity_label.

The finding provider value for the severity label.

pub fn set_finding_provider_fields_severity_label(
self,
input: Option<Vec<StringFilter>>
) -> Self

The finding provider value for the severity label.

pub fn finding_provider_fields_severity_original(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to finding_provider_fields_severity_original.

To override the contents of this collection use set_finding_provider_fields_severity_original.

The finding provider's original value for the severity.

pub fn set_finding_provider_fields_severity_original(
self,
input: Option<Vec<StringFilter>>
) -> Self

The finding provider's original value for the severity.

pub fn finding_provider_fields_types(
self,
input: impl Into<StringFilter>
) -> Self

Appends an item to finding_provider_fields_types.

To override the contents of this collection use set_finding_provider_fields_types.

One or more finding types that the finding provider assigned to the finding. Uses the format of namespace/category/classifier that classify a finding.

Valid namespace values are: Software and Configuration Checks | TTPs | Effects | Unusual Behaviors | Sensitive Data Identifications

pub fn set_finding_provider_fields_types(
self,
input: Option<Vec<StringFilter>>
) -> Self

One or more finding types that the finding provider assigned to the finding. Uses the format of namespace/category/classifier that classify a finding.

Valid namespace values are: Software and Configuration Checks | TTPs | Effects | Unusual Behaviors | Sensitive Data Identifications

pub fn build(self) -> AwsSecurityFindingFilters

Consumes the builder and constructs a AwsSecurityFindingFilters