Struct aws_sdk_securityhub::model::resource_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ResourceDetails

Implementations

impl Builder

pub fn aws_auto_scaling_auto_scaling_group(
self,
input: AwsAutoScalingAutoScalingGroupDetails
) -> Self

Details for an autoscaling group.

pub fn set_aws_auto_scaling_auto_scaling_group(
self,
input: Option<AwsAutoScalingAutoScalingGroupDetails>
) -> Self

pub fn aws_code_build_project(self, input: AwsCodeBuildProjectDetails) -> Self

Details for an CodeBuild project.

pub fn set_aws_code_build_project(
self,
input: Option<AwsCodeBuildProjectDetails>
) -> Self

pub fn aws_cloud_front_distribution(
self,
input: AwsCloudFrontDistributionDetails
) -> Self

Details about a CloudFront distribution.

pub fn set_aws_cloud_front_distribution(
self,
input: Option<AwsCloudFrontDistributionDetails>
) -> Self

pub fn aws_ec2_instance(self, input: AwsEc2InstanceDetails) -> Self

Details about an EC2 instance related to a finding.

pub fn set_aws_ec2_instance(self, input: Option<AwsEc2InstanceDetails>) -> Self

pub fn aws_ec2_network_interface(
self,
input: AwsEc2NetworkInterfaceDetails
) -> Self

Details for an EC2 network interface.

pub fn set_aws_ec2_network_interface(
self,
input: Option<AwsEc2NetworkInterfaceDetails>
) -> Self

pub fn aws_ec2_security_group(self, input: AwsEc2SecurityGroupDetails) -> Self

Details for an EC2 security group.

pub fn set_aws_ec2_security_group(
self,
input: Option<AwsEc2SecurityGroupDetails>
) -> Self

pub fn aws_ec2_volume(self, input: AwsEc2VolumeDetails) -> Self

Details for an EC2 volume.

pub fn set_aws_ec2_volume(self, input: Option<AwsEc2VolumeDetails>) -> Self

pub fn aws_ec2_vpc(self, input: AwsEc2VpcDetails) -> Self

Details for an EC2 VPC.

pub fn set_aws_ec2_vpc(self, input: Option<AwsEc2VpcDetails>) -> Self

pub fn aws_ec2_eip(self, input: AwsEc2EipDetails) -> Self

Details about an Elastic IP address.

pub fn set_aws_ec2_eip(self, input: Option<AwsEc2EipDetails>) -> Self

pub fn aws_ec2_subnet(self, input: AwsEc2SubnetDetails) -> Self

Details about a subnet in Amazon EC2.

pub fn set_aws_ec2_subnet(self, input: Option<AwsEc2SubnetDetails>) -> Self

pub fn aws_ec2_network_acl(self, input: AwsEc2NetworkAclDetails) -> Self

Details about an EC2 network access control list (ACL).

pub fn set_aws_ec2_network_acl(
self,
input: Option<AwsEc2NetworkAclDetails>
) -> Self

pub fn aws_elbv2_load_balancer(self, input: AwsElbv2LoadBalancerDetails) -> Self

Details about a load balancer.

pub fn set_aws_elbv2_load_balancer(
self,
input: Option<AwsElbv2LoadBalancerDetails>
) -> Self

pub fn aws_elastic_beanstalk_environment(
self,
input: AwsElasticBeanstalkEnvironmentDetails
) -> Self

Details about an Elastic Beanstalk environment.

pub fn set_aws_elastic_beanstalk_environment(
self,
input: Option<AwsElasticBeanstalkEnvironmentDetails>
) -> Self

pub fn aws_elasticsearch_domain(
self,
input: AwsElasticsearchDomainDetails
) -> Self

Details for an Elasticsearch domain.

pub fn set_aws_elasticsearch_domain(
self,
input: Option<AwsElasticsearchDomainDetails>
) -> Self

pub fn aws_s3_bucket(self, input: AwsS3BucketDetails) -> Self

Details about an S3 bucket related to a finding.

pub fn set_aws_s3_bucket(self, input: Option<AwsS3BucketDetails>) -> Self

pub fn aws_s3_account_public_access_block(
self,
input: AwsS3AccountPublicAccessBlockDetails
) -> Self

Details about the Amazon S3 Public Access Block configuration for an account.

pub fn set_aws_s3_account_public_access_block(
self,
input: Option<AwsS3AccountPublicAccessBlockDetails>
) -> Self

pub fn aws_s3_object(self, input: AwsS3ObjectDetails) -> Self

Details about an S3 object related to a finding.

pub fn set_aws_s3_object(self, input: Option<AwsS3ObjectDetails>) -> Self

pub fn aws_secrets_manager_secret(
self,
input: AwsSecretsManagerSecretDetails
) -> Self

Details about a Secrets Manager secret.

pub fn set_aws_secrets_manager_secret(
self,
input: Option<AwsSecretsManagerSecretDetails>
) -> Self

pub fn aws_iam_access_key(self, input: AwsIamAccessKeyDetails) -> Self

Details about an IAM access key related to a finding.

pub fn set_aws_iam_access_key(
self,
input: Option<AwsIamAccessKeyDetails>
) -> Self

pub fn aws_iam_user(self, input: AwsIamUserDetails) -> Self

Details about an IAM user.

pub fn set_aws_iam_user(self, input: Option<AwsIamUserDetails>) -> Self

pub fn aws_iam_policy(self, input: AwsIamPolicyDetails) -> Self

Details about an IAM permissions policy.

pub fn set_aws_iam_policy(self, input: Option<AwsIamPolicyDetails>) -> Self

pub fn aws_api_gateway_v2_stage(
self,
input: AwsApiGatewayV2StageDetails
) -> Self

Provides information about a version 2 stage for Amazon API Gateway.

pub fn set_aws_api_gateway_v2_stage(
self,
input: Option<AwsApiGatewayV2StageDetails>
) -> Self

pub fn aws_api_gateway_v2_api(self, input: AwsApiGatewayV2ApiDetails) -> Self

Provides information about a version 2 API in Amazon API Gateway.

pub fn set_aws_api_gateway_v2_api(
self,
input: Option<AwsApiGatewayV2ApiDetails>
) -> Self

pub fn aws_dynamo_db_table(self, input: AwsDynamoDbTableDetails) -> Self

Details about a DynamoDB table.

pub fn set_aws_dynamo_db_table(
self,
input: Option<AwsDynamoDbTableDetails>
) -> Self

pub fn aws_api_gateway_stage(self, input: AwsApiGatewayStageDetails) -> Self

Provides information about a version 1 Amazon API Gateway stage.

pub fn set_aws_api_gateway_stage(
self,
input: Option<AwsApiGatewayStageDetails>
) -> Self

pub fn aws_api_gateway_rest_api(
self,
input: AwsApiGatewayRestApiDetails
) -> Self

Provides information about a REST API in version 1 of Amazon API Gateway.

pub fn set_aws_api_gateway_rest_api(
self,
input: Option<AwsApiGatewayRestApiDetails>
) -> Self

pub fn aws_cloud_trail_trail(self, input: AwsCloudTrailTrailDetails) -> Self

Provides details about a CloudTrail trail.

pub fn set_aws_cloud_trail_trail(
self,
input: Option<AwsCloudTrailTrailDetails>
) -> Self

pub fn aws_ssm_patch_compliance(
self,
input: AwsSsmPatchComplianceDetails
) -> Self

Provides information about the state of a patch on an instance based on the patch baseline that was used to patch the instance.

pub fn set_aws_ssm_patch_compliance(
self,
input: Option<AwsSsmPatchComplianceDetails>
) -> Self

pub fn aws_certificate_manager_certificate(
self,
input: AwsCertificateManagerCertificateDetails
) -> Self

Provides details about an Certificate Manager certificate.

pub fn set_aws_certificate_manager_certificate(
self,
input: Option<AwsCertificateManagerCertificateDetails>
) -> Self

pub fn aws_redshift_cluster(self, input: AwsRedshiftClusterDetails) -> Self

Contains details about an Amazon Redshift cluster.

pub fn set_aws_redshift_cluster(
self,
input: Option<AwsRedshiftClusterDetails>
) -> Self

pub fn aws_elb_load_balancer(self, input: AwsElbLoadBalancerDetails) -> Self

Contains details about a Classic Load Balancer.

pub fn set_aws_elb_load_balancer(
self,
input: Option<AwsElbLoadBalancerDetails>
) -> Self

pub fn aws_iam_group(self, input: AwsIamGroupDetails) -> Self

Contains details about an IAM group.

pub fn set_aws_iam_group(self, input: Option<AwsIamGroupDetails>) -> Self

pub fn aws_iam_role(self, input: AwsIamRoleDetails) -> Self

Details about an IAM role.

pub fn set_aws_iam_role(self, input: Option<AwsIamRoleDetails>) -> Self

pub fn aws_kms_key(self, input: AwsKmsKeyDetails) -> Self

Details about an KMS key.

pub fn set_aws_kms_key(self, input: Option<AwsKmsKeyDetails>) -> Self

pub fn aws_lambda_function(self, input: AwsLambdaFunctionDetails) -> Self

Details about a Lambda function.

pub fn set_aws_lambda_function(
self,
input: Option<AwsLambdaFunctionDetails>
) -> Self

pub fn aws_lambda_layer_version(
self,
input: AwsLambdaLayerVersionDetails
) -> Self

Details for a Lambda layer version.

pub fn set_aws_lambda_layer_version(
self,
input: Option<AwsLambdaLayerVersionDetails>
) -> Self

pub fn aws_rds_db_instance(self, input: AwsRdsDbInstanceDetails) -> Self

Details about an Amazon RDS database instance.

pub fn set_aws_rds_db_instance(
self,
input: Option<AwsRdsDbInstanceDetails>
) -> Self

pub fn aws_sns_topic(self, input: AwsSnsTopicDetails) -> Self

Details about an SNS topic.

pub fn set_aws_sns_topic(self, input: Option<AwsSnsTopicDetails>) -> Self

pub fn aws_sqs_queue(self, input: AwsSqsQueueDetails) -> Self

Details about an SQS queue.

pub fn set_aws_sqs_queue(self, input: Option<AwsSqsQueueDetails>) -> Self

pub fn aws_waf_web_acl(self, input: AwsWafWebAclDetails) -> Self

Details for an WAF WebACL.

pub fn set_aws_waf_web_acl(self, input: Option<AwsWafWebAclDetails>) -> Self

pub fn aws_rds_db_snapshot(self, input: AwsRdsDbSnapshotDetails) -> Self

Details about an Amazon RDS database snapshot.

pub fn set_aws_rds_db_snapshot(
self,
input: Option<AwsRdsDbSnapshotDetails>
) -> Self

pub fn aws_rds_db_cluster_snapshot(
self,
input: AwsRdsDbClusterSnapshotDetails
) -> Self

Details about an Amazon RDS database cluster snapshot.

pub fn set_aws_rds_db_cluster_snapshot(
self,
input: Option<AwsRdsDbClusterSnapshotDetails>
) -> Self

pub fn aws_rds_db_cluster(self, input: AwsRdsDbClusterDetails) -> Self

Details about an Amazon RDS database cluster.

pub fn set_aws_rds_db_cluster(
self,
input: Option<AwsRdsDbClusterDetails>
) -> Self

pub fn aws_ecs_cluster(self, input: AwsEcsClusterDetails) -> Self

Details about an ECS cluster.

pub fn set_aws_ecs_cluster(self, input: Option<AwsEcsClusterDetails>) -> Self

pub fn aws_ecs_task_definition(self, input: AwsEcsTaskDefinitionDetails) -> Self

Details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

pub fn set_aws_ecs_task_definition(
self,
input: Option<AwsEcsTaskDefinitionDetails>
) -> Self

pub fn container(self, input: ContainerDetails) -> Self

Details about a container resource related to a finding.

pub fn set_container(self, input: Option<ContainerDetails>) -> Self

pub fn other(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_other(self, input: Option<HashMap<String, String>>) -> Self

pub fn aws_rds_event_subscription(
self,
input: AwsRdsEventSubscriptionDetails
) -> Self

Details about an RDS event notification subscription.

pub fn set_aws_rds_event_subscription(
self,
input: Option<AwsRdsEventSubscriptionDetails>
) -> Self

pub fn aws_ecs_service(self, input: AwsEcsServiceDetails) -> Self

Details about a service within an ECS cluster.

pub fn set_aws_ecs_service(self, input: Option<AwsEcsServiceDetails>) -> Self

pub fn aws_auto_scaling_launch_configuration(
self,
input: AwsAutoScalingLaunchConfigurationDetails
) -> Self

Provides details about a launch configuration.

pub fn set_aws_auto_scaling_launch_configuration(
self,
input: Option<AwsAutoScalingLaunchConfigurationDetails>
) -> Self

pub fn aws_ec2_vpn_connection(self, input: AwsEc2VpnConnectionDetails) -> Self

Details about an EC2 VPN connection.

pub fn set_aws_ec2_vpn_connection(
self,
input: Option<AwsEc2VpnConnectionDetails>
) -> Self

pub fn aws_ecr_container_image(self, input: AwsEcrContainerImageDetails) -> Self

information about an Amazon ECR image.

pub fn set_aws_ecr_container_image(
self,
input: Option<AwsEcrContainerImageDetails>
) -> Self

pub fn build(self) -> ResourceDetails

Consumes the builder and constructs a ResourceDetails