Struct aws_sdk_securityhub::model::AwsEcsTaskDefinitionContainerDefinitionsDetails

#[non_exhaustive]
pub struct AwsEcsTaskDefinitionContainerDefinitionsDetails {

    pub command: Option<Vec<String>>,
    pub cpu: i32,
    pub depends_on: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails>>,
    pub disable_networking: bool,
    pub dns_search_domains: Option<Vec<String>>,
    pub dns_servers: Option<Vec<String>>,
    pub docker_labels: Option<HashMap<String, String>>,
    pub docker_security_options: Option<Vec<String>>,
    pub entry_point: Option<Vec<String>>,
    pub environment: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails>>,
    pub environment_files: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails>>,
    pub essential: bool,
    pub extra_hosts: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails>>,
    pub firelens_configuration: Option<AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails>,
    pub health_check: Option<AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails>,
    pub hostname: Option<String>,
    pub image: Option<String>,
    pub interactive: bool,
    pub links: Option<Vec<String>>,
    pub linux_parameters: Option<AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails>,
    pub log_configuration: Option<AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails>,
    pub memory: i32,
    pub memory_reservation: i32,
    pub mount_points: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails>>,
    pub name: Option<String>,
    pub port_mappings: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails>>,
    pub privileged: bool,
    pub pseudo_terminal: bool,
    pub readonly_root_filesystem: bool,
    pub repository_credentials: Option<AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails>,
    pub resource_requirements: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails>>,
    pub secrets: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails>>,
    pub start_timeout: i32,
    pub stop_timeout: i32,
    pub system_controls: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails>>,
    pub ulimits: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails>>,
    pub user: Option<String>,
    pub volumes_from: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails>>,
    pub working_directory: Option<String>,

}

A container definition that describes a container in the task.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

command: Option<Vec<String>>

The command that is passed to the container.

cpu: i32

The number of CPU units reserved for the container.

depends_on: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails>>

The dependencies that are defined for container startup and shutdown.

disable_networking: bool

Whether to disable networking within the container.

dns_search_domains: Option<Vec<String>>

A list of DNS search domains that are presented to the container.

dns_servers: Option<Vec<String>>

A list of DNS servers that are presented to the container.

docker_labels: Option<HashMap<String, String>>

A key-value map of labels to add to the container.

docker_security_options: Option<Vec<String>>

A list of strings to provide custom labels for SELinux and AppArmor multi-level security systems.

entry_point: Option<Vec<String>>

The entry point that is passed to the container.

environment: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails>>

The environment variables to pass to a container.

environment_files: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails>>

A list of files containing the environment variables to pass to a container.

essential: bool

Whether the container is essential. All tasks must have at least one essential container.

extra_hosts: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails>>

A list of hostnames and IP address mappings to append to the /etc/hosts file on the container.

firelens_configuration: Option<AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails>

The FireLens configuration for the container. Specifies and configures a log router for container logs.

health_check: Option<AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails>

The container health check command and associated configuration parameters for the container.

hostname: Option<String>

The hostname to use for the container.

image: Option<String>

The image used to start the container.

interactive: bool

If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

links: Option<Vec<String>>

A list of links for the container in the form container_name:alias . Allows containers to communicate with each other without the need for port mappings.

linux_parameters: Option<AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails>

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

log_configuration: Option<AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails>

The log configuration specification for the container.

memory: i32

The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

memory_reservation: i32

The soft limit (in MiB) of memory to reserve for the container.

mount_points: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails>>

The mount points for the data volumes in the container.

name: Option<String>

The name of the container.

port_mappings: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails>>

The list of port mappings for the container.

privileged: bool

Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

pseudo_terminal: bool

Whether to allocate a TTY to the container.

readonly_root_filesystem: bool

Whether the container is given read-only access to its root file system.

repository_credentials: Option<AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails>

The private repository authentication credentials to use.

resource_requirements: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails>>

The type and amount of a resource to assign to a container. The only supported resource is a GPU.

secrets: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails>>

The secrets to pass to the container.

start_timeout: i32

The number of seconds to wait before giving up on resolving dependencies for a container.

stop_timeout: i32

The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

system_controls: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails>>

A list of namespaced kernel parameters to set in the container.

ulimits: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails>>

A list of ulimits to set in the container.

user: Option<String>

The user to use inside the container.

The value can use one of the following formats.

user

user : group

uid

uid : gid

user : gid

uid : group

volumes_from: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails>>

Data volumes to mount from another container.

working_directory: Option<String>

The working directory in which to run commands inside the container.

Implementations

impl AwsEcsTaskDefinitionContainerDefinitionsDetails

pub fn command(&self) -> Option<&[String]>

The command that is passed to the container.

pub fn cpu(&self) -> i32

The number of CPU units reserved for the container.

pub fn depends_on(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails]>

The dependencies that are defined for container startup and shutdown.

pub fn disable_networking(&self) -> bool

Whether to disable networking within the container.

pub fn dns_search_domains(&self) -> Option<&[String]>

A list of DNS search domains that are presented to the container.

pub fn dns_servers(&self) -> Option<&[String]>

A list of DNS servers that are presented to the container.

pub fn docker_labels(&self) -> Option<&HashMap<String, String>>

A key-value map of labels to add to the container.

pub fn docker_security_options(&self) -> Option<&[String]>

A list of strings to provide custom labels for SELinux and AppArmor multi-level security systems.

pub fn entry_point(&self) -> Option<&[String]>

The entry point that is passed to the container.

pub fn environment(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails]>

The environment variables to pass to a container.

pub fn environment_files(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails]>

A list of files containing the environment variables to pass to a container.

pub fn essential(&self) -> bool

Whether the container is essential. All tasks must have at least one essential container.

pub fn extra_hosts(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails]>

A list of hostnames and IP address mappings to append to the /etc/hosts file on the container.

pub fn firelens_configuration(
&self
) -> Option<&AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails>

The FireLens configuration for the container. Specifies and configures a log router for container logs.

pub fn health_check(
&self
) -> Option<&AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails>

The container health check command and associated configuration parameters for the container.

pub fn hostname(&self) -> Option<&str>

The hostname to use for the container.

pub fn image(&self) -> Option<&str>

The image used to start the container.

pub fn interactive(&self) -> bool

If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

pub fn links(&self) -> Option<&[String]>

A list of links for the container in the form container_name:alias . Allows containers to communicate with each other without the need for port mappings.

pub fn linux_parameters(
&self
) -> Option<&AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails>

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

pub fn log_configuration(
&self
) -> Option<&AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails>

The log configuration specification for the container.

pub fn memory(&self) -> i32

The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

pub fn memory_reservation(&self) -> i32

The soft limit (in MiB) of memory to reserve for the container.

pub fn mount_points(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails]>

The mount points for the data volumes in the container.

pub fn name(&self) -> Option<&str>

The name of the container.

pub fn port_mappings(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails]>

The list of port mappings for the container.

pub fn privileged(&self) -> bool

Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

pub fn pseudo_terminal(&self) -> bool

Whether to allocate a TTY to the container.

pub fn readonly_root_filesystem(&self) -> bool

Whether the container is given read-only access to its root file system.

pub fn repository_credentials(
&self
) -> Option<&AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails>

The private repository authentication credentials to use.

pub fn resource_requirements(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails]>

The type and amount of a resource to assign to a container. The only supported resource is a GPU.

pub fn secrets(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails]>

The secrets to pass to the container.

pub fn start_timeout(&self) -> i32

The number of seconds to wait before giving up on resolving dependencies for a container.

pub fn stop_timeout(&self) -> i32

The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

pub fn system_controls(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails]>

A list of namespaced kernel parameters to set in the container.

pub fn ulimits(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails]>

A list of ulimits to set in the container.

pub fn user(&self) -> Option<&str>

The user to use inside the container.

The value can use one of the following formats.

user

user : group

uid

uid : gid

user : gid

uid : group

pub fn volumes_from(
&self
) -> Option<&[AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails]>

Data volumes to mount from another container.

pub fn working_directory(&self) -> Option<&str>

The working directory in which to run commands inside the container.

impl AwsEcsTaskDefinitionContainerDefinitionsDetails

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AwsEcsTaskDefinitionContainerDefinitionsDetails