Struct aws_sdk_securityhub::model::AwsRdsDbInstanceDetails

#[non_exhaustive]
pub struct AwsRdsDbInstanceDetails {

    pub associated_roles: Option<Vec<AwsRdsDbInstanceAssociatedRole>>,
    pub ca_certificate_identifier: Option<String>,
    pub db_cluster_identifier: Option<String>,
    pub db_instance_identifier: Option<String>,
    pub db_instance_class: Option<String>,
    pub db_instance_port: i32,
    pub dbi_resource_id: Option<String>,
    pub db_name: Option<String>,
    pub deletion_protection: bool,
    pub endpoint: Option<AwsRdsDbInstanceEndpoint>,
    pub engine: Option<String>,
    pub engine_version: Option<String>,
    pub iam_database_authentication_enabled: bool,
    pub instance_create_time: Option<String>,
    pub kms_key_id: Option<String>,
    pub publicly_accessible: bool,
    pub storage_encrypted: bool,
    pub tde_credential_arn: Option<String>,
    pub vpc_security_groups: Option<Vec<AwsRdsDbInstanceVpcSecurityGroup>>,
    pub multi_az: bool,
    pub enhanced_monitoring_resource_arn: Option<String>,
    pub db_instance_status: Option<String>,
    pub master_username: Option<String>,
    pub allocated_storage: i32,
    pub preferred_backup_window: Option<String>,
    pub backup_retention_period: i32,
    pub db_security_groups: Option<Vec<String>>,
    pub db_parameter_groups: Option<Vec<AwsRdsDbParameterGroup>>,
    pub availability_zone: Option<String>,
    pub db_subnet_group: Option<AwsRdsDbSubnetGroup>,
    pub preferred_maintenance_window: Option<String>,
    pub pending_modified_values: Option<AwsRdsDbPendingModifiedValues>,
    pub latest_restorable_time: Option<String>,
    pub auto_minor_version_upgrade: bool,
    pub read_replica_source_db_instance_identifier: Option<String>,
    pub read_replica_db_instance_identifiers: Option<Vec<String>>,
    pub read_replica_db_cluster_identifiers: Option<Vec<String>>,
    pub license_model: Option<String>,
    pub iops: i32,
    pub option_group_memberships: Option<Vec<AwsRdsDbOptionGroupMembership>>,
    pub character_set_name: Option<String>,
    pub secondary_availability_zone: Option<String>,
    pub status_infos: Option<Vec<AwsRdsDbStatusInfo>>,
    pub storage_type: Option<String>,
    pub domain_memberships: Option<Vec<AwsRdsDbDomainMembership>>,
    pub copy_tags_to_snapshot: bool,
    pub monitoring_interval: i32,
    pub monitoring_role_arn: Option<String>,
    pub promotion_tier: i32,
    pub timezone: Option<String>,
    pub performance_insights_enabled: bool,
    pub performance_insights_kms_key_id: Option<String>,
    pub performance_insights_retention_period: i32,
    pub enabled_cloud_watch_logs_exports: Option<Vec<String>>,
    pub processor_features: Option<Vec<AwsRdsDbProcessorFeature>>,
    pub listener_endpoint: Option<AwsRdsDbInstanceEndpoint>,
    pub max_allocated_storage: i32,

}

Contains the details of an Amazon RDS DB instance.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

associated_roles: Option<Vec<AwsRdsDbInstanceAssociatedRole>>

The IAM roles associated with the DB instance.

ca_certificate_identifier: Option<String>

The identifier of the CA certificate for this DB instance.

db_cluster_identifier: Option<String>

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

db_instance_identifier: Option<String>

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

db_instance_class: Option<String>

Contains the name of the compute and memory capacity class of the DB instance.

db_instance_port: i32

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

dbi_resource_id: Option<String>

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

db_name: Option<String>

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

deletion_protection: bool

Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

endpoint: Option<AwsRdsDbInstanceEndpoint>

Specifies the connection endpoint.

engine: Option<String>

Provides the name of the database engine to use for this DB instance.

engine_version: Option<String>

Indicates the database engine version.

iam_database_authentication_enabled: bool

True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

instance_create_time: Option<String>

Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

kms_key_id: Option<String>

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

publicly_accessible: bool

Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

storage_encrypted: bool

Specifies whether the DB instance is encrypted.

tde_credential_arn: Option<String>

The ARN from the key store with which the instance is associated for TDE encryption.

vpc_security_groups: Option<Vec<AwsRdsDbInstanceVpcSecurityGroup>>

A list of VPC security groups that the DB instance belongs to.

multi_az: bool

Whether the DB instance is a multiple Availability Zone deployment.

enhanced_monitoring_resource_arn: Option<String>

The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

db_instance_status: Option<String>

The current status of the DB instance.

master_username: Option<String>

The master user name of the DB instance.

allocated_storage: i32

The amount of storage (in gigabytes) to initially allocate for the DB instance.

preferred_backup_window: Option<String>

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

backup_retention_period: i32

The number of days for which to retain automated backups.

db_security_groups: Option<Vec<String>>

A list of the DB security groups to assign to the DB instance.

db_parameter_groups: Option<Vec<AwsRdsDbParameterGroup>>

A list of the DB parameter groups to assign to the DB instance.

availability_zone: Option<String>

The Availability Zone where the DB instance will be created.

db_subnet_group: Option<AwsRdsDbSubnetGroup>

Information about the subnet group that is associated with the DB instance.

preferred_maintenance_window: Option<String>

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format :HH:MM-:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

pending_modified_values: Option<AwsRdsDbPendingModifiedValues>

Changes to the DB instance that are currently pending.

latest_restorable_time: Option<String>

Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

auto_minor_version_upgrade: bool

Indicates whether minor version patches are applied automatically.

read_replica_source_db_instance_identifier: Option<String>

If this DB instance is a read replica, contains the identifier of the source DB instance.

read_replica_db_instance_identifiers: Option<Vec<String>>

List of identifiers of the read replicas associated with this DB instance.

read_replica_db_cluster_identifiers: Option<Vec<String>>

List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

license_model: Option<String>

License model information for this DB instance.

iops: i32

Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

option_group_memberships: Option<Vec<AwsRdsDbOptionGroupMembership>>

The list of option group memberships for this DB instance.

character_set_name: Option<String>

The name of the character set that this DB instance is associated with.

secondary_availability_zone: Option<String>

For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

status_infos: Option<Vec<AwsRdsDbStatusInfo>>

The status of a read replica. If the instance isn't a read replica, this is empty.

storage_type: Option<String>

The storage type for the DB instance.

domain_memberships: Option<Vec<AwsRdsDbDomainMembership>>

The Active Directory domain membership records associated with the DB instance.

copy_tags_to_snapshot: bool

Whether to copy resource tags to snapshots of the DB instance.

monitoring_interval: i32

The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

monitoring_role_arn: Option<String>

The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

promotion_tier: i32

The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

timezone: Option<String>

The time zone of the DB instance.

performance_insights_enabled: bool

Indicates whether Performance Insights is enabled for the DB instance.

performance_insights_kms_key_id: Option<String>

The identifier of the KMS key used to encrypt the Performance Insights data.

performance_insights_retention_period: i32

The number of days to retain Performance Insights data.

enabled_cloud_watch_logs_exports: Option<Vec<String>>

A list of log types that this DB instance is configured to export to CloudWatch Logs.

processor_features: Option<Vec<AwsRdsDbProcessorFeature>>

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

listener_endpoint: Option<AwsRdsDbInstanceEndpoint>

Specifies the connection endpoint.

max_allocated_storage: i32

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

Implementations

impl AwsRdsDbInstanceDetails

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AwsRdsDbInstanceDetails