Struct aws_sdk_securityhub::model::ResourceDetails

#[non_exhaustive]
pub struct ResourceDetails {

    pub aws_auto_scaling_auto_scaling_group: Option<AwsAutoScalingAutoScalingGroupDetails>,
    pub aws_code_build_project: Option<AwsCodeBuildProjectDetails>,
    pub aws_cloud_front_distribution: Option<AwsCloudFrontDistributionDetails>,
    pub aws_ec2_instance: Option<AwsEc2InstanceDetails>,
    pub aws_ec2_network_interface: Option<AwsEc2NetworkInterfaceDetails>,
    pub aws_ec2_security_group: Option<AwsEc2SecurityGroupDetails>,
    pub aws_ec2_volume: Option<AwsEc2VolumeDetails>,
    pub aws_ec2_vpc: Option<AwsEc2VpcDetails>,
    pub aws_ec2_eip: Option<AwsEc2EipDetails>,
    pub aws_ec2_subnet: Option<AwsEc2SubnetDetails>,
    pub aws_ec2_network_acl: Option<AwsEc2NetworkAclDetails>,
    pub aws_elbv2_load_balancer: Option<AwsElbv2LoadBalancerDetails>,
    pub aws_elastic_beanstalk_environment: Option<AwsElasticBeanstalkEnvironmentDetails>,
    pub aws_elasticsearch_domain: Option<AwsElasticsearchDomainDetails>,
    pub aws_s3_bucket: Option<AwsS3BucketDetails>,
    pub aws_s3_account_public_access_block: Option<AwsS3AccountPublicAccessBlockDetails>,
    pub aws_s3_object: Option<AwsS3ObjectDetails>,
    pub aws_secrets_manager_secret: Option<AwsSecretsManagerSecretDetails>,
    pub aws_iam_access_key: Option<AwsIamAccessKeyDetails>,
    pub aws_iam_user: Option<AwsIamUserDetails>,
    pub aws_iam_policy: Option<AwsIamPolicyDetails>,
    pub aws_api_gateway_v2_stage: Option<AwsApiGatewayV2StageDetails>,
    pub aws_api_gateway_v2_api: Option<AwsApiGatewayV2ApiDetails>,
    pub aws_dynamo_db_table: Option<AwsDynamoDbTableDetails>,
    pub aws_api_gateway_stage: Option<AwsApiGatewayStageDetails>,
    pub aws_api_gateway_rest_api: Option<AwsApiGatewayRestApiDetails>,
    pub aws_cloud_trail_trail: Option<AwsCloudTrailTrailDetails>,
    pub aws_ssm_patch_compliance: Option<AwsSsmPatchComplianceDetails>,
    pub aws_certificate_manager_certificate: Option<AwsCertificateManagerCertificateDetails>,
    pub aws_redshift_cluster: Option<AwsRedshiftClusterDetails>,
    pub aws_elb_load_balancer: Option<AwsElbLoadBalancerDetails>,
    pub aws_iam_group: Option<AwsIamGroupDetails>,
    pub aws_iam_role: Option<AwsIamRoleDetails>,
    pub aws_kms_key: Option<AwsKmsKeyDetails>,
    pub aws_lambda_function: Option<AwsLambdaFunctionDetails>,
    pub aws_lambda_layer_version: Option<AwsLambdaLayerVersionDetails>,
    pub aws_rds_db_instance: Option<AwsRdsDbInstanceDetails>,
    pub aws_sns_topic: Option<AwsSnsTopicDetails>,
    pub aws_sqs_queue: Option<AwsSqsQueueDetails>,
    pub aws_waf_web_acl: Option<AwsWafWebAclDetails>,
    pub aws_rds_db_snapshot: Option<AwsRdsDbSnapshotDetails>,
    pub aws_rds_db_cluster_snapshot: Option<AwsRdsDbClusterSnapshotDetails>,
    pub aws_rds_db_cluster: Option<AwsRdsDbClusterDetails>,
    pub aws_ecs_cluster: Option<AwsEcsClusterDetails>,
    pub aws_ecs_task_definition: Option<AwsEcsTaskDefinitionDetails>,
    pub container: Option<ContainerDetails>,
    pub other: Option<HashMap<String, String>>,
    pub aws_rds_event_subscription: Option<AwsRdsEventSubscriptionDetails>,
    pub aws_ecs_service: Option<AwsEcsServiceDetails>,
    pub aws_auto_scaling_launch_configuration: Option<AwsAutoScalingLaunchConfigurationDetails>,
    pub aws_ec2_vpn_connection: Option<AwsEc2VpnConnectionDetails>,
    pub aws_ecr_container_image: Option<AwsEcrContainerImageDetails>,

}

Additional details about a resource related to a finding.

To provide the details, use the object that corresponds to the resource type. For example, if the resource type is AwsEc2Instance, then you use the AwsEc2Instance object to provide the details.

If the type-specific object does not contain all of the fields you want to populate, then you use the Other object to populate those additional fields.

You also use the Other object to populate the details when the selected type does not have a corresponding object.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

aws_auto_scaling_auto_scaling_group: Option<AwsAutoScalingAutoScalingGroupDetails>

Details for an autoscaling group.

aws_code_build_project: Option<AwsCodeBuildProjectDetails>

Details for an CodeBuild project.

aws_cloud_front_distribution: Option<AwsCloudFrontDistributionDetails>

Details about a CloudFront distribution.

aws_ec2_instance: Option<AwsEc2InstanceDetails>

Details about an EC2 instance related to a finding.

aws_ec2_network_interface: Option<AwsEc2NetworkInterfaceDetails>

Details for an EC2 network interface.

aws_ec2_security_group: Option<AwsEc2SecurityGroupDetails>

Details for an EC2 security group.

aws_ec2_volume: Option<AwsEc2VolumeDetails>

Details for an EC2 volume.

aws_ec2_vpc: Option<AwsEc2VpcDetails>

Details for an EC2 VPC.

aws_ec2_eip: Option<AwsEc2EipDetails>

Details about an Elastic IP address.

aws_ec2_subnet: Option<AwsEc2SubnetDetails>

Details about a subnet in Amazon EC2.

aws_ec2_network_acl: Option<AwsEc2NetworkAclDetails>

Details about an EC2 network access control list (ACL).

aws_elbv2_load_balancer: Option<AwsElbv2LoadBalancerDetails>

Details about a load balancer.

aws_elastic_beanstalk_environment: Option<AwsElasticBeanstalkEnvironmentDetails>

Details about an Elastic Beanstalk environment.

aws_elasticsearch_domain: Option<AwsElasticsearchDomainDetails>

Details for an Elasticsearch domain.

aws_s3_bucket: Option<AwsS3BucketDetails>

Details about an S3 bucket related to a finding.

aws_s3_account_public_access_block: Option<AwsS3AccountPublicAccessBlockDetails>

Details about the Amazon S3 Public Access Block configuration for an account.

aws_s3_object: Option<AwsS3ObjectDetails>

Details about an S3 object related to a finding.

aws_secrets_manager_secret: Option<AwsSecretsManagerSecretDetails>

Details about a Secrets Manager secret.

aws_iam_access_key: Option<AwsIamAccessKeyDetails>

Details about an IAM access key related to a finding.

aws_iam_user: Option<AwsIamUserDetails>

Details about an IAM user.

aws_iam_policy: Option<AwsIamPolicyDetails>

Details about an IAM permissions policy.

aws_api_gateway_v2_stage: Option<AwsApiGatewayV2StageDetails>

Provides information about a version 2 stage for Amazon API Gateway.

aws_api_gateway_v2_api: Option<AwsApiGatewayV2ApiDetails>

Provides information about a version 2 API in Amazon API Gateway.

aws_dynamo_db_table: Option<AwsDynamoDbTableDetails>

Details about a DynamoDB table.

aws_api_gateway_stage: Option<AwsApiGatewayStageDetails>

Provides information about a version 1 Amazon API Gateway stage.

aws_api_gateway_rest_api: Option<AwsApiGatewayRestApiDetails>

Provides information about a REST API in version 1 of Amazon API Gateway.

aws_cloud_trail_trail: Option<AwsCloudTrailTrailDetails>

Provides details about a CloudTrail trail.

aws_ssm_patch_compliance: Option<AwsSsmPatchComplianceDetails>

Provides information about the state of a patch on an instance based on the patch baseline that was used to patch the instance.

aws_certificate_manager_certificate: Option<AwsCertificateManagerCertificateDetails>

Provides details about an Certificate Manager certificate.

aws_redshift_cluster: Option<AwsRedshiftClusterDetails>

Contains details about an Amazon Redshift cluster.

aws_elb_load_balancer: Option<AwsElbLoadBalancerDetails>

Contains details about a Classic Load Balancer.

aws_iam_group: Option<AwsIamGroupDetails>

Contains details about an IAM group.

aws_iam_role: Option<AwsIamRoleDetails>

Details about an IAM role.

aws_kms_key: Option<AwsKmsKeyDetails>

Details about an KMS key.

aws_lambda_function: Option<AwsLambdaFunctionDetails>

Details about a Lambda function.

aws_lambda_layer_version: Option<AwsLambdaLayerVersionDetails>

Details for a Lambda layer version.

aws_rds_db_instance: Option<AwsRdsDbInstanceDetails>

Details about an Amazon RDS database instance.

aws_sns_topic: Option<AwsSnsTopicDetails>

Details about an SNS topic.

aws_sqs_queue: Option<AwsSqsQueueDetails>

Details about an SQS queue.

aws_waf_web_acl: Option<AwsWafWebAclDetails>

Details for an WAF WebACL.

aws_rds_db_snapshot: Option<AwsRdsDbSnapshotDetails>

Details about an Amazon RDS database snapshot.

aws_rds_db_cluster_snapshot: Option<AwsRdsDbClusterSnapshotDetails>

Details about an Amazon RDS database cluster snapshot.

aws_rds_db_cluster: Option<AwsRdsDbClusterDetails>

Details about an Amazon RDS database cluster.

aws_ecs_cluster: Option<AwsEcsClusterDetails>

Details about an ECS cluster.

aws_ecs_task_definition: Option<AwsEcsTaskDefinitionDetails>

Details about a task definition. A task definition describes the container and volume definitions of an Amazon Elastic Container Service task.

container: Option<ContainerDetails>

Details about a container resource related to a finding.

other: Option<HashMap<String, String>>

Details about a resource that are not available in a type-specific details object. Use the Other object in the following cases.

The type-specific object does not contain all of the fields that you want to populate. In this case, first use the type-specific object to populate those fields. Use the Other object to populate the fields that are missing from the type-specific object.

The resource type does not have a corresponding object. This includes resources for which the type is Other.

aws_rds_event_subscription: Option<AwsRdsEventSubscriptionDetails>

Details about an RDS event notification subscription.

aws_ecs_service: Option<AwsEcsServiceDetails>

Details about a service within an ECS cluster.

aws_auto_scaling_launch_configuration: Option<AwsAutoScalingLaunchConfigurationDetails>

Provides details about a launch configuration.

aws_ec2_vpn_connection: Option<AwsEc2VpnConnectionDetails>

Details about an EC2 VPN connection.

aws_ecr_container_image: Option<AwsEcrContainerImageDetails>

information about an Amazon ECR image.

Implementations

impl ResourceDetails

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ResourceDetails