Module aws_sdk_securityhub::operation

Structs

AcceptAdministratorInvitation

Accepts the invitation to be a member account and be monitored by the Security Hub administrator

AcceptInvitation

This method is deprecated. Instead, use AcceptAdministratorInvitation.

BatchDisableStandards

Disables the standards specified by the provided

BatchEnableStandards

Enables the standards specified by the provided StandardsArn. To obtain the

BatchImportFindings

Imports security findings generated from an integrated product into Security Hub.

BatchUpdateFindings

Used by Security Hub customers to update information about their investigation into a finding.

CreateActionTarget

Creates a custom action target in Security Hub.

CreateInsight

Creates a custom insight in Security Hub. An insight is a consolidation of findings that relate

CreateMembers

Creates a member association in Security Hub between the specified accounts and the account

DeclineInvitations

Declines invitations to become a member account.

DeleteActionTarget

Deletes a custom action target from Security Hub.

DeleteInsight

Deletes the insight specified by the InsightArn.

DeleteInvitations

Deletes invitations received by the Amazon Web Services account to become a member account.

DeleteMembers

Deletes the specified member accounts from Security Hub.

DescribeActionTargets

Returns a list of the custom action targets in Security Hub in your account.

DescribeHub

Returns details about the Hub resource in your account, including the

DescribeOrganizationConfiguration

Returns information about the Organizations configuration for Security Hub. Can only be

DescribeProducts

Returns information about product integrations in Security Hub.

DescribeStandards

Returns a list of the available standards in Security Hub.

DescribeStandardsControls

Returns a list of security standards controls.

DisableImportFindingsForProduct

Disables the integration of the specified product with Security Hub. After the integration is

DisableOrganizationAdminAccount

Disables a Security Hub administrator account. Can only be called by the organization

DisableSecurityHub

Disables Security Hub in your account only in the current Region. To disable Security Hub in all

DisassociateFromAdministratorAccount

Disassociates the current Security Hub member account from the associated administrator

DisassociateFromMasterAccount

This method is deprecated. Instead, use DisassociateFromAdministratorAccount.

DisassociateMembers

Disassociates the specified member accounts from the associated administrator account.

EnableImportFindingsForProduct

Enables the integration of a partner product with Security Hub. Integrated products send

EnableOrganizationAdminAccount

Designates the Security Hub administrator account for an organization. Can only be called by

EnableSecurityHub

Enables Security Hub for your account in the current Region or the Region you specify in the

GetAdministratorAccount

Provides the details for the Security Hub administrator account for the current member account.

GetEnabledStandards

Returns a list of the standards that are currently enabled.

GetFindings

Returns a list of findings that match the specified criteria.

GetInsightResults

Lists the results of the Security Hub insight specified by the insight ARN.

GetInsights

Lists and describes insights for the specified insight ARNs.

GetInvitationsCount

Returns the count of all Security Hub membership invitations that were sent to the

GetMasterAccount

This method is deprecated. Instead, use GetAdministratorAccount.

GetMembers

Returns the details for the Security Hub member accounts for the specified account IDs.

InviteMembers

Invites other Amazon Web Services accounts to become member accounts for the Security Hub administrator account that

ListEnabledProductsForImport

Lists all findings-generating solutions (products) that you are subscribed to receive

ListInvitations

Lists all Security Hub membership invitations that were sent to the current Amazon Web Services account.

ListMembers

Lists details about all member accounts for the current Security Hub administrator

ListOrganizationAdminAccounts

Lists the Security Hub administrator accounts. Can only be called by the organization

ListTagsForResource

Returns a list of tags associated with a resource.

TagResource

Adds one or more tags to a resource.

UntagResource

Removes one or more tags from a resource.

UpdateActionTarget

Updates the name and description of a custom action target in Security Hub.

UpdateFindings

UpdateInsight

Updates the Security Hub insight identified by the specified insight ARN.

UpdateOrganizationConfiguration

Used to update the configuration related to Organizations. Can only be called from a

UpdateSecurityHubConfiguration

Updates configuration options for Security Hub.

UpdateStandardsControl

Used to control whether an individual security standard control is enabled or