Struct aws_sdk_securityhub::model::aws_ecs_task_definition_container_definitions_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsEcsTaskDefinitionContainerDefinitionsDetails

Implementations

impl Builder

pub fn command(self, input: impl Into<String>) -> Self

pub fn set_command(self, input: Option<Vec<String>>) -> Self

pub fn cpu(self, input: i32) -> Self

The number of CPU units reserved for the container.

pub fn set_cpu(self, input: Option<i32>) -> Self

pub fn depends_on(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails>
) -> Self

pub fn set_depends_on(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsDependsOnDetails>>
) -> Self

pub fn disable_networking(self, input: bool) -> Self

Whether to disable networking within the container.

pub fn set_disable_networking(self, input: Option<bool>) -> Self

pub fn dns_search_domains(self, input: impl Into<String>) -> Self

pub fn set_dns_search_domains(self, input: Option<Vec<String>>) -> Self

pub fn dns_servers(self, input: impl Into<String>) -> Self

pub fn set_dns_servers(self, input: Option<Vec<String>>) -> Self

pub fn docker_labels(self, k: impl Into<String>, v: impl Into<String>) -> Self

pub fn set_docker_labels(self, input: Option<HashMap<String, String>>) -> Self

pub fn docker_security_options(self, input: impl Into<String>) -> Self

pub fn set_docker_security_options(self, input: Option<Vec<String>>) -> Self

pub fn entry_point(self, input: impl Into<String>) -> Self

pub fn set_entry_point(self, input: Option<Vec<String>>) -> Self

pub fn environment(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails>
) -> Self

pub fn set_environment(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentDetails>>
) -> Self

pub fn environment_files(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails>
) -> Self

pub fn set_environment_files(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsEnvironmentFilesDetails>>
) -> Self

pub fn essential(self, input: bool) -> Self

Whether the container is essential. All tasks must have at least one essential container.

pub fn set_essential(self, input: Option<bool>) -> Self

pub fn extra_hosts(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails>
) -> Self

pub fn set_extra_hosts(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsExtraHostsDetails>>
) -> Self

pub fn firelens_configuration(
self,
input: AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails
) -> Self

The FireLens configuration for the container. Specifies and configures a log router for container logs.

pub fn set_firelens_configuration(
self,
input: Option<AwsEcsTaskDefinitionContainerDefinitionsFirelensConfigurationDetails>
) -> Self

pub fn health_check(
self,
input: AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails
) -> Self

The container health check command and associated configuration parameters for the container.

pub fn set_health_check(
self,
input: Option<AwsEcsTaskDefinitionContainerDefinitionsHealthCheckDetails>
) -> Self

pub fn hostname(self, input: impl Into<String>) -> Self

The hostname to use for the container.

pub fn set_hostname(self, input: Option<String>) -> Self

pub fn image(self, input: impl Into<String>) -> Self

The image used to start the container.

pub fn set_image(self, input: Option<String>) -> Self

pub fn interactive(self, input: bool) -> Self

If set to true, then containerized applications can be deployed that require stdin or a tty to be allocated.

pub fn set_interactive(self, input: Option<bool>) -> Self

pub fn links(self, input: impl Into<String>) -> Self

pub fn set_links(self, input: Option<Vec<String>>) -> Self

pub fn linux_parameters(
self,
input: AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails
) -> Self

Linux-specific modifications that are applied to the container, such as Linux kernel capabilities.

pub fn set_linux_parameters(
self,
input: Option<AwsEcsTaskDefinitionContainerDefinitionsLinuxParametersDetails>
) -> Self

pub fn log_configuration(
self,
input: AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails
) -> Self

The log configuration specification for the container.

pub fn set_log_configuration(
self,
input: Option<AwsEcsTaskDefinitionContainerDefinitionsLogConfigurationDetails>
) -> Self

pub fn memory(self, input: i32) -> Self

The amount (in MiB) of memory to present to the container. If the container attempts to exceed the memory specified here, the container is shut down. The total amount of memory reserved for all containers within a task must be lower than the task memory value, if one is specified.

pub fn set_memory(self, input: Option<i32>) -> Self

pub fn memory_reservation(self, input: i32) -> Self

The soft limit (in MiB) of memory to reserve for the container.

pub fn set_memory_reservation(self, input: Option<i32>) -> Self

pub fn mount_points(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails>
) -> Self

pub fn set_mount_points(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsMountPointsDetails>>
) -> Self

pub fn name(self, input: impl Into<String>) -> Self

The name of the container.

pub fn set_name(self, input: Option<String>) -> Self

pub fn port_mappings(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails>
) -> Self

pub fn set_port_mappings(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsPortMappingsDetails>>
) -> Self

pub fn privileged(self, input: bool) -> Self

Whether the container is given elevated privileges on the host container instance. The elevated privileges are similar to the root user.

pub fn set_privileged(self, input: Option<bool>) -> Self

pub fn pseudo_terminal(self, input: bool) -> Self

Whether to allocate a TTY to the container.

pub fn set_pseudo_terminal(self, input: Option<bool>) -> Self

pub fn readonly_root_filesystem(self, input: bool) -> Self

Whether the container is given read-only access to its root file system.

pub fn set_readonly_root_filesystem(self, input: Option<bool>) -> Self

pub fn repository_credentials(
self,
input: AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails
) -> Self

The private repository authentication credentials to use.

pub fn set_repository_credentials(
self,
input: Option<AwsEcsTaskDefinitionContainerDefinitionsRepositoryCredentialsDetails>
) -> Self

pub fn resource_requirements(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails>
) -> Self

pub fn set_resource_requirements(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsResourceRequirementsDetails>>
) -> Self

pub fn secrets(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails>
) -> Self

pub fn set_secrets(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSecretsDetails>>
) -> Self

pub fn start_timeout(self, input: i32) -> Self

The number of seconds to wait before giving up on resolving dependencies for a container.

pub fn set_start_timeout(self, input: Option<i32>) -> Self

pub fn stop_timeout(self, input: i32) -> Self

The number of seconds to wait before the container is stopped if it doesn't shut down normally on its own.

pub fn set_stop_timeout(self, input: Option<i32>) -> Self

pub fn system_controls(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails>
) -> Self

pub fn set_system_controls(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsSystemControlsDetails>>
) -> Self

pub fn ulimits(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails>
) -> Self

pub fn set_ulimits(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsUlimitsDetails>>
) -> Self

pub fn user(self, input: impl Into<String>) -> Self

The user to use inside the container.

The value can use one of the following formats.

user

user : group

uid

uid : gid

user : gid

uid : group

pub fn set_user(self, input: Option<String>) -> Self

pub fn volumes_from(
self,
input: impl Into<AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails>
) -> Self

pub fn set_volumes_from(
self,
input: Option<Vec<AwsEcsTaskDefinitionContainerDefinitionsVolumesFromDetails>>
) -> Self

pub fn working_directory(self, input: impl Into<String>) -> Self

The working directory in which to run commands inside the container.

pub fn set_working_directory(self, input: Option<String>) -> Self

pub fn build(self) -> AwsEcsTaskDefinitionContainerDefinitionsDetails

Consumes the builder and constructs a AwsEcsTaskDefinitionContainerDefinitionsDetails