Struct aws_sdk_securityhub::model::aws_rds_db_instance_details::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for AwsRdsDbInstanceDetails

Implementations

impl Builder

pub fn associated_roles(
self,
input: impl Into<AwsRdsDbInstanceAssociatedRole>
) -> Self

Appends an item to associated_roles.

To override the contents of this collection use set_associated_roles.

The IAM roles associated with the DB instance.

pub fn set_associated_roles(
self,
input: Option<Vec<AwsRdsDbInstanceAssociatedRole>>
) -> Self

The IAM roles associated with the DB instance.

pub fn ca_certificate_identifier(self, input: impl Into<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn set_ca_certificate_identifier(self, input: Option<String>) -> Self

The identifier of the CA certificate for this DB instance.

pub fn db_cluster_identifier(self, input: impl Into<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn set_db_cluster_identifier(self, input: Option<String>) -> Self

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

pub fn db_instance_identifier(self, input: impl Into<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn set_db_instance_identifier(self, input: Option<String>) -> Self

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

pub fn db_instance_class(self, input: impl Into<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn set_db_instance_class(self, input: Option<String>) -> Self

Contains the name of the compute and memory capacity class of the DB instance.

pub fn db_instance_port(self, input: i32) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn set_db_instance_port(self, input: Option<i32>) -> Self

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pub fn dbi_resource_id(self, input: impl Into<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

pub fn set_dbi_resource_id(self, input: Option<String>) -> Self

The Amazon Web Services Region-unique, immutable identifier for the DB instance. This identifier is found in CloudTrail log entries whenever the KMS key for the DB instance is accessed.

pub fn db_name(self, input: impl Into<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

pub fn set_db_name(self, input: Option<String>) -> Self

The meaning of this parameter differs according to the database engine you use.

MySQL, MariaDB, SQL Server, PostgreSQL

Contains the name of the initial database of this instance that was provided at create time, if one was specified when the DB instance was created. This same name is returned for the life of the DB instance.

Oracle

Contains the Oracle System ID (SID) of the created DB instance. Not shown when the returned parameters do not apply to an Oracle DB instance.

pub fn deletion_protection(self, input: bool) -> Self

Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

pub fn set_deletion_protection(self, input: Option<bool>) -> Self

Indicates whether the DB instance has deletion protection enabled.

When deletion protection is enabled, the database cannot be deleted.

pub fn endpoint(self, input: AwsRdsDbInstanceEndpoint) -> Self

Specifies the connection endpoint.

pub fn set_endpoint(self, input: Option<AwsRdsDbInstanceEndpoint>) -> Self

Specifies the connection endpoint.

pub fn engine(self, input: impl Into<String>) -> Self

Provides the name of the database engine to use for this DB instance.

pub fn set_engine(self, input: Option<String>) -> Self

Provides the name of the database engine to use for this DB instance.

pub fn engine_version(self, input: impl Into<String>) -> Self

Indicates the database engine version.

pub fn set_engine_version(self, input: Option<String>) -> Self

Indicates the database engine version.

pub fn iam_database_authentication_enabled(self, input: bool) -> Self

True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

pub fn set_iam_database_authentication_enabled(
self,
input: Option<bool>
) -> Self

True if mapping of IAM accounts to database accounts is enabled, and otherwise false.

IAM database authentication can be enabled for the following database engines.

For MySQL 5.6, minor version 5.6.34 or higher

For MySQL 5.7, minor version 5.7.16 or higher

Aurora 5.6 or higher

pub fn instance_create_time(self, input: impl Into<String>) -> Self

Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_instance_create_time(self, input: Option<String>) -> Self

Indicates when the DB instance was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn kms_key_id(self, input: impl Into<String>) -> Self

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

pub fn set_kms_key_id(self, input: Option<String>) -> Self

If StorageEncrypted is true, the KMS key identifier for the encrypted DB instance.

pub fn publicly_accessible(self, input: bool) -> Self

Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

pub fn set_publicly_accessible(self, input: Option<bool>) -> Self

Specifies the accessibility options for the DB instance.

A value of true specifies an Internet-facing instance with a publicly resolvable DNS name, which resolves to a public IP address.

A value of false specifies an internal instance with a DNS name that resolves to a private IP address.

pub fn storage_encrypted(self, input: bool) -> Self

Specifies whether the DB instance is encrypted.

pub fn set_storage_encrypted(self, input: Option<bool>) -> Self

Specifies whether the DB instance is encrypted.

pub fn tde_credential_arn(self, input: impl Into<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn set_tde_credential_arn(self, input: Option<String>) -> Self

The ARN from the key store with which the instance is associated for TDE encryption.

pub fn vpc_security_groups(
self,
input: impl Into<AwsRdsDbInstanceVpcSecurityGroup>
) -> Self

Appends an item to vpc_security_groups.

To override the contents of this collection use set_vpc_security_groups.

A list of VPC security groups that the DB instance belongs to.

pub fn set_vpc_security_groups(
self,
input: Option<Vec<AwsRdsDbInstanceVpcSecurityGroup>>
) -> Self

A list of VPC security groups that the DB instance belongs to.

pub fn multi_az(self, input: bool) -> Self

Whether the DB instance is a multiple Availability Zone deployment.

pub fn set_multi_az(self, input: Option<bool>) -> Self

Whether the DB instance is a multiple Availability Zone deployment.

pub fn enhanced_monitoring_resource_arn(self, input: impl Into<String>) -> Self

The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

pub fn set_enhanced_monitoring_resource_arn(self, input: Option<String>) -> Self

The ARN of the CloudWatch Logs log stream that receives the enhanced monitoring metrics data for the DB instance.

pub fn db_instance_status(self, input: impl Into<String>) -> Self

The current status of the DB instance.

pub fn set_db_instance_status(self, input: Option<String>) -> Self

The current status of the DB instance.

pub fn master_username(self, input: impl Into<String>) -> Self

The master user name of the DB instance.

pub fn set_master_username(self, input: Option<String>) -> Self

The master user name of the DB instance.

pub fn allocated_storage(self, input: i32) -> Self

The amount of storage (in gigabytes) to initially allocate for the DB instance.

pub fn set_allocated_storage(self, input: Option<i32>) -> Self

The amount of storage (in gigabytes) to initially allocate for the DB instance.

pub fn preferred_backup_window(self, input: impl Into<String>) -> Self

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

pub fn set_preferred_backup_window(self, input: Option<String>) -> Self

The range of time each day when automated backups are created, if automated backups are enabled.

Uses the format HH:MM-HH:MM. For example, 04:52-05:22.

pub fn backup_retention_period(self, input: i32) -> Self

The number of days for which to retain automated backups.

pub fn set_backup_retention_period(self, input: Option<i32>) -> Self

The number of days for which to retain automated backups.

pub fn db_security_groups(self, input: impl Into<String>) -> Self

Appends an item to db_security_groups.

To override the contents of this collection use set_db_security_groups.

A list of the DB security groups to assign to the DB instance.

pub fn set_db_security_groups(self, input: Option<Vec<String>>) -> Self

A list of the DB security groups to assign to the DB instance.

pub fn db_parameter_groups(
self,
input: impl Into<AwsRdsDbParameterGroup>
) -> Self

Appends an item to db_parameter_groups.

To override the contents of this collection use set_db_parameter_groups.

A list of the DB parameter groups to assign to the DB instance.

pub fn set_db_parameter_groups(
self,
input: Option<Vec<AwsRdsDbParameterGroup>>
) -> Self

A list of the DB parameter groups to assign to the DB instance.

pub fn availability_zone(self, input: impl Into<String>) -> Self

The Availability Zone where the DB instance will be created.

pub fn set_availability_zone(self, input: Option<String>) -> Self

The Availability Zone where the DB instance will be created.

pub fn db_subnet_group(self, input: AwsRdsDbSubnetGroup) -> Self

Information about the subnet group that is associated with the DB instance.

pub fn set_db_subnet_group(self, input: Option<AwsRdsDbSubnetGroup>) -> Self

Information about the subnet group that is associated with the DB instance.

pub fn preferred_maintenance_window(self, input: impl Into<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format :HH:MM-:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Uses the format :HH:MM-:HH:MM.

For the day values, use mon|tue|wed|thu|fri|sat|sun.

For example, sun:09:32-sun:10:02.

pub fn pending_modified_values(
self,
input: AwsRdsDbPendingModifiedValues
) -> Self

Changes to the DB instance that are currently pending.

pub fn set_pending_modified_values(
self,
input: Option<AwsRdsDbPendingModifiedValues>
) -> Self

Changes to the DB instance that are currently pending.

pub fn latest_restorable_time(self, input: impl Into<String>) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn set_latest_restorable_time(self, input: Option<String>) -> Self

Specifies the latest time to which a database can be restored with point-in-time restore.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn auto_minor_version_upgrade(self, input: bool) -> Self

Indicates whether minor version patches are applied automatically.

pub fn set_auto_minor_version_upgrade(self, input: Option<bool>) -> Self

Indicates whether minor version patches are applied automatically.

pub fn read_replica_source_db_instance_identifier(
self,
input: impl Into<String>
) -> Self

If this DB instance is a read replica, contains the identifier of the source DB instance.

pub fn set_read_replica_source_db_instance_identifier(
self,
input: Option<String>
) -> Self

If this DB instance is a read replica, contains the identifier of the source DB instance.

pub fn read_replica_db_instance_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_instance_identifiers.

To override the contents of this collection use set_read_replica_db_instance_identifiers.

List of identifiers of the read replicas associated with this DB instance.

pub fn set_read_replica_db_instance_identifiers(
self,
input: Option<Vec<String>>
) -> Self

List of identifiers of the read replicas associated with this DB instance.

pub fn read_replica_db_cluster_identifiers(
self,
input: impl Into<String>
) -> Self

Appends an item to read_replica_db_cluster_identifiers.

To override the contents of this collection use set_read_replica_db_cluster_identifiers.

List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

pub fn set_read_replica_db_cluster_identifiers(
self,
input: Option<Vec<String>>
) -> Self

List of identifiers of Aurora DB clusters to which the RDS DB instance is replicated as a read replica.

pub fn license_model(self, input: impl Into<String>) -> Self

License model information for this DB instance.

pub fn set_license_model(self, input: Option<String>) -> Self

License model information for this DB instance.

pub fn iops(self, input: i32) -> Self

Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

pub fn set_iops(self, input: Option<i32>) -> Self

Specifies the provisioned IOPS (I/O operations per second) for this DB instance.

pub fn option_group_memberships(
self,
input: impl Into<AwsRdsDbOptionGroupMembership>
) -> Self

Appends an item to option_group_memberships.

To override the contents of this collection use set_option_group_memberships.

The list of option group memberships for this DB instance.

pub fn set_option_group_memberships(
self,
input: Option<Vec<AwsRdsDbOptionGroupMembership>>
) -> Self

The list of option group memberships for this DB instance.

pub fn character_set_name(self, input: impl Into<String>) -> Self

The name of the character set that this DB instance is associated with.

pub fn set_character_set_name(self, input: Option<String>) -> Self

The name of the character set that this DB instance is associated with.

pub fn secondary_availability_zone(self, input: impl Into<String>) -> Self

For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

pub fn set_secondary_availability_zone(self, input: Option<String>) -> Self

For a DB instance with multi-Availability Zone support, the name of the secondary Availability Zone.

pub fn status_infos(self, input: impl Into<AwsRdsDbStatusInfo>) -> Self

Appends an item to status_infos.

To override the contents of this collection use set_status_infos.

The status of a read replica. If the instance isn't a read replica, this is empty.

pub fn set_status_infos(self, input: Option<Vec<AwsRdsDbStatusInfo>>) -> Self

The status of a read replica. If the instance isn't a read replica, this is empty.

pub fn storage_type(self, input: impl Into<String>) -> Self

The storage type for the DB instance.

pub fn set_storage_type(self, input: Option<String>) -> Self

The storage type for the DB instance.

pub fn domain_memberships(
self,
input: impl Into<AwsRdsDbDomainMembership>
) -> Self

Appends an item to domain_memberships.

To override the contents of this collection use set_domain_memberships.

The Active Directory domain membership records associated with the DB instance.

pub fn set_domain_memberships(
self,
input: Option<Vec<AwsRdsDbDomainMembership>>
) -> Self

The Active Directory domain membership records associated with the DB instance.

pub fn copy_tags_to_snapshot(self, input: bool) -> Self

Whether to copy resource tags to snapshots of the DB instance.

pub fn set_copy_tags_to_snapshot(self, input: Option<bool>) -> Self

Whether to copy resource tags to snapshots of the DB instance.

pub fn monitoring_interval(self, input: i32) -> Self

The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

pub fn set_monitoring_interval(self, input: Option<i32>) -> Self

The interval, in seconds, between points when enhanced monitoring metrics are collected for the DB instance.

pub fn monitoring_role_arn(self, input: impl Into<String>) -> Self

The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

pub fn set_monitoring_role_arn(self, input: Option<String>) -> Self

The ARN for the IAM role that permits Amazon RDS to send enhanced monitoring metrics to CloudWatch Logs.

pub fn promotion_tier(self, input: i32) -> Self

The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

pub fn set_promotion_tier(self, input: Option<i32>) -> Self

The order in which to promote an Aurora replica to the primary instance after a failure of the existing primary instance.

pub fn timezone(self, input: impl Into<String>) -> Self

The time zone of the DB instance.

pub fn set_timezone(self, input: Option<String>) -> Self

The time zone of the DB instance.

pub fn performance_insights_enabled(self, input: bool) -> Self

Indicates whether Performance Insights is enabled for the DB instance.

pub fn set_performance_insights_enabled(self, input: Option<bool>) -> Self

Indicates whether Performance Insights is enabled for the DB instance.

pub fn performance_insights_kms_key_id(self, input: impl Into<String>) -> Self

The identifier of the KMS key used to encrypt the Performance Insights data.

pub fn set_performance_insights_kms_key_id(self, input: Option<String>) -> Self

The identifier of the KMS key used to encrypt the Performance Insights data.

pub fn performance_insights_retention_period(self, input: i32) -> Self

The number of days to retain Performance Insights data.

pub fn set_performance_insights_retention_period(
self,
input: Option<i32>
) -> Self

The number of days to retain Performance Insights data.

pub fn enabled_cloud_watch_logs_exports(self, input: impl Into<String>) -> Self

Appends an item to enabled_cloud_watch_logs_exports.

To override the contents of this collection use set_enabled_cloud_watch_logs_exports.

A list of log types that this DB instance is configured to export to CloudWatch Logs.

pub fn set_enabled_cloud_watch_logs_exports(
self,
input: Option<Vec<String>>
) -> Self

A list of log types that this DB instance is configured to export to CloudWatch Logs.

pub fn processor_features(
self,
input: impl Into<AwsRdsDbProcessorFeature>
) -> Self

Appends an item to processor_features.

To override the contents of this collection use set_processor_features.

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn set_processor_features(
self,
input: Option<Vec<AwsRdsDbProcessorFeature>>
) -> Self

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pub fn listener_endpoint(self, input: AwsRdsDbInstanceEndpoint) -> Self

Specifies the connection endpoint.

pub fn set_listener_endpoint(
self,
input: Option<AwsRdsDbInstanceEndpoint>
) -> Self

Specifies the connection endpoint.

pub fn max_allocated_storage(self, input: i32) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

pub fn set_max_allocated_storage(self, input: Option<i32>) -> Self

The upper limit to which Amazon RDS can automatically scale the storage of the DB instance.

pub fn build(self) -> AwsRdsDbInstanceDetails

Consumes the builder and constructs a AwsRdsDbInstanceDetails