Struct aws_sdk_securityhub::model::AwsRedshiftClusterDetails

#[non_exhaustive]
pub struct AwsRedshiftClusterDetails {

    pub allow_version_upgrade: bool,
    pub automated_snapshot_retention_period: i32,
    pub availability_zone: Option<String>,
    pub cluster_availability_status: Option<String>,
    pub cluster_create_time: Option<String>,
    pub cluster_identifier: Option<String>,
    pub cluster_nodes: Option<Vec<AwsRedshiftClusterClusterNode>>,
    pub cluster_parameter_groups: Option<Vec<AwsRedshiftClusterClusterParameterGroup>>,
    pub cluster_public_key: Option<String>,
    pub cluster_revision_number: Option<String>,
    pub cluster_security_groups: Option<Vec<AwsRedshiftClusterClusterSecurityGroup>>,
    pub cluster_snapshot_copy_status: Option<AwsRedshiftClusterClusterSnapshotCopyStatus>,
    pub cluster_status: Option<String>,
    pub cluster_subnet_group_name: Option<String>,
    pub cluster_version: Option<String>,
    pub db_name: Option<String>,
    pub deferred_maintenance_windows: Option<Vec<AwsRedshiftClusterDeferredMaintenanceWindow>>,
    pub elastic_ip_status: Option<AwsRedshiftClusterElasticIpStatus>,
    pub elastic_resize_number_of_node_options: Option<String>,
    pub encrypted: bool,
    pub endpoint: Option<AwsRedshiftClusterEndpoint>,
    pub enhanced_vpc_routing: bool,
    pub expected_next_snapshot_schedule_time: Option<String>,
    pub expected_next_snapshot_schedule_time_status: Option<String>,
    pub hsm_status: Option<AwsRedshiftClusterHsmStatus>,
    pub iam_roles: Option<Vec<AwsRedshiftClusterIamRole>>,
    pub kms_key_id: Option<String>,
    pub maintenance_track_name: Option<String>,
    pub manual_snapshot_retention_period: i32,
    pub master_username: Option<String>,
    pub next_maintenance_window_start_time: Option<String>,
    pub node_type: Option<String>,
    pub number_of_nodes: i32,
    pub pending_actions: Option<Vec<String>>,
    pub pending_modified_values: Option<AwsRedshiftClusterPendingModifiedValues>,
    pub preferred_maintenance_window: Option<String>,
    pub publicly_accessible: bool,
    pub resize_info: Option<AwsRedshiftClusterResizeInfo>,
    pub restore_status: Option<AwsRedshiftClusterRestoreStatus>,
    pub snapshot_schedule_identifier: Option<String>,
    pub snapshot_schedule_state: Option<String>,
    pub vpc_id: Option<String>,
    pub vpc_security_groups: Option<Vec<AwsRedshiftClusterVpcSecurityGroup>>,

}

Details about an Amazon Redshift cluster.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

allow_version_upgrade: bool

Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

automated_snapshot_retention_period: i32

The number of days that automatic cluster snapshots are retained.

availability_zone: Option<String>

The name of the Availability Zone in which the cluster is located.

cluster_availability_status: Option<String>

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

cluster_create_time: Option<String>

Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

cluster_identifier: Option<String>

The unique identifier of the cluster.

cluster_nodes: Option<Vec<AwsRedshiftClusterClusterNode>>

The nodes in the cluster.

cluster_parameter_groups: Option<Vec<AwsRedshiftClusterClusterParameterGroup>>

The list of cluster parameter groups that are associated with this cluster.

cluster_public_key: Option<String>

The public key for the cluster.

cluster_revision_number: Option<String>

The specific revision number of the database in the cluster.

cluster_security_groups: Option<Vec<AwsRedshiftClusterClusterSecurityGroup>>

A list of cluster security groups that are associated with the cluster.

cluster_snapshot_copy_status: Option<AwsRedshiftClusterClusterSnapshotCopyStatus>

Information about the destination Region and retention period for the cross-Region snapshot copy.

cluster_status: Option<String>

The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

cluster_subnet_group_name: Option<String>

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

cluster_version: Option<String>

The version ID of the Amazon Redshift engine that runs on the cluster.

db_name: Option<String>

The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.

deferred_maintenance_windows: Option<Vec<AwsRedshiftClusterDeferredMaintenanceWindow>>

List of time windows during which maintenance was deferred.

elastic_ip_status: Option<AwsRedshiftClusterElasticIpStatus>

Information about the status of the Elastic IP (EIP) address.

elastic_resize_number_of_node_options: Option<String>

The number of nodes that you can use the elastic resize method to resize the cluster to.

encrypted: bool

Indicates whether the data in the cluster is encrypted at rest.

endpoint: Option<AwsRedshiftClusterEndpoint>

The connection endpoint.

enhanced_vpc_routing: bool

Indicates whether to create the cluster with enhanced VPC routing enabled.

expected_next_snapshot_schedule_time: Option<String>

Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

expected_next_snapshot_schedule_time_status: Option<String>

The status of the next expected snapshot.

Valid values: OnTrack | Pending

hsm_status: Option<AwsRedshiftClusterHsmStatus>

Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

iam_roles: Option<Vec<AwsRedshiftClusterIamRole>>

A list of IAM roles that the cluster can use to access other Amazon Web Services services.

kms_key_id: Option<String>

The identifier of the KMS encryption key that is used to encrypt data in the cluster.

maintenance_track_name: Option<String>

The name of the maintenance track for the cluster.

manual_snapshot_retention_period: i32

The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

master_username: Option<String>

The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

next_maintenance_window_start_time: Option<String>

Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

node_type: Option<String>

The node type for the nodes in the cluster.

number_of_nodes: i32

The number of compute nodes in the cluster.

pending_actions: Option<Vec<String>>

A list of cluster operations that are waiting to start.

pending_modified_values: Option<AwsRedshiftClusterPendingModifiedValues>

A list of changes to the cluster that are currently pending.

preferred_maintenance_window: Option<String>

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: :HH:MM-:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

publicly_accessible: bool

Whether the cluster can be accessed from a public network.

resize_info: Option<AwsRedshiftClusterResizeInfo>

Information about the resize operation for the cluster.

restore_status: Option<AwsRedshiftClusterRestoreStatus>

Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

snapshot_schedule_identifier: Option<String>

A unique identifier for the cluster snapshot schedule.

snapshot_schedule_state: Option<String>

The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

vpc_id: Option<String>

The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

vpc_security_groups: Option<Vec<AwsRedshiftClusterVpcSecurityGroup>>

The list of VPC security groups that the cluster belongs to, if the cluster is in a VPC.

Implementations

impl AwsRedshiftClusterDetails

pub fn allow_version_upgrade(&self) -> bool

Indicates whether major version upgrades are applied automatically to the cluster during the maintenance window.

pub fn automated_snapshot_retention_period(&self) -> i32

The number of days that automatic cluster snapshots are retained.

pub fn availability_zone(&self) -> Option<&str>

The name of the Availability Zone in which the cluster is located.

pub fn cluster_availability_status(&self) -> Option<&str>

The availability status of the cluster for queries. Possible values are the following:

Available - The cluster is available for queries.

Unavailable - The cluster is not available for queries.

Maintenance - The cluster is intermittently available for queries due to maintenance activities.

Modifying -The cluster is intermittently available for queries due to changes that modify the cluster.

Failed - The cluster failed and is not available for queries.

pub fn cluster_create_time(&self) -> Option<&str>

Indicates when the cluster was created.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn cluster_identifier(&self) -> Option<&str>

The unique identifier of the cluster.

pub fn cluster_nodes(&self) -> Option<&[AwsRedshiftClusterClusterNode]>

The nodes in the cluster.

pub fn cluster_parameter_groups(
&self
) -> Option<&[AwsRedshiftClusterClusterParameterGroup]>

The list of cluster parameter groups that are associated with this cluster.

pub fn cluster_public_key(&self) -> Option<&str>

The public key for the cluster.

pub fn cluster_revision_number(&self) -> Option<&str>

The specific revision number of the database in the cluster.

pub fn cluster_security_groups(
&self
) -> Option<&[AwsRedshiftClusterClusterSecurityGroup]>

A list of cluster security groups that are associated with the cluster.

pub fn cluster_snapshot_copy_status(
&self
) -> Option<&AwsRedshiftClusterClusterSnapshotCopyStatus>

Information about the destination Region and retention period for the cross-Region snapshot copy.

pub fn cluster_status(&self) -> Option<&str>

The current status of the cluster.

Valid values: available | available, prep-for-resize | available, resize-cleanup | cancelling-resize | creating | deleting | final-snapshot | hardware-failure | incompatible-hsm | incompatible-network | incompatible-parameters | incompatible-restore | modifying | paused | rebooting | renaming | resizing | rotating-keys | storage-full | updating-hsm

pub fn cluster_subnet_group_name(&self) -> Option<&str>

The name of the subnet group that is associated with the cluster. This parameter is valid only when the cluster is in a VPC.

pub fn cluster_version(&self) -> Option<&str>

The version ID of the Amazon Redshift engine that runs on the cluster.

pub fn db_name(&self) -> Option<&str>

The name of the initial database that was created when the cluster was created.

The same name is returned for the life of the cluster.

If an initial database is not specified, a database named devdev is created by default.

pub fn deferred_maintenance_windows(
&self
) -> Option<&[AwsRedshiftClusterDeferredMaintenanceWindow]>

List of time windows during which maintenance was deferred.

pub fn elastic_ip_status(&self) -> Option<&AwsRedshiftClusterElasticIpStatus>

Information about the status of the Elastic IP (EIP) address.

pub fn elastic_resize_number_of_node_options(&self) -> Option<&str>

The number of nodes that you can use the elastic resize method to resize the cluster to.

pub fn encrypted(&self) -> bool

Indicates whether the data in the cluster is encrypted at rest.

pub fn endpoint(&self) -> Option<&AwsRedshiftClusterEndpoint>

The connection endpoint.

pub fn enhanced_vpc_routing(&self) -> bool

Indicates whether to create the cluster with enhanced VPC routing enabled.

pub fn expected_next_snapshot_schedule_time(&self) -> Option<&str>

Indicates when the next snapshot is expected to be taken. The cluster must have a valid snapshot schedule and have backups enabled.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn expected_next_snapshot_schedule_time_status(&self) -> Option<&str>

The status of the next expected snapshot.

Valid values: OnTrack | Pending

pub fn hsm_status(&self) -> Option<&AwsRedshiftClusterHsmStatus>

Information about whether the Amazon Redshift cluster finished applying any changes to hardware security module (HSM) settings that were specified in a modify cluster command.

pub fn iam_roles(&self) -> Option<&[AwsRedshiftClusterIamRole]>

A list of IAM roles that the cluster can use to access other Amazon Web Services services.

pub fn kms_key_id(&self) -> Option<&str>

The identifier of the KMS encryption key that is used to encrypt data in the cluster.

pub fn maintenance_track_name(&self) -> Option<&str>

The name of the maintenance track for the cluster.

pub fn manual_snapshot_retention_period(&self) -> i32

The default number of days to retain a manual snapshot.

If the value is -1, the snapshot is retained indefinitely.

This setting doesn't change the retention period of existing snapshots.

Valid values: Either -1 or an integer between 1 and 3,653

pub fn master_username(&self) -> Option<&str>

The master user name for the cluster. This name is used to connect to the database that is specified in as the value of DBName.

pub fn next_maintenance_window_start_time(&self) -> Option<&str>

Indicates the start of the next maintenance window.

Uses the date-time format specified in RFC 3339 section 5.6, Internet Date/Time Format. The value cannot contain spaces. For example, 2020-03-22T13:22:13.933Z.

pub fn node_type(&self) -> Option<&str>

The node type for the nodes in the cluster.

pub fn number_of_nodes(&self) -> i32

The number of compute nodes in the cluster.

pub fn pending_actions(&self) -> Option<&[String]>

A list of cluster operations that are waiting to start.

pub fn pending_modified_values(
&self
) -> Option<&AwsRedshiftClusterPendingModifiedValues>

A list of changes to the cluster that are currently pending.

pub fn preferred_maintenance_window(&self) -> Option<&str>

The weekly time range, in Universal Coordinated Time (UTC), during which system maintenance can occur.

Format: :HH:MM-:HH:MM

For the day values, use mon | tue | wed | thu | fri | sat | sun

For example, sun:09:32-sun:10:02

pub fn publicly_accessible(&self) -> bool

Whether the cluster can be accessed from a public network.

pub fn resize_info(&self) -> Option<&AwsRedshiftClusterResizeInfo>

Information about the resize operation for the cluster.

pub fn restore_status(&self) -> Option<&AwsRedshiftClusterRestoreStatus>

Information about the status of a cluster restore action. Only applies to a cluster that was created by restoring a snapshot.

pub fn snapshot_schedule_identifier(&self) -> Option<&str>

A unique identifier for the cluster snapshot schedule.

pub fn snapshot_schedule_state(&self) -> Option<&str>

The current state of the cluster snapshot schedule.

Valid values: MODIFYING | ACTIVE | FAILED

pub fn vpc_id(&self) -> Option<&str>

The identifier of the VPC that the cluster is in, if the cluster is in a VPC.

pub fn vpc_security_groups(
&self
) -> Option<&[AwsRedshiftClusterVpcSecurityGroup]>

The list of VPC security groups that the cluster belongs to, if the cluster is in a VPC.

impl AwsRedshiftClusterDetails

pub fn builder() -> Builder

Creates a new builder-style object to manufacture AwsRedshiftClusterDetails